% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add a tooltip to the label of an input}
\usage{
add_tooltip(label, message)
}
\arguments{
\item{label}{character. The text to display initially}

\item{message}{character. The text to display in the tooltip}
}
\description{
Add a tooltip to the label of an input
}
\keyword{internal}
