% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{AddStudyIds}
\alias{AddStudyIds}
\title{Add a new column in the data for study IDs, both for long and wide formats.}
\usage{
AddStudyIds(data)
}
\arguments{
\item{data}{Data frame in which to search for treatment names}
}
\value{
Vector of all treatment names
}
\description{
Add a new column in the data for study IDs, both for long and wide formats.
}
