% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_compare_f.R
\name{baseline_compare}
\alias{baseline_compare}
\title{Compare treatment pairs}
\usage{
baseline_compare(model, logger = NULL)
}
\arguments{
\item{model}{list. Object produced by \code{baseline_model()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
Relative effects table
}
\description{
Produce a table of comparisons of all treatment pairs for baseline risk
models using \code{bnma::relative.effects.table()}
}
\examples{
\donttest{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

fitted_baseline_model <- baseline_model(configured_data = configured_data,
                                        regressor_type = "shared")

baseline_compare(model = fitted_baseline_model)
}
}
