% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_deviance_f.R
\name{baseline_deviance}
\alias{baseline_deviance}
\title{Produce deviance plots for baseline risk models}
\usage{
baseline_deviance(model, async = FALSE)
}
\arguments{
\item{model}{Output model produced by \code{baseline_model()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
list containing:
\item{deviance_mtc}{equivalent summary to that produced by \code{gemtc::mtc.deviance()}}
\item{stem_plot}{plotly object}
\item{lev_plot}{plotly object}
}
\description{
Produce deviance plotly plots for baseline risk models. Unlike for
\code{bayes_model} output, only stem and leverage plots are produced.
}
\examples{
\donttest{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

fitted_baseline_model <- baseline_model(configured_data = configured_data,
                                        regressor_type = "shared")

baseline_deviance(model = fitted_baseline_model)
}
}
