% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_forest_f.R
\name{baseline_forest}
\alias{baseline_forest}
\title{Produce a forest plot for baseline risk models}
\usage{
baseline_forest(
  model,
  xmin = NULL,
  xmax = NULL,
  title = "Baseline risk regression analysis",
  ranking = FALSE,
  logger = NULL
)
}
\arguments{
\item{model}{Output produced by \code{baseline_model()}}

\item{xmin}{numeric. Minimum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{xmax}{numeric. Maximum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{title}{character. Title for the plot. Defaults to \verb{Baseline risk regression analysis}}

\item{ranking}{logical. Whether the function is being used in \code{baseline_ranking}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Produce a forest plot for a baseline risk model using \code{gemtc::blobbogram()}
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_iter, max_iter and check_iter are set low to run quickly, but should
# be left as the default values in real use

fitted_baseline_model <- baseline_model(configured_data = configured_data,
                                        regressor_type = "shared",
                                        n_iter = 120,
                                        max_iter = 120,
                                        check_iter = 10)

baseline_forest(model = fitted_baseline_model)
}
