% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_forest_f.R
\name{baseline_forest_limits}
\alias{baseline_forest_limits}
\title{Extract the default x-axis limits using the same method internal to
\code{gemtc::forest()}}
\usage{
baseline_forest_limits(forest_data)
}
\arguments{
\item{forest_data}{data.frame. Created by \code{format_baseline_forest()}}
}
\description{
Extract the default x-axis limits using the same method internal to
\code{gemtc::forest()}
}
