% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_regression_f.R
\name{baseline_regression}
\alias{baseline_regression}
\title{Baseline regression data}
\usage{
baseline_regression(model, configured_data, async = FALSE)
}
\arguments{
\item{model}{Output produced by \code{baseline_model()}}

\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{directness}{list. Output from \code{CalculateDirectness()}}
\item{credible_regions}{list. Output from \code{CalculateCredibleRegions()}}
}
\description{
Generate data required to produce a metaregression plot
for a baseline risk model.
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_iter, max_iter and check_iter are set low to run quickly, but should
# be left as the default values in real use

fitted_baseline_model <- baseline_model(configured_data = configured_data,
                                        regressor_type = "shared",
                                        n_iter = 120,
                                        max_iter = 120,
                                        check_iter = 10)

regression_data <- baseline_regression(model = fitted_baseline_model,
                                       configured_data = configured_data)
}
