% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_summary_f.R
\name{baseline_summary_block}
\alias{baseline_summary_block}
\title{Add the treatment effect lines, confidence intervals, studies, and comparison header, for a single comparison.}
\usage{
baseline_summary_block(
  df,
  label_x_position,
  outcome_x_position,
  y_positions,
  treatment
)
}
\arguments{
\item{df}{Dataframe of the data to plot}

\item{label_x_position}{x co-ordinate for the start of the labels.}

\item{outcome_x_position}{x co-ordinate for the start of the text.}

\item{y_positions}{Vector of y co-ordinates for the study labels and confidence intervals.}

\item{treatment}{Treatment.}
}
\description{
Add the treatment effect lines, confidence intervals, studies, and comparison header, for a single comparison.
}
