% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_summary_f.R
\name{baseline_summary_outcome_text}
\alias{baseline_summary_outcome_text}
\title{Add the effects size and confidence interval text for a single comparison in a single study.}
\usage{
baseline_summary_outcome_text(df, x_position, studlab, treatment)
}
\arguments{
\item{df}{Dataframe of the data to plot}

\item{x_position}{x co-ordinate for the start of the text.}

\item{studlab}{Study.}

\item{treatment}{Treatment.}

\item{y_position}{y co-ordinate for the effect size and confidence interval.}
}
\description{
Add the effects size and confidence interval text for a single comparison in a single study.
}
