% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_summary_f.R
\name{baseline_summary_treatment_effect}
\alias{baseline_summary_treatment_effect}
\title{Add the treatment effect line and confidence interval for a single comparison in a single study.}
\usage{
baseline_summary_treatment_effect(
  df,
  y_position,
  studlab,
  treatment,
  ci_limit_height = 0.3
)
}
\arguments{
\item{df}{Dataframe of the data to plot}

\item{y_position}{y co-ordinate for the study label and confidence interval.}

\item{studlab}{Study.}

\item{treatment}{Treatment}

\item{ci_limit_height}{Height of the vertical lines at the end of the confidence intervals. Defaults to 0.3.}
}
\description{
Add the treatment effect line and confidence interval for a single comparison in a single study.
}
