% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_summary_f.R
\name{baseline_summary_y_position}
\alias{baseline_summary_y_position}
\title{Convert a vector of treatments to y positions by adding one
when the previous treatment is identical, or three if it is different}
\usage{
baseline_summary_y_position(treatments)
}
\arguments{
\item{treatments}{vector of treatments}
}
\value{
vector of y positions
}
\description{
Convert a vector of treatments to y positions by adding one
when the previous treatment is identical, or three if it is different
}
