% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{BaselineRiskDicTable}
\alias{BaselineRiskDicTable}
\title{Creates a DIC table in gemtc format from a bnma model}
\usage{
BaselineRiskDicTable(br_model)
}
\arguments{
\item{br_model}{Output from bnma::network.run(), typically created from BaselineRiskRegression().}
}
\value{
A DIC table in the same format as from gemtc.
}
\description{
Creates a DIC table in gemtc format from a bnma model
}
