% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{BaselineRiskModelOutput}
\alias{BaselineRiskModelOutput}
\title{Function to collate model output to be used in the novel graph.}
\usage{
BaselineRiskModelOutput(connected_data, treatment_df, model, outcome_measure)
}
\arguments{
\item{connected_data}{Data frame from which model was calculated.}

\item{treatment_df}{Data frame containing treatment IDs (Number) and names (Label).}

\item{model}{Completed model object after running BaselineRiskRegression().}

\item{outcome_measure}{The outcome measure for the analysis: One of: "OR", "RR", "MD".}
}
\value{
List of bnma related output:
mtcResults = model object itself carried through (needed to match existing code).
covariate_value = The mean covariate value, used for centring.
reference_name = The name of the reference treatment.
comparator_names = Vector containing the names of the comparators.
a = text output stating whether fixed or random effects.
cov_value_sentence = text output stating the value for which the covariate has been set to for producing output.
slopes = named list of slopes for the regression equations (unstandardised - equal to one 'increment').
intercepts = named list of intercepts for the regression equations at cov_value.
outcome = The outcome type for the analysis eg. "MD" or "OR".
model = effects type, "fixed" or "random".
covariate_min = Vector of minimum covariate values directly contributing to the regression.
covariate_max = Vector of maximum covariate values directly contributing to the regression.
dic = Summary of model fit
}
\description{
Function to collate model output to be used in the novel graph.
}
