% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{BaselineRiskNetwork}
\alias{BaselineRiskNetwork}
\title{Creates the baseline risk meta-regression network in BNMA}
\usage{
BaselineRiskNetwork(br_data, outcome, model_type, cov_parameters)
}
\arguments{
\item{br_data}{A list of data in the format produced by FormatForBnma().}

\item{outcome}{"Continuous" or "Binary".}

\item{model_type}{"fixed" or "random".}

\item{cov_parameters}{"shared", "exchangable", or "unrelated".}
}
\value{
Output from bnma::network.data().
}
\description{
Creates the baseline risk meta-regression network in BNMA
}
