% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{BaselineRiskRegression}
\alias{BaselineRiskRegression}
\title{Fits the baseline risk meta-regression model in BNMA}
\usage{
BaselineRiskRegression(
  connected_data,
  treatment_df,
  outcome,
  reference_treatment,
  model_type,
  cov_parameters,
  seed = 123
)
}
\arguments{
\item{connected_data}{A data frame of data in MetaInsight format.}

\item{treatment_df}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively.}

\item{outcome}{"Continuous" or "Binary".}

\item{model_type}{"fixed" or "random".}

\item{cov_parameters}{"shared", "exchangable", or "unrelated".}

\item{seed}{Seed. Defaults to 123.}

\item{reference}{treatment An element of treatment_df$Label, the .}
}
\value{
Output from bnma::network.run().
}
\description{
Fits the baseline risk meta-regression model in BNMA
}
