% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{BaselineRiskRelativeEffectsTable}
\alias{BaselineRiskRelativeEffectsTable}
\title{Puts a relative effects table from bnma into gemtc format}
\usage{
BaselineRiskRelativeEffectsTable(median_ci_table)
}
\arguments{
\item{median_ci_table}{Output from bnma::relative.effects.table(, summary_stat = "ci")}
}
\value{
A relative effects table in the same format as from gemtc.
}
\description{
Puts a relative effects table from bnma into gemtc format
}
