% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{baye_comp}
\alias{baye_comp}
\title{3b Creates a table of comparisons of all treatment pairs}
\usage{
baye_comp(model, outcome_measure)
}
\arguments{
\item{model}{Various model output created by baye().}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}

\item{metaoutcome}{"Continuous" or "Binary".}
}
\value{
Relative effects table created by gemtc::relative.effect.table().
}
\description{
3b Creates a table of comparisons of all treatment pairs
}
