% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_compare_f.R
\name{bayes_compare}
\alias{bayes_compare}
\alias{covariate_compare}
\title{Compare treatment pairs}
\usage{
bayes_compare(model, logger = NULL)

covariate_compare(...)
}
\arguments{
\item{model}{list. Object created by \code{bayes_model()} or \code{covariate_model()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{...}{Parameters passed to \code{bayes_compare()}}
}
\value{
Relative effects table
}
\description{
Produce a table of comparisons of all treatment pairs for Bayesian
models using \code{gemtc::relative.effect.table()}
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_adapt and n_iter are set low to run quickly, but should be left as the
# default values in real use

fitted_bayes_model <- bayes_model(configured_data = configured_data,
                                  n_adapt = 100,
                                  n_iter = 100)

bayes_compare(model = fitted_bayes_model)
}
