% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_details_f.R
\name{bayes_details}
\alias{bayes_details}
\alias{covariate_details}
\alias{baseline_details}
\title{Produces a summary of a Bayesian model}
\usage{
bayes_details(model, logger = NULL)

covariate_details(...)

baseline_details(...)
}
\arguments{
\item{model}{list. Output produced by \code{baseline_model()}, \code{bayes_model()} or \code{covariate_model()}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
HTML summary of the model
}
\description{
Produces a summary of a Bayesian model
}
