% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_deviance_f.R
\name{bayes_deviance}
\alias{bayes_deviance}
\alias{covariate_deviance}
\title{Produce deviance plots}
\usage{
bayes_deviance(model, n_adapt = 5000, n_iter = 20000, async = FALSE)

covariate_deviance(...)
}
\arguments{
\item{model}{Bayesian model produced by \code{bayes_model()} or \code{covariate_model()}}

\item{n_adapt}{numeric. Number of adaptation iterations.
Defaults to \code{5000} and can normally be left unchanged}

\item{n_iter}{numeric. Number of simulation iterations.
Defaults to \code{20000} and can normally be left unchanged}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}

\item{...}{Parameters passed to \code{bayes_deviance()}}
}
\value{
A list containing different elements depending on the input model:

When \code{model} was created by \code{bayes_model()} containing:
\item{deviance_mtc}{results from \code{gemtc::mtc.deviance()} for model$mtcResults}
\item{deviance_ume}{results from \code{gemtc::mtc.deviance()} for UME model}
\item{scat_plot}{plotly object}
\item{stem_plot}{plotly object}
\item{lev_plot}{plotly object}

When \code{model} was created by \code{covariate_model()} containing:
\item{deviance_mtc}{results from \code{gemtc::mtc.deviance()} for model$mtcResults}
\item{stem_plot}{plotly object}
\item{lev_plot}{plotly object}
}
\description{
Produce deviance plots using the output of \code{gemtc::mtc.deviance()}
for Bayesian and covariate models. Because these plots are interactive, it is
not currently possible to download them, although they can be included in
html reports.
}
\examples{
\donttest{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_adapt and n_iter are set low to run quickly, but should be left as the
# default values in real use

fitted_bayes_model <- bayes_model(configured_data = configured_data,
                                  n_adapt = 100,
                                  n_iter = 100)

bayes_deviance(model = fitted_bayes_model,
               n_adapt = 100,
               n_iter = 100)
}
}
