% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_forest_f.R
\name{bayes_forest}
\alias{bayes_forest}
\alias{covariate_forest}
\title{Bayesian forest plot}
\usage{
bayes_forest(
  model,
  xmin = NULL,
  xmax = NULL,
  title = "",
  ranking = FALSE,
  logger = NULL
)

covariate_forest(...)
}
\arguments{
\item{model}{list. Object created by \code{bayes_model()} or \code{covariate_model()}}

\item{xmin}{numeric. Minimum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{xmax}{numeric. Maximum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{title}{character. Title for the plot. Default is no title}

\item{ranking}{logical. Whether the function is being used in \code{bayes_ranking}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{...}{Parameters passed to \code{bayes_forest()}}
}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Produce a Bayesian forest plot with \code{gemtc::forest()}
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

fitted_bayes_model <- bayes_model(configured_data = configured_data)
bayes_forest(model = fitted_bayes_model)
}
