% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_forest_f.R
\name{bayes_forest_limits}
\alias{bayes_forest_limits}
\title{Extract the default x-axis limits using the same method internal to
\code{gemtc::forest()}}
\usage{
bayes_forest_limits(model, reference_treatment)
}
\arguments{
\item{model}{list. Object created by \code{bayes_model()} or \code{covariate_model()}}

\item{reference_treatment}{character. The reference treatment of the dataset}
}
\description{
Extract the default x-axis limits using the same method internal to
\code{gemtc::forest()}
}
