% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{bayes_mcmc}
\alias{bayes_mcmc}
\alias{covariate_mcmc}
\alias{baseline_mcmc}
\title{Produce Markov chain Monte Carlo plots for Bayesian models}
\usage{
bayes_mcmc(model, async = FALSE)

covariate_mcmc(...)

baseline_mcmc(...)
}
\arguments{
\item{model}{Output from \code{baseline_model()}, \code{bayes_model()} or \code{covariate_model()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}

\item{...}{Parameters passed to \code{bayes_mcmc()}}
}
\value{
list containing:
\item{gelman_plots}{Gelman plots}
\item{trace_plots}{Trace plots}
\item{density_plots}{Density plots}
\item{n_rows}{The number of rows for each plot}
}
\description{
Produce Markov chain Monte Carlo plots for Bayesian models
}
