% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model_f.R
\name{bayes_model}
\alias{bayes_model}
\title{Fit a Bayesian model}
\usage{
bayes_model(configured_data, n_adapt = 5000, n_iter = 20000, async = FALSE)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{n_adapt}{numeric. Number of adaptation iterations.
Defaults to \code{5000} and can normally be left unchanged}

\item{n_iter}{numeric. Number of simulation iterations.
Defaults to \code{20000} and can normally be left unchanged}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{mtcResults}{mtc.result. Output from \code{gemtc::mtc.run()}}
\item{mtcRelEffects}{mtc.result. Output from \code{gemtc::relative.effect()}}
\item{rel_eff_tbl}{mtc.relative.effect.table. Output from \code{gemtc::relative.effect.table()}}
\item{sumresults}{summary.mtc.result. Output from \code{summary(mtcRelEffects)}}
\item{mtcNetwork}{mtc.network. Output from \code{gemtc::mtc.network()}}
\item{dic}{dataframe. Containing the statistics 'Dbar', 'pD', 'DIC', and 'data points'}
\item{outcome}{character. The \code{outcome} from \code{configured_data}}
\item{outcome_measure}{character. The \code{outcome_measure}from \code{configured_data}}
\item{reference_treatment}{character. The \code{reference_treatment}from \code{configured_data}}
\item{effects}{character. The \code{effects} from \code{configured_data}}
\item{seed}{numeric. The \code{seed} from \code{configured_data}}
}
\description{
Fit a Bayesian model using \CRANpkg{gemtc}
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_adapt and n_iter are set low to run quickly, but should be left as the
# default values in real use

fitted_bayes_model <- bayes_model(configured_data = configured_data,
                                  n_adapt = 100,
                                  n_iter = 100)

}
