% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_results_f.R
\name{bayes_results}
\alias{bayes_results}
\alias{covariate_results}
\alias{baseline_results}
\title{Summarise a Bayesian model}
\usage{
bayes_results(model, logger = NULL)

covariate_results(...)

baseline_results(...)
}
\arguments{
\item{model}{list. Output produced by \code{baseline_model()}, \code{bayes_model()} or \code{covariate_model()}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{...}{Parameters passed to \code{bayes_results()}}
}
\value{
HTML summary of the model
}
\description{
Produce a table summarising Bayesian models
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_adapt and n_iter are set low to run quickly, but should be left as the
# default values in real use

fitted_bayes_model <- bayes_model(configured_data = configured_data,
                                  n_adapt = 100,
                                  n_iter = 100)

bayes_results(fitted_bayes_model)
}
