% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buttons.R
\name{BayesInch}
\alias{BayesInch}
\title{Calculate the inch height of a forest plot for a given number of treatments}
\usage{
BayesInch(notrt)
}
\arguments{
\item{notrt}{The number of treatments in the plot}
}
\value{
The height of the plot in inches
}
\description{
Calculate the inch height of a forest plot for a given number of treatments
}
