% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buttons.R
\name{BayesPixels}
\alias{BayesPixels}
\title{Calculate the pixel height of a forest plot for a given number of treatments}
\usage{
BayesPixels(notrt, title = FALSE)
}
\arguments{
\item{notrt}{The number of treatments in the plot}

\item{title}{TRUE if the title is included in the plot}
}
\value{
The height of the plot in pixels
}
\description{
Calculate the pixel height of a forest plot for a given number of treatments
}
