% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{BnmaGelmanPlot}
\alias{BnmaGelmanPlot}
\title{Creates a Gelman plot for a BNMA baseline-risk model.}
\usage{
BnmaGelmanPlot(gelman_plot, parameter)
}
\arguments{
\item{gelman_plot}{Output from coda::gelman.plot(bnma_model$samples[, parm]), where parm is a parameter from 'bnma_model'.}

\item{parameter}{The parameter from the previous argument, used as the title.}
}
\value{
Reproduces the Gelman plot mentioned in @param gelman_plot as a plot that can be put in a grid.
}
\description{
Creates a Gelman plot for a BNMA baseline-risk model.
}
