% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{BnmaGelmanPlots}
\alias{BnmaGelmanPlots}
\title{Creates Gelman plots for a BNMA baseline-risk model.}
\usage{
BnmaGelmanPlots(gelman_plots, parameters)
}
\arguments{
\item{gelman_plots}{List of outputs from \code{coda::gelman.plot(bnma_model$samples[, parm])}, where parm is a parameter from bnma_model.}

\item{parameters}{Vector of parameters mentioned in the previous argument.}
}
\value{
Plots the Gelman plots mentioned in @param gelman_plots.
}
\description{
Creates Gelman plots for a BNMA baseline-risk model.
}
