% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{BnmaSwitchRanking}
\alias{BnmaSwitchRanking}
\title{Change the ranking direction in a bnma ranking table.}
\usage{
BnmaSwitchRanking(ranking_table)
}
\arguments{
\item{ranking_table}{The $rank.tx table from output from bnma::network.run()}
}
\value{
A relative effects table in the same format as from gemtc.
}
\description{
Change the ranking direction in a bnma ranking table.
}
