% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{bugsnetdata}
\alias{bugsnetdata}
\title{Reformats the treatment labels then calls dataform.df().}
\usage{
bugsnetdata(data, metaoutcome, treatment_list)
}
\arguments{
\item{data}{Input dataset.}

\item{metaoutcome}{"Continuous" or "Binary".}

\item{treatment_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}
}
\value{
Output from dataform.df().
}
\description{
Reformats the treatment labels then calls dataform.df().
}
