% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scaling.R
\name{calculate_plot_pdf}
\alias{calculate_plot_pdf}
\title{Function to decide inch size of pdf plots based on the number of studies}
\usage{
calculate_plot_pdf(studies)
}
\arguments{
\item{studies}{The number of studies to be plotted (a positive int)}
}
\value{
Height or width in inches for pdf plots (a positive int)
}
\description{
Function to decide inch size of pdf plots based on the number of studies
}
