% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scaling.R
\name{calculate_plot_pixel}
\alias{calculate_plot_pixel}
\title{Function to decide pixel size of plots based on the number of studies}
\usage{
calculate_plot_pixel(studies)
}
\arguments{
\item{studies}{The number of studies to be plotted (a positive int)}
}
\value{
Height or width in pixels for png plots (a positive int)
}
\description{
Function to decide pixel size of plots based on the number of studies
}
