% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CalculateContributions}
\alias{CalculateContributions}
\title{Create the contribution matrix in a convenient format.}
\usage{
CalculateContributions(
  data,
  covariate_title,
  treatment_ids,
  outcome_type,
  outcome_measure,
  effects_type,
  std_dev_d = NULL,
  cov_parameters,
  cov_centre = NULL,
  std_dev_beta = NULL,
  study_or_arm_level,
  absolute_or_percentage,
  basic_or_all_parameters = "basic",
  weight_or_contribution,
  treatment_or_covariate_effect
)
}
\arguments{
\item{data}{Input data in long format.}

\item{covariate_title}{Title of covariate column in data.}

\item{treatment_ids}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively.}

\item{outcome_type}{"Continuous" or "Binary".}

\item{outcome_measure}{"MD", "OR", "RR" or "RD".}

\item{effects_type}{"fixed" or "random".}

\item{std_dev_d}{Between-study standard deviation. Only required when @param effects_type == "random". Defaults to NULL.}

\item{cov_parameters}{Type of regression coefficient. One of: "shared", "exchangeable", or "unrelated".}

\item{cov_centre}{Centring value for the covariate, defaults to the mean.}

\item{std_dev_beta}{Standard deviation of covariate parameters. Only required when @param cov_parameters == "exchangeable". Defaults to NULL.}

\item{study_or_arm_level}{"study" for study-level contributions, "arm" for arm-level contributions.}

\item{absolute_or_percentage}{"percentage" for percentage contributions, "absolute" for absolute contributions.}

\item{basic_or_all_parameters}{"basic" for one column per basic parameter, "all" for one column per parameter. Defaults to "basic".}

\item{weight_or_contribution}{"weight" for coefficients or "contribution" for coefficients multiplied by observed treatment effects.}

\item{treatment_or_covariate_effect}{Whether contributions are for treatment effect or covariate effect. One of: "Treatment Effect", "Covariate Effect".}
}
\value{
List of contributions:
- "direct"
  - Matrix of direct contributions to the regression. Rows are studies, columns are treatments
- "indirect"
  - Matrix of indirect contributions to the regression. Rows are studies, columns are treatments
- "relative_effect"
  - Matrix of relative effects of treatments compared to the reference. Rows are studies, columns are treatments
- "covariate_value"
  - Vector of covariate values from the studies.
}
\description{
Create the contribution matrix in a convenient format.
}
