% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_regression_f.R
\name{CalculateCredibleRegionsBnma}
\alias{CalculateCredibleRegionsBnma}
\title{Calculate the credible regions within direct evidence for the baseline risk model.}
\usage{
CalculateCredibleRegionsBnma(model_output)
}
\arguments{
\item{model_output}{Return from \code{BaselineRiskModelOutput()}.}
}
\value{
list of credible region objects and credible interval objects.
Regions cover treatments with a non-zero covariate range of direct contributions,
intervals cover treatments with a single covariate value from direct contributions.
Any treatment with no direct contributions will not be present in either list.
Each is a list of data frames for each treatment name. Each data frame contains 3 columns:
\itemize{
\item cov_value: The covariate value at which the credible region is calculated.
\item lower: the 2.5\% quantile.
\item upper: the 97.5\% quantile.
Each data frame in "regions" contains 11 rows creating a 10-polygon region.
Each data frame in "intervals" contains a single row at the covariate value of that single contribution.
}
}
\description{
Calculate the credible regions within direct evidence for the baseline risk model.
}
