% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CalculateDirectness}
\alias{CalculateDirectness}
\title{Determine whether comparisons contribute to relative treatment effects directly, indirectly, or not at all.}
\usage{
CalculateDirectness(
  data,
  covariate_title,
  treatment_ids,
  outcome,
  outcome_measure,
  effects_type
)
}
\arguments{
\item{data}{Input data in long format.}

\item{covariate_title}{Title of covariate column in data. Enter NULL if there is no covariate.}

\item{treatment_ids}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively.}

\item{outcome_measure}{"MD", "OR", "RR" or "RD".}

\item{effects_type}{"fixed" or "random".}

\item{outcome_type}{"Continuous" or "Binary".}
}
\value{
List of contributions:
\itemize{
\item "direct"
\itemize{
\item Matrix containing TRUE, FALSE, or NA corresponding to direct, indirect, or no contribution. Rows are studies, columns are treatments
}
\item "relative_effect"
\itemize{
\item Matrix of relative effects of treatments compared to the reference. Rows are studies, columns are treatments
}
\item "covariate_value"
\itemize{
\item Vector of covariate values from the studies.
}
}
}
\description{
Determine whether comparisons contribute to relative treatment effects directly, indirectly, or not at all.
}
