% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_param_classes}
\alias{check_param_classes}
\title{Utility function for checking classes of function parameters}
\usage{
check_param_classes(params, classes, logger)
}
\arguments{
\item{params}{character. Vector of parameters to check}

\item{classes}{character. Vector of classes for each parameter}

\item{logger}{Stores all notification messages to be displayed in the Log
Window or returned as errors.}
}
\value{
\code{TRUE} if any errors are found. \code{FALSE} if not
}
\description{
Utility function for checking classes of function parameters
}
\keyword{internal}
