% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CheckSingularMatrix}
\alias{CheckSingularMatrix}
\title{Check if the given matrix is singular, which would prevent the contribution matrix from being obtainable, and return an error if it is.}
\usage{
CheckSingularMatrix(matrix)
}
\arguments{
\item{matrix}{The matrix that needs to be inverted to obtain the contribution matrix.}
}
\value{
Error if the matrix is singular.
}
\description{
Check if the given matrix is singular, which would prevent the contribution matrix from being obtainable, and return an error if it is.
}
