% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{CleanData}
\alias{CleanData}
\title{Remove leading and trailing whitespace and collapse multiple whitespace characters between words.}
\usage{
CleanData(data)
}
\arguments{
\item{data}{Data frame to clean}
}
\value{
Cleaned data frame
}
\description{
Remove leading and trailing whitespace and collapse multiple whitespace characters between words.
}
