% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{CleanStrings}
\alias{CleanStrings}
\title{Clean strings by replacing all characters that are not a number, letter, or underscore, with an underscore.}
\usage{
CleanStrings(strings)
}
\arguments{
\item{strings}{Vector of the strings to be cleaned.}
}
\value{
Cleaned version of the strings.
}
\description{
Clean strings by replacing all characters that are not a number, letter, or underscore, with an underscore.
}
