% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{CleanStudies}
\alias{CleanStudies}
\title{Clean study names by replacing all characters that are not a number, letter, or underscore, with an underscore.}
\usage{
CleanStudies(data)
}
\arguments{
\item{data}{Data frame with the column 'Study' that contains study names.}
}
\value{
The data frame with 'Study' cleaned and a new column 'RawStudy' which is a copy of the original 'Study'.
}
\description{
Clean study names by replacing all characters that are not a number, letter, or underscore, with an underscore.
}
