% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{CleanTreatmentIds}
\alias{CleanTreatmentIds}
\title{Clean the treatment labels to replace all characters which are not a number, letter, or underscore, with an underscore.}
\usage{
CleanTreatmentIds(treatment_ids)
}
\arguments{
\item{treatment_ids}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively}
}
\value{
Cleaned version of treatment_ids
}
\description{
Clean the treatment labels to replace all characters which are not a number, letter, or underscore, with an underscore.
}
