% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{common_params}
\alias{common_params}
\title{Common parameters}
\arguments{
\item{treatment_df}{dataframe. Treatments}

\item{outcome}{character. Outcome type for the dataset. Either \code{Binary} or
\code{Continuous}.}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{ranking_option}{character. \code{good} if the treatment effect is desirable, else \code{bad}}

\item{freq}{list. NMA results created by freq_wrap().}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{seed}{numeric. Seed used to fit the model}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\description{
Common parameters
}
\keyword{internal}
