% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_analysis.R
\name{contrastform.df}
\alias{contrastform.df}
\title{Puts data in contrast form using netmeta::pairwise().}
\usage{
contrastform.df(wide_data, outcome_measure, outcome)
}
\arguments{
\item{wide_data}{non-covariate data in a wide format}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{outcome}{character. Outcome type for the dataset. Either \code{Binary} or
\code{Continuous}.}
}
\value{
Input data in contrast form.
}
\description{
Puts data in contrast form using netmeta::pairwise().
}
