% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_comparison_f.R
\name{covariate_comparison}
\alias{covariate_comparison}
\title{covariate_comparison}
\usage{
covariate_comparison(...)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- covariate_comparison(1)
}
}
