% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_deviance_f.R
\name{covariate_deviance}
\alias{covariate_deviance}
\title{covariate_deviance}
\usage{
covariate_deviance(...)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- covariate_deviance(1)
}
}
