% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_forest_f.R
\name{covariate_forest}
\alias{covariate_forest}
\title{covariate_forest}
\usage{
covariate_forest(...)
}
\arguments{
\item{...}{parameters passed to \code{bayes_forest()}}
}
\value{
List containing:
\item{svg}{character. SVG code to produce the plot}
\item{height}{numeric. Plot height in pixels}
\item{width}{numeric. Plot width in pixels}
}
\description{
Produce a forest plot from the covariate model. This is
a wrapper for \code{bayes_forest()} - see that for documentation.
}
