% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{covariate_model}
\alias{covariate_model}
\title{Fit a covariate regression model using \CRANpkg{gemtc}.}
\usage{
covariate_model(
  configured_data,
  covariate_value,
  regressor_type,
  covariate_model_output = NULL,
  async = FALSE
)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{covariate_value}{numeric. The value at which to fit the model. Must be greater
than or equal to the minimum value and less than or equal to the maximum
value in \code{configured_data}}

\item{regressor_type}{character. Type of regression coefficient, either \code{shared}, \code{unrelated}, or \code{exchangeable}}

\item{covariate_model_output}{list. The output of the function. Default \code{NULL}.
When supplied, only the output is recalculated for a given covariate value,
rather than refitting the model.}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List of \CRANpkg{gemtc} related output:
\item{mtcResults}{model object from \code{gemtc::mtc.run()} carried through (needed to match existing code)}
\item{mtcRelEffects}{data relating to presenting relative effects}
\item{rel_eff_tbl}{table of relative effects for each comparison}
\item{covariate_value}{The covariate value originally passed into this function}
\item{reference_treatment}{character. The \code{reference_treatment}from \code{configured_data}}
\item{comparator_names}{Vector containing the names of the comparators}
\item{a}{text output stating whether fixed or random effects}
\item{sumresults}{summary output of relative effects}
\item{dic}{data frame of model fit statistics}
\item{cov_value_sentence}{text output stating the value for which the covariate has been set to for producing output}
\item{slopes}{named list of slopes for the regression equations (unstandardised - equal to one 'increment')}
\item{intercepts}{named list of intercepts for the regression equations at covariate_value}
\item{outcome}{character. The \code{outcome} from \code{configured_data}}
\item{outcome_measure}{character. The \code{outcome_measure}from \code{configured_data}}
\item{effects}{character. The \code{effects} from \code{configured_data}}
\item{mtcNetwork}{The network object from GEMTC}
\item{covariate_min}{Vector of minimum covariate values directly contributing to the regression}
\item{covariate_max}{Vector of maximum covariate values directly contributing to the regression}
}
\description{
Fit a covariate regression model using \CRANpkg{gemtc}.
}
