% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{covariate_regression}
\alias{covariate_regression}
\title{Covariate regression data}
\usage{
covariate_regression(model, configured_data, async = FALSE)
}
\arguments{
\item{model}{list. Output created by \code{covariate_model()}}

\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{directness}{list. Output from \code{CalculateDirectness()}}
\item{credible_regions}{list. Output from \code{CalculateCredibleRegions()}}
}
\description{
Generate data from a covariate model required to produce a
metaregression plot
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

fitted_covariate_model <- covariate_model(configured_data = configured_data,
                                          covariate_value = 98,
                                          regressor_type = "shared")

regression_data <- covariate_regression(model = fitted_covariate_model,
                                        configured_data = configured_data)
}
