% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_results_f.R
\name{covariate_results}
\alias{covariate_results}
\title{covariate_results}
\usage{
covariate_results(x, logger = NULL)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- covariate_results(1)  
}
}
