% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_summary_f.R
\name{covariate_summary}
\alias{covariate_summary}
\title{Produce a plot summarising the covariate value for each study arm}
\usage{
covariate_summary(configured_data, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Produce a plot summarising the covariate value for each study arm
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

covariate_summary(configured_data)

}
