% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{CovariateModelOutput}
\alias{CovariateModelOutput}
\title{Function to collate all the model output to be used in other existing functions}
\usage{
CovariateModelOutput(
  connected_data,
  treatment_df,
  model,
  covariate_title,
  cov_value,
  outcome_measure
)
}
\arguments{
\item{connected_data}{Data frame from which model was calculated}

\item{treatment_df}{Data frame containing treatment IDs (Number) and names (Label)}

\item{model}{Completed model object after running RunCovariateRegression()}

\item{covariate_title}{Covariate name as per uploaded data}

\item{cov_value}{Value of covariate for which to give output (default value the mean of study covariates)}

\item{outcome_measure}{The outcome measure for the analysis: One of: "OR", "RR", "MD"}
}
\value{
List of gemtc related output:
mtcResults = model object itself carried through (needed to match existing code)
mtcRelEffects = data relating to presenting relative effects;
rel_eff_tbl = table of relative effects for each comparison;
covariate_value = The covariate value originally passed into this function
reference_name = The name of the reference treatment
comparator_names = Vector containing the names of the comparators.
a = text output stating whether fixed or random effects;
sumresults = summary output of relative effects
dic = data frame of model fit statistics
cov_value_sentence = text output stating the value for which the covariate has been set to for producing output
slopes = named list of slopes for the regression equations (unstandardised - equal to one 'increment')
intercepts = named list of intercepts for the regression equations at cov_value
outcome = The outcome type for the analysis eg. "MD" or "OR"
mtcNetwork = The network object from GEMTC
model = The type of linear model, either "fixed" or "random"
covariate_min = Vector of minimum covariate values directly contributing to the regression.
covariate_max = Vector of maximum covariate values directly contributing to the regression.
}
\description{
Function to collate all the model output to be used in other existing functions
}
