% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buttons.R
\name{create_raw_data_download_handler}
\alias{create_raw_data_download_handler}
\title{Create a download handler for raw data}
\usage{
create_raw_data_download_handler(
  metaoutcome,
  filename,
  continuous_file,
  binary_file
)
}
\arguments{
\item{metaoutcome}{function to get th outcome to decide which file to download}

\item{filename}{Name of file to which to download}

\item{continuous_file}{File to download if continuous}

\item{binary_file}{File to download if binary}
}
\value{
The created download handler
}
\description{
Create a download handler for raw data
}
