% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateCompositeMetaRegressionPlot}
\alias{CreateCompositeMetaRegressionPlot}
\title{Create a composite meta-regression plot which comprises plots showing direct and indirect evidence.}
\usage{
CreateCompositeMetaRegressionPlot(
  model_output,
  treatment_df,
  outcome_measure,
  comparators,
  contribution_matrix,
  contribution_type,
  confidence_regions,
  include_covariate = FALSE,
  include_ghosts = FALSE,
  include_extrapolation = FALSE,
  include_confidence = FALSE,
  confidence_opacity = 0.2,
  include_contributions = TRUE,
  contribution_multiplier = 1,
  legend_position = "BR"
)
}
\arguments{
\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{outcome_measure}{Outcome measure of analysis (OR, RR, RD or MD)}

\item{comparators}{Vector of names of comparison treatments to plot in colour.}

\item{contribution_matrix}{Contributions from function \code{CalculateContributions()}.}

\item{contribution_type}{Type of contribution, used to calculate sizes for the study contribution circles.}

\item{confidence_regions}{List of confidence region data frames from function \code{CalculateConfidenceRegions()}.}

\item{include_covariate}{TRUE if the value of the covariate is to be plotted as a vertical line. Defaults to FALSE.}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot. Defaults to FALSE.}

\item{include_extrapolation}{TRUE if regression lines should be extrapolated beyond the range of the given data. These will appear as dashed lines.
Defaults to FALSE.}

\item{include_confidence}{TRUE if the confidence regions should be plotted for the specified comparators. These will be partially transparent regions.
Defaults to FALSE.}

\item{confidence_opacity}{The opacity of the confidence regions. Can be any value between 0 and 1, inclusive. Defaults to 0.2.}

\item{include_contributions}{TRUE if the contributions should be plotted as a circle for each study. Defaults to TRUE.}

\item{contribution_multiplier}{Factor by which to scale the sizes of the study contribution circles. Defaults to 1.0.}

\item{legend_position}{String informing the position of the legend. Acceptable values are:
\itemize{
\item "BR" - Bottom-right of the plot area
\item "BL" - Bottom-left of the plot area
\item "TR" - Top-right of the plot area
\item "TL" - Top-left of the plot area
}}
}
\value{
Created ggplot2 object.
}
\description{
Create a composite meta-regression plot which comprises plots showing direct and indirect evidence.
}
