% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateContributionMatrix}
\alias{CreateContributionMatrix}
\title{Create the contribution matrix.}
\usage{
CreateContributionMatrix(
  data,
  treatment_ids,
  outcome_type,
  outcome_measure,
  effects_type,
  std_dev_d = NULL,
  cov_parameters,
  cov_centre = NULL,
  std_dev_beta = NULL,
  study_or_arm_level,
  absolute_or_percentage,
  basic_or_all_parameters = "basic",
  weight_or_contribution,
  full_output = FALSE
)
}
\arguments{
\item{data}{Input data in long format.}

\item{treatment_ids}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively.}

\item{outcome_type}{"Continuous" or "Binary".}

\item{outcome_measure}{"MD", "OR", "RR" or "RD".}

\item{effects_type}{"fixed" or "random".}

\item{std_dev_d}{Between-study standard deviation. Only required when @param effects_type == "random". Defaults to NULL.}

\item{cov_parameters}{"shared", "exchangeable", or "unrelated".}

\item{cov_centre}{Centring value for the covariate, defaults to the mean.}

\item{std_dev_beta}{Standard deviation of covariate parameters. Only required when @param cov_parameters == "exchangeable". Defaults to NULL.}

\item{study_or_arm_level}{"study" for arm-level contributions, "comparison" for basic-comparison-level contributions.}

\item{absolute_or_percentage}{"percentage" for percentage contributions, "absolute" for absolute contributions.}

\item{basic_or_all_parameters}{"basic" for one column per basic parameter, "all" for one column per parameter. Defaults to "basic".}

\item{weight_or_contribution}{"weight" for coefficients or "contribution" for coefficients multiplied by observed treatment effects.}

\item{full_output}{TRUE or FALSE, defaults to FALSE. See @return for details.}
}
\value{
If @param full_output = FALSE:
          The contribution matrix.
        If @param full_output = TRUE:
          List
           - 'contribution' = contribution matrix.
           - 'V' = The treatment effect variance matrix.
           - 'X' = The design matrix.
           - 'Z' = The Z matrix.
           - 'Lambda_tau' = The Lambda_tau matrix (only included if @param effects_type == "random").
           - 'Lambda_beta' = The Lambda_beta matrix (only included if @param cov_parameters == "exchangeable").
}
\description{
Create the contribution matrix.
}
