% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CreateForestPlot}
\alias{CreateForestPlot}
\title{3a Bayesian forest plot}
\usage{
CreateForestPlot(model, metaoutcome, bayesmin, bayesmax)
}
\arguments{
\item{model}{Various model output created by baye().}

\item{metaoutcome}{"Continuous" or "Binary".}

\item{bayesmin}{x-axis limit minimum.}

\item{bayesmax}{x-axis limit maximum.}
}
\value{
Forest plot created by gemtc::forest().
}
\description{
3a Bayesian forest plot
}
