% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{CreateGemtcModel}
\alias{CreateGemtcModel}
\title{Function for setting up covariate \CRANpkg{gemtc} model object}
\usage{
CreateGemtcModel(
  data,
  model_type,
  outcome_measure,
  regressor_type,
  ref_choice,
  seed
)
}
\arguments{
\item{data}{list containing armData and studyData (as created by PrepDataGemtc)}

\item{model_type}{Whether the model is 'fixed' or 'random'}

\item{outcome_measure}{Type of outcome measure ('OR', 'RR', or 'MD')}

\item{regressor_type}{Type of regression coefficient, either "shared", "unrelated", or "exchangeable"}

\item{ref_choice}{The choice of reference treatment as selected by user}

\item{seed}{Seed value to use when fitting model}
}
\value{
An object of class mtc.model
}
\description{
Function for setting up covariate \CRANpkg{gemtc} model object
}
