% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_networks.R
\name{CreateGraph}
\alias{CreateGraph}
\title{Create a graph where the nodes are treatments and the links are studies comparing those treatments.}
\usage{
CreateGraph(data)
}
\arguments{
\item{data}{Data for which to create graph.}
}
\value{
Created igraph object.
}
\description{
Create a graph where the nodes are treatments and the links are studies comparing those treatments.
}
