% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateIndirectContributionPlot}
\alias{CreateIndirectContributionPlot}
\title{Create a covariate regression plot where multiple comparisons can be plotted, and the contributions from each study are shown as circles.}
\usage{
CreateIndirectContributionPlot(
  model_output,
  treatment_df,
  comparators,
  contribution_matrix,
  contribution_type,
  include_covariate = FALSE,
  include_ghosts = FALSE,
  contribution_multiplier = 1
)
}
\arguments{
\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{comparators}{Vector of names of comparison treatments to plot in colour.}

\item{contribution_matrix}{Contributions from function \code{CalculateContributions()}.}

\item{contribution_type}{Type of contribution, used to calculate sizes for the study contribution circles.}

\item{include_covariate}{TRUE if the value of the covariate is to be plotted as a vertical line. Defaults to FALSE.}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot. Defaults to FALSE.}

\item{contribution_multiplier}{Factor by which to scale the sizes of the study contribution circles. Defaults to 1.0.}
}
\value{
Created ggplot2 object.
}
\description{
Create a covariate regression plot where multiple comparisons can be plotted, and the contributions from each study are shown as circles.
}
