% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregression_plot_f.R
\name{CreateIndirectCovariatePlot}
\alias{CreateIndirectCovariatePlot}
\title{Create a covariate regression plot where multiple comparisons can be plotted, and the contributions from each study are shown as circles.}
\usage{
CreateIndirectCovariatePlot(
  model_output,
  treatment_df,
  comparators,
  directness,
  include_covariate = FALSE,
  include_ghosts = FALSE,
  covariate_symbol = "circle open",
  covariate_symbol_size = 10
)
}
\arguments{
\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{comparators}{Vector of names of comparison treatments to plot in colour.}

\item{directness}{Contributions from function \code{CalculateDirectness()}.}

\item{include_covariate}{TRUE if the value of the covariate is to be plotted as a vertical line. Defaults to FALSE.}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot. Defaults to FALSE.}

\item{covariate_symbol}{The selected symbol for displaying covariates. Defaults to "circle open".}

\item{covariate_symbol_size}{Size of the covariate symbols. Defaults to 10.}
}
\value{
Created ggplot2 object.
}
\description{
Create a covariate regression plot where multiple comparisons can be plotted, and the contributions from each study are shown as circles.
}
