% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateLambdaBetaMatrix}
\alias{CreateLambdaBetaMatrix}
\title{Create the matrix of variances for exchangeable covariate parameters.}
\usage{
CreateLambdaBetaMatrix(treatments, std_dev_beta)
}
\arguments{
\item{treatments}{Vector of treatments with the reference treatment first.}

\item{std_dev_beta}{Standard deviation of covariate parameters.}
}
\value{
Lambda_beta matrix.
}
\description{
Create the matrix of variances for exchangeable covariate parameters.
}
