% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateLambdaTauMatrix}
\alias{CreateLambdaTauMatrix}
\title{Create the matrix of random effects variances.}
\usage{
CreateLambdaTauMatrix(data, studies, treatments, std_dev_d)
}
\arguments{
\item{data}{Input data in long format.}

\item{studies}{Vector of studies.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{std_dev_d}{Between-study standard deviation.}
}
\value{
Lambda_tau matrix.
}
\description{
Create the matrix of random effects variances.
}
