% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{CreateListOfWideColumns}
\alias{CreateListOfWideColumns}
\title{Create a list of all columns starting with the given column prefix.}
\usage{
CreateListOfWideColumns(wide_data, column_prefix)
}
\arguments{
\item{wide_data}{Data in wide format.}

\item{column_prefix}{String containing the opening letters of one or more columns in 'wide_data'.}
}
\value{
List with one element per selected column in 'wide_data', with each element named by the column name.
}
\description{
Create a list of all columns starting with the given column prefix.
}
