% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_plot_f.R
\name{CreateMainRegressionPlot}
\alias{CreateMainRegressionPlot}
\title{Create a covariate regression plot where multiple comparisons can be plotted, and the contributions from each study are shown as circles.}
\usage{
CreateMainRegressionPlot(
  model_output,
  treatment_df,
  outcome_measure,
  comparators,
  directness,
  credible_regions,
  include_covariate = FALSE,
  include_ghosts = FALSE,
  include_extrapolation = FALSE,
  include_credible = FALSE,
  credible_opacity = 0.2,
  covariate_symbol = "circle open",
  covariate_symbol_size = 10,
  legend_position = "BR"
)
}
\arguments{
\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{outcome_measure}{Outcome measure of analysis (OR, RR, RD, MD)}

\item{comparators}{Vector of names of comparison treatments to plot in colour.}

\item{directness}{Contributions from function \code{CalculateDirectness()}.}

\item{credible_regions}{List of credible region data frames from function \code{CalculateCredibleRegions()}.}

\item{include_covariate}{TRUE if the value of the covariate is to be plotted as a vertical line. Defaults to FALSE.}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot. Defaults to FALSE.}

\item{include_extrapolation}{TRUE if regression lines should be extrapolated beyond the range of the given data. These will appear as dashed lines.
Defaults to FALSE.}

\item{include_credible}{TRUE if the credible regions should be plotted for the specified comparators. These will be partially transparent regions.
Defaults to FALSE.}

\item{credible_opacity}{The opacity of the credible regions. Can be any value between 0 and 1, inclusive. Defaults to 0.2.}

\item{covariate_symbol}{The selected symbol for displaying covariates. Defaults to "circle open".}

\item{covariate_symbol_size}{Size of the covariate symbols. Defaults to 10.}

\item{legend_position}{String informing the position of the legend. Acceptable values are:
\itemize{
\item "BR" - Bottom-right of the plot area
\item "BL" - Bottom-left of the plot area
\item "TR" - Top-right of the plot area
\item "TL" - Top-left of the plot area
Defaults to "BR"
}}
}
\value{
Created ggplot2 object.
}
\description{
Create a covariate regression plot where multiple comparisons can be plotted, and the contributions from each study are shown as circles.
}
