% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{CreateTreatmentIds}
\alias{CreateTreatmentIds}
\title{Create a data frame with treatment IDs and treatment names.}
\usage{
CreateTreatmentIds(all_treatments, reference_treatment = all_treatments[1])
}
\arguments{
\item{all_treatments}{Vector of all treatment names}

\item{reference_treatment}{Name of treatment to be assigned ID 1}
}
\value{
Data frame containing the treatment ID ('Number') and the treatment name ('Label')
}
\description{
Create a data frame with treatment IDs and treatment names.
}
