% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update_f.R
\name{CreateTreatmentsDataFrame}
\alias{CreateTreatmentsDataFrame}
\title{Create data frame containing treatment IDs and names.}
\usage{
CreateTreatmentsDataFrame(treatment_names_string)
}
\arguments{
\item{treatment_names_string}{String containing treatment names, separated by commas}
}
\value{
Data frame with 'Number' column defining treatment IDs and 'Label' column defining treatment names
}
\description{
Create data frame containing treatment IDs and names.
}
