% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateVMatrix}
\alias{CreateVMatrix}
\title{Create the variance-covariance matrix of outcomes.}
\usage{
CreateVMatrix(data, studies, treatments, outcome_type, outcome_measure)
}
\arguments{
\item{data}{Input data in long format plus the column 'Treatment', a text version of 'T'.}

\item{studies}{Vector of studies.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{outcome_type}{"Continuous" or "Binary".}

\item{outcome_measure}{"MD" (when outcome_type == "Continuous"), "OR", "RR" or "RD" (when outcome_type == "Binary")}
}
\value{
V matrix.
}
\description{
Create the variance-covariance matrix of outcomes.
}
