% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateXMatrix}
\alias{CreateXMatrix}
\title{Create the design matrix.}
\usage{
CreateXMatrix(data, studies, treatments, covar_centred, cov_parameters)
}
\arguments{
\item{data}{Input data in long format plus the column 'Treatment', a text version of 'T'.}

\item{studies}{Vector of studies.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{covar_centred}{Vector of centred covariate values, named by study.}

\item{cov_parameters}{"shared", "exchangeable", or "unrelated".}
}
\value{
X matrix.
}
\description{
Create the design matrix.
}
