% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{CreateZMatrix}
\alias{CreateZMatrix}
\title{Create the Z matrix.}
\usage{
CreateZMatrix(studies, treatments, cov_parameters)
}
\arguments{
\item{studies}{Vector of studies.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{cov_parameters}{"shared", "exchangeable", or "unrelated".}
}
\value{
Z matrix.
}
\description{
Create the Z matrix.
}
