% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crop_svg}
\alias{crop_svg}
\title{Crops an svg object, by rendering it, locating the non-white pixels and
editing the svg viewBox property. The width and height are also returned
to facilitate rendering to other formats.}
\usage{
crop_svg(svg, margin = 10)
}
\arguments{
\item{svg}{xml_document. Output from \code{svglite::xmlSVG()}}

\item{margin}{numeric. The margin in pixels to leave around the edge
of the plot content. Defaults to 10.}
}
\value{
html. The cropped svg
}
\description{
Crops an svg object, by rendering it, locating the non-white pixels and
editing the svg viewBox property. The width and height are also returned
to facilitate rendering to other formats.
}
