% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_options_panel.R
\name{data_analysis_options_panel_ui}
\alias{data_analysis_options_panel_ui}
\title{Module UI for the data analysis options panel}
\usage{
data_analysis_options_panel_ui(id)
}
\arguments{
\item{id}{ID of the module}
}
\value{
Div for the panel
}
\description{
Module UI for the data analysis options panel
}
