% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_options_panel.R
\name{data_analysis_options_server}
\alias{data_analysis_options_server}
\title{Module server for the data analysis options panel}
\usage{
data_analysis_options_server(
  id,
  data,
  is_default_data,
  treatment_df,
  metaoutcome,
  OpenDataTable
)
}
\arguments{
\item{id}{ID of the module}

\item{data}{Reactive containing data to analyse}

\item{is_default_data}{Reactive containing TRUE if data is an example dataset, loaded by default}

\item{treatment_df}{Reactive containing data frame containing treatment IDs (Number) and names (Label)}

\item{metaoutcome}{Reactive containing meta analysis outcome: "continuous" or "binary"}

\item{OpenDataTable}{Function to open the data table}
}
\value{
List of reactives:
\itemize{
\item "outcome_measure" contains the outcome measure being analysed.
This will be related to the outcome in "metaoutcome"
\item "model_effects" contains "random" or "fixed"
\item "exclusions" contains vector of names of studies being excluded from the sensitivity analysis
\item "rank_option" contains "good" or "bad" for whether a small value is desirable or not
\item "continuous_outcome" contains acronym of the continuous outcome:
"MD" for mean difference, or "SMD" for standardised mean difference
\item "binary_outcome" contains acronym of the binary outcome:
"OR" for odds ratio, "RR" for risk ratio, or "RD" for risk difference
\item "initial_data" contains the data frame of the connected subnetwork of the uploaded data
\item "sensitivity_data" contains the data frame of the connected subnetwork for the sensitivity analysis
\item "sensitivity_treatment_list" contains the treatment names ("Label") and IDs ("Number") for the sensitivity analysis data
}
}
\description{
Module server for the data analysis options panel
}
