% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_page.R
\name{data_analysis_page_server}
\alias{data_analysis_page_server}
\title{Module server for the data analysis page.}
\usage{
data_analysis_page_server(id, data, is_default_data, treatment_df, metaoutcome)
}
\arguments{
\item{id}{ID of the module}

\item{data}{Reactive containing data to analyse}

\item{is_default_data}{Reactive containing TRUE if data is an example dataset, loaded by default}

\item{treatment_df}{Reactive containing data frame containing treatment IDs (Number) and names (Label)}

\item{metaoutcome}{Reactive containing meta analysis outcome: "continuous" or "binary"}
}
\description{
Module server for the data analysis page.
}
