% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_page.R
\name{data_analysis_page_ui}
\alias{data_analysis_page_ui}
\title{Module UI for the data analysis page.}
\usage{
data_analysis_page_ui(id)
}
\arguments{
\item{id}{ID of the module}
}
\value{
Div for the data analysis page
}
\description{
Module UI for the data analysis page.
}
