% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_panel.R
\name{data_summary_panel_server}
\alias{data_summary_panel_server}
\title{Module server for the data summary panel.}
\usage{
data_summary_panel_server(
  id,
  metaoutcome,
  outcome_measure,
  bugsnetdt,
  bugsnetdt_sub,
  freq_all,
  freq_sub
)
}
\arguments{
\item{id}{ID of the module}

\item{metaoutcome}{Reactive containing meta analysis outcome: "Continuous" or "Binary"}

\item{outcome_measure}{Reactive containing meta analysis outcome measure: "MD", "SMD", "OR, "RR", or "RD"}

\item{bugsnetdt}{Reactive containing bugsnet meta-analysis}

\item{bugsnetdt_sub}{Reactive containing bugsnet meta-analysis for sensitivity analysis}

\item{freq_all}{Reactive containing frequentist meta-analysis}

\item{freq_sub}{Reactive containing frequentist meta-analysis for the sensitivity analysis}

\item{data}{Reactive containing data to analyse}
}
\description{
Module server for the data summary panel.
}
