% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary_panel.R
\name{data_summary_panel_ui}
\alias{data_summary_panel_ui}
\title{Module UI for the data summary panel.}
\usage{
data_summary_panel_ui(id, page_numbering)
}
\arguments{
\item{id}{ID of the module}
}
\value{
Div for the panel
}
\description{
Module UI for the data summary panel.
}
