% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_generic.R
\name{dataform.df}
\alias{dataform.df}
\title{Converts wide data to long, adds the variable 'se' for a continuous outcomes, and does some formatting.
(TM: to be modified to use WideToLong() when it becomes available from meta-regression.)}
\usage{
dataform.df(newData1, treat_list, CONBI)
}
\arguments{
\item{newData1}{Input dataset.}

\item{treat_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}

\item{CONBI}{"Continuous" or "Binary".}
}
\value{
Input data set in long format with the variable 'se' for a continuous outcome.
}
\description{
Converts wide data to long, adds the variable 'se' for a continuous outcomes, and does some formatting.
(TM: to be modified to use WideToLong() when it becomes available from meta-regression.)
}
