% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{DensityPlots}
\alias{DensityPlots}
\title{Creates posterior density plots of MCMC samples.}
\usage{
DensityPlots(model, parameters)
}
\arguments{
\item{model}{Model output.}

\item{parameters}{Vector of parameters to create density plots for.}
}
\value{
List of ggplot density plots.
}
\description{
Creates posterior density plots of MCMC samples.
}
