% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dic_table}
\alias{dic_table}
\title{Create a summary table of deviance information criterion stats for Bayesian models}
\usage{
dic_table(dic, analysis = "all")
}
\arguments{
\item{dic}{dataframe of DIC stats from \code{baseline_model}, \code{bayes_model()}
or \code{covariate_model()}}

\item{analysis}{Whether the analysis is using all studies (\code{all}) or a subset (\code{sub})}
}
\description{
Create a summary table of deviance information criterion stats for Bayesian models
}
