% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.AddRiskOfBias}
\alias{.AddRiskOfBias}
\title{Add risk of bias circles for a single comparison in a single study.}
\usage{
.AddRiskOfBias(
  pairwise,
  outcome_type,
  x_rob_positions,
  y_position,
  studlab,
  treatment1,
  treatment2,
  rob_variables
)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}

\item{outcome_type}{"binary" or "continuous".}

\item{x_rob_positions}{Vector of x co-ordinates for the circles.}

\item{y_position}{y co-ordinate for the circles.}

\item{studlab}{Study.}

\item{treatment1}{First treatment in the comparison.}

\item{treatment2}{Second treatment in the comparison.}

\item{rob_variables}{Vector of RoB and indirectness variable names from 'pairwise'.}
}
\description{
Add risk of bias circles for a single comparison in a single study.
}
