% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_deviance_f.R
\name{.AddTraceToUmePlot}
\alias{.AddTraceToUmePlot}
\title{Apply plotly::add_trace to the UME plot when there are more than two arms.}
\usage{
.AddTraceToUmePlot(ume_plot, all, arm_number)
}
\arguments{
\item{ume_plot}{The current UME plot.}

\item{all}{The full data frame of residual deviances.}

\item{arm_number}{Current arm number.}
}
\value{
'ume_plot' with additional points added, corresponding to 'arm_number'.
}
\description{
Apply plotly::add_trace to the UME plot when there are more than two arms.
}
