% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.AddTreatmentEffect}
\alias{.AddTreatmentEffect}
\title{Add the treatment effect line and confidence interval for a single comparison in a single study.}
\usage{
.AddTreatmentEffect(
  pairwise,
  y_position,
  studlab,
  treatment1,
  treatment2,
  ci_limit_height = 0.3
)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}

\item{y_position}{y co-ordinate for the study label and confidence interval.}

\item{studlab}{Study.}

\item{treatment1}{First treatment in the comparison.}

\item{treatment2}{Second treatment in the comparison.}

\item{ci_limit_height}{Height of the vertical lines at the end of the confidence intervals. Defaults to 0.3.}
}
\description{
Add the treatment effect line and confidence interval for a single comparison in a single study.
}
