% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.AddTreatmentEffectBlock}
\alias{.AddTreatmentEffectBlock}
\title{Add the treatment effect lines, confidence intervals, studies, and comparison header, for a single comparison.}
\usage{
.AddTreatmentEffectBlock(
  pairwise,
  outcome_type,
  x_label_position,
  x_outcome_position,
  x_rob_positions,
  y_header_position,
  y_positions,
  treatment1,
  treatment2,
  rob_variables
)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}

\item{outcome_type}{"binary" or "continuous".}

\item{x_label_position}{x co-ordinate for the header and study labels.}

\item{x_outcome_position}{x co-ordinate for the comparison-within-study outcomes.}

\item{x_rob_positions}{Vector of x co-ordinates for the circles.}

\item{y_header_position}{y co-ordinate for the header.}

\item{y_positions}{Vector of y co-ordinates for the study labels and confidence intervals.}

\item{treatment1}{First treatment in the comparison.}

\item{treatment2}{Second treatment in the comparison.}

\item{rob_variables}{Vector of RoB and indirectness variable names from 'pairwise'.}
}
\description{
Add the treatment effect lines, confidence intervals, studies, and comparison header, for a single comparison.
}
