% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.binaryOrder}
\alias{.binaryOrder}
\title{The column order for a binary outcome, for wide and long data.}
\usage{
.binaryOrder(max_arms)
}
\arguments{
\item{max_arms}{Maximum number of arms in the data set.}
}
\value{
Vector of column names in order.
}
\description{
The column order for a binary outcome, for wide and long data.
}
