% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{.CalculateSideif}
\alias{.CalculateSideif}
\title{Creates a named list containing SIDE inconsistency factor results for a comparison. The results are point estimate, confidence interval limits, Z-value and p-value.}
\usage{
.CalculateSideif(
  direct,
  indirect,
  direct_lower,
  direct_upper,
  indirect_lower,
  indirect_upper
)
}
\arguments{
\item{direct}{Direct treatment effect estimate.}

\item{indirect}{Indirect treatment effect estimate.}

\item{direct_lower}{Lower 95\% CI limit of direct treatment effect.}

\item{direct_upper}{Upper 95\% CI limit of direct treatment effect.}

\item{indirect_lower}{Lower 95\% CI limit of indirect treatment effect.}

\item{indirect_upper}{Upper 95\% CI limit of indirect treatment effect.}
}
\value{
List of SIDE IF results with the described structure.
}
\description{
Creates a named list containing SIDE inconsistency factor results for a comparison. The results are point estimate, confidence interval limits, Z-value and p-value.
}
