% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{.CorrectColumnName}
\alias{.CorrectColumnName}
\title{Correct a column name to match the expected letter casing.}
\usage{
.CorrectColumnName(original_name, column_names)
}
\arguments{
\item{original_name}{Column name to fix}

\item{column_names}{Named vector where each name is a regular expression to match, and the value is the replacement string.}
}
\value{
The corrected column name.
}
\description{
Correct a column name to match the expected letter casing.
}
