% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.FindConfidenceInterval}
\alias{.FindConfidenceInterval}
\title{Find the confidence interval at a given covariate value.}
\usage{
.FindConfidenceInterval(mtc_results, reference_name, cov_value, parameter_name)
}
\arguments{
\item{mtc_results}{Meta-analysis object from which to find confidence interval.}

\item{reference_name}{Name of reference treatment.}

\item{cov_value}{Covariate value at which to find the confidence interval.}

\item{parameter_name}{Name of the parameter for which to get the confidence interval.}
}
\value{
Named vector of "2.5%" and "97.5" quantiles.
}
\description{
Find the confidence interval at a given covariate value.
}
