% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{.FindConfidenceIntervalBnma}
\alias{.FindConfidenceIntervalBnma}
\title{Find the confidence interval at a given covariate value.}
\usage{
.FindConfidenceIntervalBnma(mtc_results, cov_value, parameter_name)
}
\arguments{
\item{mtc_results}{Meta-analysis object from which to find confidence interval.}

\item{cov_value}{Covariate value at which to find the confidence interval.}

\item{parameter_name}{Name of the parameter for which to get the confidence interval.}

\item{reference_name}{Name of reference treatment.}
}
\value{
Named vector of "2.5\%" and "97.5" quantiles.
}
\description{
Find the confidence interval at a given covariate value.
}
