% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.FindCredibleInterval}
\alias{.FindCredibleInterval}
\title{Find the credible interval at a given covariate value.}
\usage{
.FindCredibleInterval(
  mtc_results,
  reference_treatment,
  cov_value,
  parameter_name
)
}
\arguments{
\item{mtc_results}{Meta-analysis object from which to find credible interval.}

\item{reference_treatment}{Name of reference treatment.}

\item{cov_value}{Covariate value at which to find the credible interval.}

\item{parameter_name}{Name of the parameter for which to get the credible interval.}
}
\value{
Named vector of "2.5\%" and "97.5" quantiles.
}
\description{
Find the credible interval at a given covariate value.
}
