% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_regression_f.R
\name{.FindCredibleIntervalBnma}
\alias{.FindCredibleIntervalBnma}
\title{Find the credible interval at a given covariate value.}
\usage{
.FindCredibleIntervalBnma(mtc_results, cov_value, parameter_name)
}
\arguments{
\item{mtc_results}{Meta-analysis object from which to find credible interval.}

\item{cov_value}{Covariate value at which to find the credible interval.}

\item{parameter_name}{Name of the parameter for which to get the credible interval.}

\item{reference_name}{Name of reference treatment.}
}
\value{
Named vector of "2.5\%" and "97.5" quantiles.
}
\description{
Find the credible interval at a given covariate value.
}
