% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregression_plot_f.R
\name{.FindDirectRegressionContributions}
\alias{.FindDirectRegressionContributions}
\title{Find the contributions to the regression analysis.}
\usage{
.FindDirectRegressionContributions(
  model_output,
  reference,
  comparator,
  directness
)
}
\arguments{
\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{reference}{Name of reference treatment.}

\item{comparator}{Name of comparison treatment for which to find the contributions.}

\item{directness}{Contributions from function \code{CalculateDirectness()}.}
}
\value{
Data frame containing contribution details. Each row represents a study contributing to a given treatment. Columns are:
\itemize{
\item Treatment: The treatment for which this contribution relates.
\item covariate_value: Value of the covariate for this study.
\item relative_effect Relative effect for this study.
\item contribution: Size of contribution for this study.
}
}
\description{
Find the contributions to the regression analysis.
}
