% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.FindIndirectRegressionContributions}
\alias{.FindIndirectRegressionContributions}
\title{Find the contributions to the regression analysis.}
\usage{
.FindIndirectRegressionContributions(
  model_output,
  reference,
  comparator,
  contribution_matrix,
  contribution_type
)
}
\arguments{
\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{reference}{Name of reference treatment.}

\item{comparator}{Name of comparison treatment for which to find the contributions.}

\item{contribution_type}{Type of contribution to find.}
}
\value{
Data frame containing contribution details. Each row represents a study contributing to a given treatment. Columns are:
\itemize{
\item Treatment: The treatment for which this contribution relates.
\item covariate_value: Value of the covariate for this study.
\item contribution: Size of contribution for this study.
}
}
\description{
Find the contributions to the regression analysis.
}
