% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.FindMinMax}
\alias{.FindMinMax}
\title{Find the default minimum and maximum values for the x-axis and calculate
a sensible step to use in the numeric input}
\usage{
.FindMinMax(pairwise)
}
\arguments{
\item{pairwise}{Results of pairwise analysis}
}
\value{
List containing:
\item{x_min}{numeric. Minimum value for the x-axis}
\item{x_max}{numeric. Maximum value for the x-axis}
\item{step}{numeric. Step value to use in numericInput}
}
\description{
Find the default minimum and maximum values for the x-axis and calculate
a sensible step to use in the numeric input
}
