% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{.FindMissingColumns}
\alias{.FindMissingColumns}
\title{Find the names of any required columns which are missing from the data frame.}
\usage{
.FindMissingColumns(data, required_columns)
}
\arguments{
\item{data}{Data frame in which to find missing columns.}

\item{required_columns}{Data frame defining required columns. Columns in this data frame are:
- pattern = A regular expression defining the generic column title for wide or long format.
- name = The generic name of the column.}
}
\value{
Vector containing names of any columns which are required, but are missing.
}
\description{
Find the names of any required columns which are missing from the data frame.
}
