% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.FindMistypedColumns}
\alias{.FindMistypedColumns}
\title{Identify any columns which are not of the expected type.}
\usage{
.FindMistypedColumns(data, outcome_columns)
}
\arguments{
\item{data}{Data frame to validate.}

\item{outcome_columns}{Data frame containing data definitions.}
}
\value{
A vector of column titles which contain data of the wrong type.
}
\description{
Identify any columns which are not of the expected type.
}
