% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_generic.R
\name{.FindTreatmentCount}
\alias{.FindTreatmentCount}
\title{Find the number of treatment columns in a wide data frame. This assumes that all columns are numbered sequentially, starting at 1.}
\usage{
.FindTreatmentCount(wide_data)
}
\arguments{
\item{wide_data}{Data frame in which to find treatment column count.}
}
\value{
The number of treatment columns.
}
\description{
Find the number of treatment columns in a wide data frame. This assumes that all columns are numbered sequentially, starting at 1.
}
