% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_generic.R
\name{.FindVaryingColumnIndices}
\alias{.FindVaryingColumnIndices}
\title{Find the titles of all varying column to be reshaped.}
\usage{
.FindVaryingColumnIndices(wide_data)
}
\arguments{
\item{wide_data}{Data frame win which to find varying column names.}
}
\value{
Vector of titles of all numbered columns.
}
\description{
Find the titles of all varying column to be reshaped.
}
