% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.FindXTicks}
\alias{.FindXTicks}
\title{Finds suitable tick positions for the x-axis.}
\usage{
.FindXTicks(lower, upper)
}
\arguments{
\item{lower}{x-axis minimum.}

\item{upper}{x-axis maximum.}
}
\value{
Vector of x-axis ticks.
}
\description{
Finds suitable tick positions for the x-axis.
}
