% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{.FixColumnNameCases}
\alias{.FixColumnNameCases}
\title{Rename the columns of a data frame to match the expected letter casing.}
\usage{
.FixColumnNameCases(data, outcome_type)
}
\arguments{
\item{data}{Data frame to fix}

\item{outcome_type}{Type of outcome for which to reorder, either 'Continuous' or 'Binary'}
}
\value{
Data frame with renamed columns.
}
\description{
Rename the columns of a data frame to match the expected letter casing.
}
