% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.FormatRegressionConfidenceRegion}
\alias{.FormatRegressionConfidenceRegion}
\title{Format the confidence regions for the regression analysis into a plottable data frame.}
\usage{
.FormatRegressionConfidenceRegion(confidence_regions, comparator)
}
\arguments{
\item{confidence_regions}{List of confidence region data frames from function \code{CalculateConfidenceRegions()}.}

\item{comparator}{Name of comparison treatment for which to find the contributions.}
}
\value{
Data frame containing contribution details. Each row represents a confidence interval at a specific covariate value, for a given treatment. Columns are:
\itemize{
\item Treatment: The treatment for which this confidence interval relates.
\item covariate_value: Value of the covariate for this interval
\item y_min Relative effect of the lower end of this interval
\item y_max: Relative effect of the upper end of this interval
}
}
\description{
Format the confidence regions for the regression analysis into a plottable data frame.
}
