% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregression_plot_f.R
\name{.FormatRegressionCredibleRegion}
\alias{.FormatRegressionCredibleRegion}
\title{Format the credible regions for the regression analysis into a plottable data frame.}
\usage{
.FormatRegressionCredibleRegion(credible_regions, comparator)
}
\arguments{
\item{credible_regions}{List of credible region data frames from function \code{CalculateCredibleRegions()}.}

\item{comparator}{Name of comparison treatment for which to find the contributions.}
}
\value{
Data frame containing contribution details. Each row represents a credible interval at a specific covariate value, for a given treatment. Columns are:
\itemize{
\item Treatment: The treatment for which this credible interval relates.
\item covariate_value: Value of the covariate for this interval
\item y_min Relative effect of the lower end of this interval
\item y_max: Relative effect of the upper end of this interval
}
}
\description{
Format the credible regions for the regression analysis into a plottable data frame.
}
