% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.PlotConfidenceRegions}
\alias{.PlotConfidenceRegions}
\title{Plot the confidence regions and intervals on the plot.}
\usage{
.PlotConfidenceRegions(
  plot,
  confidence_regions,
  comparators,
  confidence_opacity
)
}
\arguments{
\item{plot}{object to which to add elements.}

\item{confidence_regions}{List of confidence region data frames from function \code{CalculateConfidenceRegions()}.}

\item{comparators}{Vector of names of comparison treatments to plot.}
}
\value{
The modified ggplot2 object.
}
\description{
Plot the confidence regions and intervals on the plot.
}
