% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregression_plot_f.R
\name{.PlotCredibleRegions}
\alias{.PlotCredibleRegions}
\title{Plot the credible regions and intervals on the plot.}
\usage{
.PlotCredibleRegions(
  plot,
  credible_regions,
  comparators,
  credible_opacity = 0.2
)
}
\arguments{
\item{plot}{object to which to add elements.}

\item{credible_regions}{List of credible region data frames from function \code{CalculateCredibleRegions()}.}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{credible_opacity}{The opacity of the credible regions. Can be any value between 0 and 1, inclusive. Defaults to 0.2.}
}
\value{
The modified ggplot2 object.
}
\description{
Plot the credible regions and intervals on the plot.
}
