% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.PlotDirectContributionCircles}
\alias{.PlotDirectContributionCircles}
\title{Plot the contribution circles for direct evidence on the plot.}
\usage{
.PlotDirectContributionCircles(
  plot,
  model_output,
  treatment_df,
  reference,
  comparators,
  contribution_matrix,
  contribution_type,
  contribution_multiplier,
  ghosted = FALSE
)
}
\arguments{
\item{plot}{ggplot2 object to which to add elements.}

\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{reference}{Name of reference treatment.}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{contribution_matrix}{Contributions from function \code{CalculateContributions()}.}

\item{contribution_type}{Name of the type of contribution, used to calculate sizes for the study contribution circles.}

\item{contribution_multiplier}{Multiplication factor by which to scale the sizes of the study contribution circles. Defaults to 1.0.}

\item{ghosted}{TRUE if studies should be plotted in grey. Defaults to FALSE.}
}
\value{
The modified ggplot2 object.
}
\description{
Plot the contribution circles for direct evidence on the plot.
}
