% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregression_plot_f.R
\name{.PlotIndirectCovariateCircles}
\alias{.PlotIndirectCovariateCircles}
\title{Plot the contribution circles on the plot.}
\usage{
.PlotIndirectCovariateCircles(
  plot,
  model_output,
  treatment_df,
  reference,
  comparators,
  directness,
  covariate_symbol = "circle open",
  covariate_symbol_size = 10,
  ghosted = FALSE
)
}
\arguments{
\item{plot}{object to which to add elements.}

\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{reference}{Name of reference treatment.}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{directness}{Contributions from function \code{CalculateDirectness()}.}

\item{covariate_symbol}{The selected symbol for displaying covariates. Defaults to "circle open".}

\item{covariate_symbol_size}{Size of the covariate symbols. Defaults to 10.}

\item{ghosted}{TRUE if studies should be plotted in grey. Defaults to FALSE.}
}
\value{
The modified ggplot2 object.
}
\description{
Plot the contribution circles on the plot.
}
