% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_plot_f.R
\name{.PlotRegressionLines}
\alias{.PlotRegressionLines}
\title{Plot the regression lines on the plot.}
\usage{
.PlotRegressionLines(
  plot,
  model_output,
  treatment_df,
  reference,
  comparators,
  extrapolate,
  ghosted = FALSE
)
}
\arguments{
\item{plot}{ggplot2 object to which to add elements.}

\item{model_output}{GEMTC model results found by calling \code{CovariateModelOutput()}.}

\item{treatment_df}{Data frame containing treatment IDs (Number), sanitised names (Label), and original names (RawLabel).}

\item{reference}{Name of reference treatment.}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{extrapolate}{TRUE if regression lines should be extrapolated beyond the range of the data. These will be plotted as dashed lines.
Defaults to FALSE.}

\item{ghosted}{TRUE if studies should be plotted in grey. Defaults to FALSE.}
}
\value{
The modified ggplot2 object.
}
\description{
Plot the regression lines on the plot.
}
