% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{.PrepareAnalysisForCinema}
\alias{.PrepareAnalysisForCinema}
\title{Prepare analysis into a format that CINeMA can read.
This is a named list of lists. The top level list contains items:
\itemize{
\item "hatmatrix" Hat matrix from {netmeta}
\itemize{
\item "colNames" Vector of column names in form "<TREATMENT_N>:<TREATMENT_M>"
\item "H" Matrix of matrices
\item "model" Type of model: "fixed" or "random"
\item "rowNames" Vector of row names in form "<TREATMENT_N>:<TREATMENT_M>"
\item "sm" Outcome measure, one of: "OR", "RR", "RD", "MD", "SMD"
}
\item "studycontributions" Output from \code{.PrepareStudyContibutionsForCinema()}
}}
\usage{
.PrepareAnalysisForCinema(
  contributions,
  model_type,
  outcome_measure,
  gemtc_results = NULL
)
}
\arguments{
\item{contributions}{Contributions from {netmeta}.}

\item{model_type}{Type of model: "fixed" or "random".}

\item{outcome_measure}{Outcome measure, one of: "OR", "RR", "RD", "MD", "SMD".}

\item{gemtc_results}{Output from gemtc::mtc.run, as returned in the 'mtcResults' list element from baye(). If this parameter is NULL then the frequentist analysis results found in 'contributions' are used. If it is not NULL then the Bayesian analysis results contained in this parameter are used. Defaults to NULL.}
}
\value{
A List with the described structure.
}
\description{
Prepare analysis into a format that CINeMA can read.
This is a named list of lists. The top level list contains items:
\itemize{
\item "hatmatrix" Hat matrix from {netmeta}
\itemize{
\item "colNames" Vector of column names in form "<TREATMENT_N>:<TREATMENT_M>"
\item "H" Matrix of matrices
\item "model" Type of model: "fixed" or "random"
\item "rowNames" Vector of row names in form "<TREATMENT_N>:<TREATMENT_M>"
\item "sm" Outcome measure, one of: "OR", "RR", "RD", "MD", "SMD"
}
\item "studycontributions" Output from \code{.PrepareStudyContibutionsForCinema()}
}
}
