% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{.PrepareComparisonForCinema}
\alias{.PrepareComparisonForCinema}
\title{Creates a list containing the results of a comparison. The elements are analysis  results including treatment effects, standard errors, confidence intervals, predictions intervals, and p-values, for NMA effects, direct and indirect effects, and SIDE inconsistency factors. If there is no such data (e.g. if there is no indirect evidence) then the list element does not appear.}
\usage{
.PrepareComparisonForCinema(
  freq_results,
  model_type,
  comparison,
  gemtc_results = NULL
)
}
\arguments{
\item{freq_results}{Results from a frequentist model as contained in contribution output from {netmeta}.}

\item{model_type}{Type of model: "fixed" or "random".}

\item{comparison}{A comparison of the form \if{html}{\out{<treatment1>}}:\if{html}{\out{<treatment2>}}.}

\item{gemtc_results}{Output from gemtc::mtc.run, as returned in the 'mtcResults' list element from baye(). If this parameter is NULL then the frequentist analysis results found in 'contributions' are used. If it is not NULL then the Bayesian analysis results contained in this parameter are used. Defaults to NULL.}
}
\value{
List of analysis results with the described structure.
}
\description{
Creates a list containing the results of a comparison. The elements are analysis  results including treatment effects, standard errors, confidence intervals, predictions intervals, and p-values, for NMA effects, direct and indirect effects, and SIDE inconsistency factors. If there is no such data (e.g. if there is no indirect evidence) then the list element does not appear.
}
