% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{.PrepareComparisonsForCinema}
\alias{.PrepareComparisonsForCinema}
\title{Creates a list, each element of which is a named list containing the results of a comparison. In the outer list there is one element per comparison. Each inner list contains results including treatment effects, standard errors, confidence intervals, predictions intervals, and p-values, for NMA effects, direct and indirect effects, and SIDE inconsistency factors. If there is no such data (e.g. if there is no indirect evidence) then the list element does not appear.}
\usage{
.PrepareComparisonsForCinema(
  contributions,
  model_type,
  outcome_measure,
  gemtc_results = NULL
)
}
\arguments{
\item{contributions}{Contributions from {netmeta}.}

\item{model_type}{Type of model: "fixed" or "random".}

\item{outcome_measure}{Outcome measure, one of: "OR", "RR", "RD", "MD", "SMD".}

\item{gemtc_results}{Output from gemtc::mtc.run, as returned in the 'mtcResults' list element from baye(). If this parameter is NULL then the frequentist analysis results found in 'contributions' are used. If it is not NULL then the Bayesian analysis results contained in this parameter are used. Defaults to NULL.}
}
\value{
List of results with the described structure.
}
\description{
Creates a list, each element of which is a named list containing the results of a comparison. In the outer list there is one element per comparison. Each inner list contains results including treatment effects, standard errors, confidence intervals, predictions intervals, and p-values, for NMA effects, direct and indirect effects, and SIDE inconsistency factors. If there is no such data (e.g. if there is no indirect evidence) then the list element does not appear.
}
