% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{.PrepareDataForCinema}
\alias{.PrepareDataForCinema}
\title{Prepare data into a format that CINeMA can read.
This is an unnamed list of lists. The top level list contains an item for each study arm,
as in long format.
Each item in the second level is a named list of study arm properties:
\itemize{
\item "study" Study name
\item "id" Unique study ID
\item "t" Name of treatment
\item "n" Number of participants
\item "r" Number of events (binary outcome)
\item "mean" Mean treatment effect (continuous outcome)
\item "sd" Standard deviation of treatment effect (continuous outcome)
\item "indirectness" Indirectness of study
\item "rob" Risk of bias of study
}}
\usage{
.PrepareDataForCinema(data, treatment_ids, outcome_type)
}
\arguments{
\item{data}{The dataframe used in MetaInsight.}

\item{treatment_ids}{Data frame containing treatment names (Label), original tratment names (RawLabel) and IDs (Number).}

\item{outcome_type}{Type of outcome for which to reorder, either 'continuous' or 'binary'.}
}
\value{
A List with the described structure.
}
\description{
Prepare data into a format that CINeMA can read.
This is an unnamed list of lists. The top level list contains an item for each study arm,
as in long format.
Each item in the second level is a named list of study arm properties:
\itemize{
\item "study" Study name
\item "id" Unique study ID
\item "t" Name of treatment
\item "n" Number of participants
\item "r" Number of events (binary outcome)
\item "mean" Mean treatment effect (continuous outcome)
\item "sd" Standard deviation of treatment effect (continuous outcome)
\item "indirectness" Indirectness of study
\item "rob" Risk of bias of study
}
}
