% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{.PrepareStudyContibutionsForCinema}
\alias{.PrepareStudyContibutionsForCinema}
\title{Prepare study contributions into a format that CINeMA can read.
This is a named list of lists. The top level list has item named "<TREATMENT_N>:<TREATMENT_M>"
where "<TREATMENT_N>" and "<TREATMENT_M>" are the names of 2 treatments. There is an item
for every possible comparison in the network.
Each item in the second level is a named list of numerical values for the contribution of each
study towards the comparison. The name of each item is the name of the relevant study, and
every study is included in the list. The values sum to 1.}
\usage{
.PrepareStudyContibutionsForCinema(study_contributions, data)
}
\arguments{
\item{study_contributions}{The contribution matrix calculated by {netmeta}.}
}
\value{
A List with the described structure.
}
\description{
Prepare study contributions into a format that CINeMA can read.
This is a named list of lists. The top level list has item named "<TREATMENT_N>:<TREATMENT_M>"
where "<TREATMENT_N>" and "<TREATMENT_M>" are the names of 2 treatments. There is an item
for every possible comparison in the network.
Each item in the second level is a named list of numerical values for the contribution of each
study towards the comparison. The name of each item is the name of the relevant study, and
every study is included in the list. The values sum to 1.
}
