% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_plot_f.R
\name{.SetupIndirectCovariatePlot}
\alias{.SetupIndirectCovariatePlot}
\title{Setup the main components of the plot panel.}
\usage{
.SetupIndirectCovariatePlot(reference, comparators, include_ghosts)
}
\arguments{
\item{reference}{Name of the reference treatment.}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{include_ghosts}{TRUE if all otherc omparator studies should be plotted in grey in the background of the plot. Defaults to FALSE.}
}
\value{
Created ggplot2 object.
}
\description{
Setup the main components of the plot panel.
}
