% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_plot_f.R
\name{.SetupMainRegressionPlot}
\alias{.SetupMainRegressionPlot}
\title{Setup the main components of the plot panel.}
\usage{
.SetupMainRegressionPlot(
  reference,
  comparators,
  outcome_measure,
  include_ghosts,
  credible_opacity,
  legend_position
)
}
\arguments{
\item{reference}{Name of the reference treatment.}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{outcome_measure}{Outcome measure of analysis (OR, RR, RD, MD)}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot. Defaults to FALSE.}

\item{credible_opacity}{The opacity of the credible regions. Can be any value between 0 and 1, inclusive. Defaults to 0.2.}

\item{legend_position}{String informing the position of the legend. Acceptable values are:
\itemize{
\item "BR" - Bottom-right of the plot area
\item "BL" - Bottom-left of the plot area
\item "TR" - Top-right of the plot area
\item "TL" - Top-left of the plot area
}}
}
\value{
Created ggplot2 object.
}
\description{
Setup the main components of the plot panel.
}
