% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ThrowErrorForMatchingStudies}
\alias{.ThrowErrorForMatchingStudies}
\title{Throw an error if any study meets given criteria.}
\usage{
.ThrowErrorForMatchingStudies(values, condition, message)
}
\arguments{
\item{values}{covariate values in which to check.}

\item{condition}{Function taking the covariate values to check.
This should return TRUE in the error case.}

\item{message}{Message to show before listing the problem studies or FALSE
if no problems are found}
}
\description{
Throw an error if any study meets given criteria.
}
