% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{.TidyStringItem}
\alias{.TidyStringItem}
\title{Tidy a character column.}
\usage{
.TidyStringItem(string)
}
\arguments{
\item{string}{A character vector to be tidied.}
}
\value{
Tidied character vector.
}
\description{
Tidy a character column.
}
