% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{.ValidateColumnTypes}
\alias{.ValidateColumnTypes}
\title{Validate that all columns in the data are of the expected type.}
\usage{
.ValidateColumnTypes(data, outcome_columns)
}
\arguments{
\item{data}{Data frame to validate.}

\item{outcome_columns}{Data frame containing data definitions.}
}
\value{
Validation result in the form of a list:
- "valid" = TRUE or FALSE defining whether data is valid
- "message" = String describing any issues causing the data to be invalid
}
\description{
Validate that all columns in the data are of the expected type.
}
