% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ValidateCovariate}
\alias{.ValidateCovariate}
\title{Validate the covariate and infer the type of the covariate from the data in the column.
In error cases, this function will throw exceptions:
\itemize{
\item If the data has any NAs
\item If the data has any non-numeric values
\item If every study has the same covariate value
\item If any study contains multiple different covariate values
}}
\usage{
.ValidateCovariate(data, covariate_title)
}
\arguments{
\item{data}{Data frame containing all study data.}

\item{covariate_title}{Name of the covariate column.}
}
\value{
list containing \code{valid} and \code{message} to pass to logger
}
\description{
Validate the covariate and infer the type of the covariate from the data in the column.
In error cases, this function will throw exceptions:
\itemize{
\item If the data has any NAs
\item If the data has any non-numeric values
\item If every study has the same covariate value
\item If any study contains multiple different covariate values
}
}
