% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ValidateMatchingWideColumns}
\alias{.ValidateMatchingWideColumns}
\title{Validate wide columns that all columns are numbered sequentially from 1 to the same number across all numbered columns.}
\usage{
.ValidateMatchingWideColumns(uploaded_data, required_columns)
}
\arguments{
\item{uploaded_data}{Data frame to validate.}

\item{required_columns}{Data frame representing required numbered columns. Columns in this data frame are:
\itemize{
\item pattern = A regular expression defining the generic column title for wide or long format.
\item number_group = Regular expression replacement to extract the number from a column.
}}
}
\value{
TRUE if the columns are correctly numbered, else FALSE.
}
\description{
Validate wide columns that all columns are numbered sequentially from 1 to the same number across all numbered columns.
}
