% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ValidateMissingColumns}
\alias{.ValidateMissingColumns}
\title{Validate that there are no missing columns in the data.}
\usage{
.ValidateMissingColumns(data, required_columns, outcome)
}
\arguments{
\item{data}{Data frame to validate.}

\item{required_columns}{Data frame containing data definitions.}

\item{outcome}{Outcome type selected for the data. Either "Binary" or "Continuous".}
}
\value{
Validation result in the form of a list:
\itemize{
\item "valid" = TRUE or FALSE defining whether data is valid
\item "message" = String describing any issues causing the data to be invalid
}
}
\description{
Validate that there are no missing columns in the data.
}
