% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ValidateNumberedColumns}
\alias{.ValidateNumberedColumns}
\title{Validate that numbered columns in the data are seqential from 1.}
\usage{
.ValidateNumberedColumns(data, required_columns)
}
\arguments{
\item{data}{Data frame to validate.}

\item{required_columns}{Data frame containing data definitions.}
}
\value{
Validation result in the form of a list:
\itemize{
\item "valid" = TRUE or FALSE defining whether data is valid
\item "message" = String describing any issues causing the data to be invalid
}
}
\description{
Validate that numbered columns in the data are seqential from 1.
}
