% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ValidateQualityColumns}
\alias{.ValidateQualityColumns}
\title{Validate that the quality assessment columns in the data are of the expected format.}
\usage{
.ValidateQualityColumns(data)
}
\arguments{
\item{data}{Data frame to validate.}
}
\value{
Validation result in the form of a list:
\itemize{
\item "valid" = TRUE or FALSE defining whether data is valid
\item "message" = String describing any issues causing the data to be invalid
}
}
\description{
Validate that the quality assessment columns in the data are of the expected format.
}
