% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{.ValidateSingleArmStudies}
\alias{.ValidateSingleArmStudies}
\title{Validate that all studies have at least 2 arms.}
\usage{
.ValidateSingleArmStudies(data)
}
\arguments{
\item{data}{Data frame to validate.}
}
\value{
Validation result in the form of a list:
- "valid" = TRUE or FALSE defining whether data is valid
- "message" = String describing any issues causing the data to be invalid
}
\description{
Validate that all studies have at least 2 arms.
}
