% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.WeightMatrixFixedExchangeable}
\alias{.WeightMatrixFixedExchangeable}
\title{Intermediate function called within CreateContributionMatrix(), dealing with the fixed effects, exchangeable case.}
\usage{
.WeightMatrixFixedExchangeable(
  X,
  V,
  Z,
  basic_or_all_parameters,
  studies,
  treatments,
  std_dev_beta
)
}
\arguments{
\item{X}{X matrix.}

\item{V}{V matrix.}

\item{Z}{Z matrix.}

\item{basic_or_all_parameters}{"basic" for one column per basic parameter, "all" for one column per parameter.}

\item{studies}{Vector of studies.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{std_dev_beta}{Standard deviation of covariate parameters.}
}
\value{
The weight matrix.
}
\description{
Intermediate function called within CreateContributionMatrix(), dealing with the fixed effects, exchangeable case.
}
