% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{.WeightMatrixFixedUnrelatedShared}
\alias{.WeightMatrixFixedUnrelatedShared}
\title{Intermediate function called within CreateContributionMatrix(), dealing with the fixed effects, unrelated or shared cases.}
\usage{
.WeightMatrixFixedUnrelatedShared(X, V, Z, basic_or_all_parameters)
}
\arguments{
\item{X}{X matrix.}

\item{V}{V matrix.}

\item{Z}{Z matrix.}

\item{basic_or_all_parameters}{"basic" for one column per basic parameter, "all" for one column per parameter.}
}
\value{
The weight matrix.
}
\description{
Intermediate function called within CreateContributionMatrix(), dealing with the fixed effects, unrelated or shared cases.
}
