% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{download_button_pair}
\alias{download_button_pair}
\title{Create a pair of download buttons}
\usage{
download_button_pair(id)
}
\arguments{
\item{id}{The id of the module}
}
\value{
tagList of downloadButtons
}
\description{
Create a pair of download buttons
}
