% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{entry.df}
\alias{entry.df}
\title{Converts long data to wide, leaves wide data unchanged.}
\usage{
entry.df(data, CONBI)
}
\arguments{
\item{data}{Input dataset.}

\item{CONBI}{"Continuous" or "Binary".}
}
\value{
Input data in long format.
}
\description{
Converts long data to wide, leaves wide data unchanged.
}
