% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cinema_f.R
\name{export_cinema}
\alias{export_cinema}
\title{export_cinema}
\usage{
export_cinema(configured_data, gemtc_results = NULL, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{gemtc_results}{Output from \code{gemtc::mtc.run()}, as returned in the \code{mtcResults} list element from \code{bayes_model()}.
If this parameter is \code{NULL} then the frequentist analysis results found in 'contributions' are used.
If it is not \code{NULL} then the Bayesian analysis results contained in this parameter are used. Defaults to \code{NULL}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
JSON string with the following structure:
A named list of lists. The top level list contains items:
\itemize{
\item "project" Information for CINeMA project
\itemize{
\item "CM" Contribution matrices
\itemize{
\item "contributionMatrices" output from \code{.PrepareAnalysisForCinema()}
}
\item "format" Data format. Always "long"
\item "type" Outcome type. Either "binary" or "continuous"
\item "Studies" Study data
\itemize{
\item "long" Output from \code{.PrepareDataForCinema()}
}
}
}
}
\description{
Prepare project into a JSON format that CINeMA can read.
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

cinema_project <- export_cinema(configured_data = configured_data)

writeLines(cinema_project, tempfile(fileext = ".json"))

}
