% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{extract_ci}
\alias{extract_ci}
\title{Extract the minimum and maximum confidence intervals from the summary produced by netmeta}
\usage{
extract_ci(freq, outcome)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{outcome}{character. `Binary` or `Continuous`}
}
\value{
List containing:
 \item{xmin}{numeric. Minimum confidence interval}
 \item{xmax}{numeric. Maximum confidence interval}
}
\description{
Extract the minimum and maximum confidence intervals from the summary produced by netmeta
}
