% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{ExtractGemtcStats}
\alias{ExtractGemtcStats}
\title{Extract the analysis statistics required for CINeMA from GEMTC output. This includes simulating a predictive distribution in order to get a prediction interval, which is not done in GEMTC.}
\usage{
ExtractGemtcStats(gemtc_results, treatments)
}
\arguments{
\item{gemtc_results}{Model results from gemtc::mtc.run().}

\item{treatments}{Vector of two treatments}
}
\value{
Named vector with names 'median', 'se', 'ci_lower', 'ci_upper', 'pi_lower', and 'pi_upper'.
}
\description{
Extract the analysis statistics required for CINeMA from GEMTC output. This includes simulating a predictive distribution in order to get a prediction interval, which is not done in GEMTC.
}
