% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{FindAllTreatments}
\alias{FindAllTreatments}
\title{Find all of the treatment names in the data, both for long and wide formats.
If a study is specified, then the treatments are only found for that study.
Otherwise, treatments are found for all studies.}
\usage{
FindAllTreatments(data, treatment_ids = NULL, study = NULL)
}
\arguments{
\item{data}{Data frame in which to search for treatment names.}

\item{treatment_ids}{Data frame containing names of treatments ("Label") and IDs ("Number").
This is only needed if the "T" column in the data frame contains IDs instead of names. Defaults to NULL.}

\item{study}{Name of study for which to find treatment names. Defaults to NULL.}
}
\value{
Vector of all treatment names.
}
\description{
Find all of the treatment names in the data, both for long and wide formats.
If a study is specified, then the treatments are only found for that study.
Otherwise, treatments are found for all studies.
}
