% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{FindDataShape}
\alias{FindDataShape}
\title{Find which shape the data takes: either wide or long.}
\usage{
FindDataShape(data)
}
\arguments{
\item{data}{Data for which to check shape}
}
\value{
Either "wide" or "long"
}
\description{
Find which shape the data takes: either wide or long.
}
