% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{FindExpectedReferenceTreatment}
\alias{FindExpectedReferenceTreatment}
\title{Find the expected reference treatment from a vector.
This is done by comparing treatment names to expected reference treatment names.}
\usage{
FindExpectedReferenceTreatment(treatments)
}
\arguments{
\item{treatments}{vector containing all treatment names}
}
\value{
Name of the expected reference treatment if one is found, else NULL
}
\description{
Find the expected reference treatment from a vector.
This is done by comparing treatment names to expected reference treatment names.
}
