% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{FindMaxArms}
\alias{FindMaxArms}
\title{Find the maximum number of arms in a study, for long or wide data.}
\usage{
FindMaxArms(data)
}
\arguments{
\item{data}{Long or wide format data.}
}
\value{
The maximum number of arms.
}
\description{
Find the maximum number of arms in a study, for long or wide data.
}
