% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{FindRobIndividualNames}
\alias{FindRobIndividualNames}
\title{Find the names of all columns which contain an individual RoB variable.}
\usage{
FindRobIndividualNames(df)
}
\arguments{
\item{df}{Data frame in which to find covariate columns.}
}
\value{
Names of all individual RoB columns
}
\description{
Find the names of all columns which contain an individual RoB variable.
}
