% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{FindRobNames}
\alias{FindRobNames}
\title{Find the names of all columns which contain a RoB or indirectness variable.}
\usage{
FindRobNames(df)
}
\arguments{
\item{df}{Data frame in which to find covariate columns.}
}
\value{
Names of all RoB or indirectness columns
}
\description{
Find the names of all columns which contain a RoB or indirectness variable.
}
