% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{FindStartingXLimits}
\alias{FindStartingXLimits}
\title{Finds suitable starting limits for the x-axis.}
\usage{
FindStartingXLimits(pairwise)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}
}
\value{
Named vector in the form c(lower, upper).
}
\description{
Finds suitable starting limits for the x-axis.
}
