% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{FindStudiesIncludingTreatments}
\alias{FindStudiesIncludingTreatments}
\title{Find all of the study names which include the given treatment names, both for long and wide formats.}
\usage{
FindStudiesIncludingTreatments(data, treatments)
}
\arguments{
\item{data}{Data frame in which to search for study names}

\item{treatments}{Vector of matching treatments}
}
\value{
Vector of all matching study names
}
\description{
Find all of the study names which include the given treatment names, both for long and wide formats.
}
