% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{forest.df}
\alias{forest.df}
\title{Creates the forest plot from netmeta::netmeta() output.}
\usage{
forest.df(netresult, model, lstx, ref, min, max)
}
\arguments{
\item{netresult}{NMA results from netmeta::netmeta(), as produced by freq_wrap().}

\item{model}{"fixed" or "random".}

\item{lstx}{Vector of treatment labels.}

\item{ref}{Reference treatment.}

\item{min}{Minimum x-axis limit.}

\item{max}{Maximum x-axis limit.}

\item{ntx}{Number of treatments.}
}
\value{
Forest plot created using metafor::forest().
}
\description{
Creates the forest plot from netmeta::netmeta() output.
}
