% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{forest_height_pixels}
\alias{forest_height_pixels}
\title{Calculate the pixel height of a forest plot for a given number of treatments}
\usage{
forest_height_pixels(notrt, title = FALSE, annotation = FALSE)
}
\arguments{
\item{notrt}{The number of treatments in the plot}

\item{title}{TRUE if the title is included in the plot}

\item{annotation}{TRUE if an annotation is included in the plot}
}
\value{
The height of the plot in pixels
}
\description{
Calculate the pixel height of a forest plot for a given number of treatments
}
