% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{forest_width}
\alias{forest_width}
\title{Calculate the width in inches of a forest plot for a given label.}
\usage{
forest_width(n_chars)
}
\arguments{
\item{n_chars}{The number of characters in the widest label.
For frequentist plots this should be the longest treatment name.
For Bayesian plots this should be 'Compared to {reference treatment}'}
}
\value{
The width of the plot in inches
}
\description{
Calculate the width in inches of a forest plot for a given label.
}
