% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_forest_f.R
\name{format_baseline_forest}
\alias{format_baseline_forest}
\title{Convert output of \code{bnma::relative.effects.table()} into a suitable format
to pass to \code{gemtc::blobbogram()}}
\usage{
format_baseline_forest(median_ci_table, reference_treatment)
}
\arguments{
\item{median_ci_table}{matrix. Created by \code{bnma::relative.effects.table()}}

\item{reference_treatment}{character. The reference treatment of the dataset}
}
\description{
Convert output of \code{bnma::relative.effects.table()} into a suitable format
to pass to \code{gemtc::blobbogram()}
}
