% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_step}
\alias{format_step}
\title{Calculate a sensible step value to use in numericInput}
\usage{
format_step(x)
}
\arguments{
\item{x}{numeric. The value to create a step for}
}
\description{
Calculate a sensible step value to use in numericInput
}
\keyword{internal}
