% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_xlim}
\alias{format_xlim}
\title{Format xlimits to pretty values. Adapted from \code{gemtc::blobbogram()}}
\usage{
format_xlim(x, limit, log.scale)
}
\arguments{
\item{x}{numeric. The value to format.}

\item{limit}{character. Either \code{min} or \code{max}}

\item{log.scale}{logical. Whether the values are on a log scale.}
}
\description{
Format xlimits to pretty values. Adapted from \code{gemtc::blobbogram()}
}
