% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{FormatForBnma}
\alias{FormatForBnma}
\title{Takes MetaInsight data and converts it into \CRANpkg{bnma} format}
\usage{
FormatForBnma(connected_data, treatment_df, outcome, reference_treatment)
}
\arguments{
\item{connected_data}{A data frame of data in MetaInsight format}

\item{treatment_df}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively}

\item{outcome}{"continuous" or "binary"}

\item{reference_treatment}{An element of treatment_df$Label, the reference treatment.}
}
\value{
List:
\itemize{
\item 'ArmLevel' = Data frame containing 'Study', 'Treat', 'N', 'Outcomes', and (for outcome_type="continuous") 'SD'
\item 'Treat.order' = Vector of (unique) treatments, with the reference treatment first
}
}
\description{
Takes MetaInsight data and converts it into \CRANpkg{bnma} format
}
