% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_forest_f.R
\name{FormatForCreateTauSentence}
\alias{FormatForCreateTauSentence}
\title{Put the output from a bnma model into the format of a gemtc model, in order to apply CreateTauSentence().}
\usage{
FormatForCreateTauSentence(model)
}
\arguments{
\item{br_model}{Output from bnma::network.run.}
}
\value{
List:
\itemize{
\item 'sumresults' = List:
\itemize{
\item 'summaries' = equivalent of summary(gemtc model)$summaries.
\item 'a' = "fixed effect" or "random effect".
}
}
}
\description{
Put the output from a bnma model into the format of a gemtc model, in order to apply CreateTauSentence().
}
