% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{freq.df}
\alias{freq.df}
\title{Frequentist NMA.}
\usage{
freq.df(model, outcome, dataf, lstx, ref)
}
\arguments{
\item{model}{"fixed" or "random".}

\item{outcome}{"MD", "SMD", "OR", "RR", or "RD".}

\item{dataf}{Data in contrast form with treatment labels, typically output from labelmatching.df().}

\item{lstx}{Vector of treatment labels. (TM: this parameter is not used. Removing it has knock on effects due to parameters not being named when this function is called. To be removed in a later task).}

\item{ref}{Reference treatment.}
}
\value{
NMA results from netmeta::netmeta().
}
\description{
Frequentist NMA.
}
