% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_compare_f.R
\name{freq_compare}
\alias{freq_compare}
\title{Produce comparison tables}
\usage{
freq_compare(freq, model_type, ranking_option, logger = NULL)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{ranking_option}{character. \code{good} if the treatment effect is desirable, else \code{bad}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
Ranking table created by \code{netmeta::netleague()}
}
\description{
Produce comparison tables
}
