% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{freq_forest}
\alias{freq_forest}
\title{Produce a frequentist forest plot}
\usage{
freq_forest(
  configured_data,
  xmin = NULL,
  xmax = NULL,
  title = "",
  logger = NULL
)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{xmin}{numeric. Minimum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{xmax}{numeric. Maximum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{title}{character. Title for the plot.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Produce an annotated frequentist forest plot using
\code{meta::forest()}
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

freq_forest(configured_data = configured_data)

}
