% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{freq_forest_annotation}
\alias{freq_forest_annotation}
\title{Creates the text to be displayed underneath the forest plots, with between-study SD, number of studies and number of treatments.}
\usage{
freq_forest_annotation(freq, effects, outcome_measure)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{effects}{"fixed" or "random".}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}
}
\value{
Text as described above.
}
\description{
Creates the text to be displayed underneath the forest plots, with between-study SD, number of studies and number of treatments.
}
