% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{freq_forest_limits}
\alias{freq_forest_limits}
\title{Extract the minimum and maximum confidence intervals from the summary produced by netmeta}
\usage{
freq_forest_limits(freq, outcome)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{outcome}{character. \code{binary} or \code{continuous}}
}
\value{
List containing:
\item{xmin}{numeric. Minimum confidence interval}
\item{xmax}{numeric. Maximum confidence interval}
}
\description{
Extract the minimum and maximum confidence intervals from the summary produced by netmeta
}
