% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{freq_wrap}
\alias{freq_wrap}
\title{Calls preliminary functions, finishing with frequentist analysis in freq.df(), then outputs the model results and related objects.}
\usage{
freq_wrap(data, treat_list, model, outcome, CONBI, ref)
}
\arguments{
\item{data}{Input dataset in wide format.}

\item{treat_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}

\item{model}{"fixed" or "random".}

\item{outcome}{"MD", "SMD", "OR", "RR", or "RD".}

\item{CONBI}{"Continuous" or "Binary".}

\item{ref}{Reference treatment.}
}
\value{
List
\itemize{
\item 'net1': NMA results from netmeta::netmeta().
\item 'lstx': Vector of treatment labels.
\item 'ntx': Number of treatments.
\item 'd0': Data in contrast form.
\item 'd1': Same as 'd0' but with treatment labels.
}
}
\description{
Calls preliminary functions, finishing with frequentist analysis in freq.df(), then outputs the model results and related objects.
}
