% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_analysis.R
\name{frequentist}
\alias{frequentist}
\title{Fit a frequentist model}
\usage{
frequentist(
  non_covariate_data,
  outcome,
  treatments,
  outcome_measure,
  effects,
  reference_treatment
)
}
\arguments{
\item{non_covariate_data}{Input dataset with any covariates removed.}

\item{outcome}{character. Outcome type for the dataset. Either \code{binary} or
\code{continuous}.}

\item{treatments}{dataframe. Treatments}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{continuous}}

\item{effects}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{reference_treatment}{character. The reference treatment of the dataset}
}
\value{
List containing:
\item{netmeta}{list. NMA results from netmeta::netmeta()}
\item{pairwise}{dataframe. Results from meta::pairwise() but with treatment labels}
\item{pairwise_reversed}{dataframe. pairwise as if the treatments had been the other way round}
}
\description{
Fits a frequentist model with \code{netmeta::netmeta()}
}
\keyword{internal}
