% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_analysis_panel.R
\name{frequentist_analysis_panel_server}
\alias{frequentist_analysis_panel_server}
\title{Module server for the frequentist analysis panel.}
\usage{
frequentist_analysis_panel_server(
  id,
  metaoutcome,
  outcome_measure,
  model_effects,
  rank_option,
  freq_all,
  freq_sub,
  bugsnetdt,
  bugsnetdt_sub,
  treatment_df,
  reference_alter
)
}
\arguments{
\item{id}{ID of the module}

\item{metaoutcome}{Reactive containing meta analysis outcome: "Continuous" or "Binary"}

\item{outcome_measure}{Reactive containing meta analysis outcome measure: "MD", "SMD", "OR, "RR", or "RD"}

\item{model_effects}{Reactive containing model effects: either "random" or "fixed"}

\item{rank_option}{Reactive containing ranking option: "good" or "bad" depending on whether small values are desirable or not}

\item{freq_all}{Reactive containing frequentist meta-analysis}

\item{freq_sub}{Reactive containing frequentist meta-analysis for the sensitivity analysis}

\item{bugsnetdt}{Reactive containing bugsnet meta-analysis}

\item{bugsnetdt_sub}{Reactive containing bugsnet meta-analysis for sensitivity analysis}

\item{treatment_df}{Reactive data frame containing treatment IDs (Number) and names (Label)}

\item{reference_alter}{Reactive containing the name of the reference treatment for the sensitivity
analysis accounting for if the chosen reference treatment has been excluded}
}
\description{
Module server for the frequentist analysis panel.
}
