% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_analysis_panel.R
\name{frequentist_analysis_panel_ui}
\alias{frequentist_analysis_panel_ui}
\title{Module UI for the frquentist analysis panel.}
\usage{
frequentist_analysis_panel_ui(id, page_numbering)
}
\arguments{
\item{id}{ID of the module}
}
\value{
Div for the panel
}
\description{
Module UI for the frquentist analysis panel.
}
