% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{GelmanPlot}
\alias{GelmanPlot}
\title{Creates a Gelman plot for a gemtc or bnma model.}
\usage{
GelmanPlot(gelman_data, parameter)
}
\arguments{
\item{gelman_data}{Output from \code{coda::gelman.plot(model$samples[, parm])}, where \code{parm} is a parameter from the model.}

\item{parameter}{The parameter from the previous argument, used as the title.}
}
\value{
A function that reproduces the Gelman plot mentioned in @param gelman_plot as a plot that can be put in a grid.
}
\description{
Creates a Gelman plot for a gemtc or bnma model.
}
