% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{GelmanPlots}
\alias{GelmanPlots}
\title{Creates Gelman plots for a gemtc or bnma model.}
\usage{
GelmanPlots(gelman_data, parameters)
}
\arguments{
\item{parameters}{Vector of parameters mentioned in the previous argument.}

\item{gelman_plots}{List of outputs from \code{coda::gelman.plot(model$samples[, parm])}, where parm is a parameter from the model.}
}
\value{
A list of functions that produce Gelman plots mentioned in @param gelman_plots.
}
\description{
Creates Gelman plots for a gemtc or bnma model.
}
