% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_details_f.R
\name{GetBnmaMcmcCharacteristics}
\alias{GetBnmaMcmcCharacteristics}
\title{MCMC characteristics from a BNMA model.}
\usage{
GetBnmaMcmcCharacteristics(model)
}
\arguments{
\item{model}{BNMA model output.}
}
\value{
Data frame with four MCMC characteristics.
}
\description{
MCMC characteristics from a BNMA model.
}
