% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{GetBnmaParameters}
\alias{GetBnmaParameters}
\title{Get the parameters that are to be displayed in Gelman plots.}
\usage{
GetBnmaParameters(all_parameters, effects_type, cov_parameters)
}
\arguments{
\item{all_parameters}{Vector of monitored parameters from a bnma model.}

\item{effects_type}{"fixed" or "random".}

\item{cov_parameters}{"shared", "exchangeable", or "unrelated".}
}
\value{
Vector of treatment effect and covariate parameter names, plus random effects sd and/or exchangeable covariate sd.
}
\description{
Get the parameters that are to be displayed in Gelman plots.
}
