% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_details_f.R
\name{GetBnmaPriors}
\alias{GetBnmaPriors}
\title{Prior distributions from a BNMA model.}
\usage{
GetBnmaPriors(model)
}
\arguments{
\item{model}{BNMA model output.}
}
\value{
Data frame with prior distribution information.
}
\description{
Prior distributions from a BNMA model.
}
