% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_details_f.R
\name{GetGemtcMcmcCharacteristics}
\alias{GetGemtcMcmcCharacteristics}
\title{MCMC characteristics from a GEMTC model.}
\usage{
GetGemtcMcmcCharacteristics(model)
}
\arguments{
\item{model}{GEMTC model output.}
}
\value{
Data frame with four MCMC characteristics.
}
\description{
MCMC characteristics from a GEMTC model.
}
