% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_details_f.R
\name{GetGemtcPriors}
\alias{GetGemtcPriors}
\title{Prior distributions from a GEMTC model.}
\usage{
GetGemtcPriors(model)
}
\arguments{
\item{model}{GEMTC model output.}
}
\value{
Data frame with prior distribution information.
}
\description{
Prior distributions from a GEMTC model.
}
