% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{GetOutcomesAndVariances}
\alias{GetOutcomesAndVariances}
\title{Get observed outcomes and variances.}
\usage{
GetOutcomesAndVariances(data, treatments, outcome_type, outcome_measure)
}
\arguments{
\item{data}{Input data in long format plus the column 'Treatment', a text version of 'T'.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{outcome_type}{"Continuous" or "Binary".}

\item{outcome_measure}{"MD" (when outcome_type == "Continuous"), "OR", "RR" or "RD" (when outcome_type == "Binary").}
}
\value{
Data frame with columns 'Study', 'Treatment', 'Outcome', and 'Variance'.
}
\description{
Get observed outcomes and variances.
}
