% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{GetReferenceOutcome}
\alias{GetReferenceOutcome}
\title{Get the outcome in the reference arm.}
\usage{
GetReferenceOutcome(
  connected_data,
  treatment_df,
  outcome,
  observed,
  model = NULL
)
}
\arguments{
\item{connected_data}{Data in long format.}

\item{treatment_df}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively.}

\item{outcome}{"binary" or "continuous".}

\item{observed}{"Observed" or "Imputed". See return for details.}

\item{model}{Model created by \code{bnma::network.run()}. Only required when \code{observed} = \code{Imputed}. Defaults to NULL.}
}
\value{
Vector of reference arm outcomes, named by study.
If a study contains the reference treatment then the value returned is the observed outcome in the reference arm.
If a study does not contain the reference treatment then the value returned is:
- NA if \code{observed} = \code{Observed};
- The median of the study-specific intercept parameter if \code{observed} = \code{Imputed}.
}
\description{
Get the outcome in the reference arm.
}
\keyword{internal}
