% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{groupforest.df}
\alias{groupforest.df}
\title{Produces the forest plot of raw outcomes, with one section per treatment contrast. For a given treatment contrast, the section contains one confidence interval for each study that directly compares the two treatments in question.}
\usage{
groupforest.df(d1, ntx, lstx, outcome, HeaderSize, TitleSize)
}
\arguments{
\item{d1}{Data in contrast form, typically output from contrastform.df() or labelmatching.df().}

\item{ntx}{Number of treatments.}

\item{lstx}{Vector of treatment labels.}

\item{outcome}{"MD", "SMD", "OR", "RR", or "RD".}

\item{HeaderSize}{Multiplier for size of text in treatment contrast headers.}

\item{TitleSize}{Multiplier for size of title.}
}
\value{
The forest plot, created using metafor::forest().
}
\description{
Produces the forest plot of raw outcomes, with one section per treatment contrast. For a given treatment contrast, the section contains one confidence interval for each study that directly compares the two treatments in question.
}
