% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{hide_and_show}
\alias{hide_and_show}
\title{hide_and_show}
\usage{
hide_and_show(module_id, show = TRUE)
}
\arguments{
\item{module_id}{character. The module identifier.}

\item{show}{logical. Whether to show the div when \code{trigger(module_id)} is
called}
}
\description{
For internal use. Hides content inside the <module_id>_div class
when a module has not been run and shows it once \code{trigger(module_id)} has
been called. The show option is \code{TRUE} by default but can be set to \code{FALSE}
to manually control when the content is displayed e.g.  if it should
be when an extendedtask completes instead of when the run button is
pressed.
}
\keyword{internal}
