% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\name{home_page_server}
\alias{home_page_server}
\title{Module server for the home page.}
\usage{
home_page_server(id)
}
\arguments{
\item{id}{ID of the module}
}
\description{
Module server for the home page.
}
