% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\name{home_page_ui}
\alias{home_page_ui}
\title{Module UI for the home page.}
\usage{
home_page_ui(id)
}
\arguments{
\item{id}{ID of the module}
}
\value{
Div for the home page
}
\description{
Module UI for the home page.
}
