% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_networks.R
\name{IdentifySubNetworks}
\alias{IdentifySubNetworks}
\title{Identify all of the disconnected subnetworks contained in the data.}
\usage{
IdentifySubNetworks(
  data,
  treatment_df,
  reference_treatment_name = NULL,
  subnet_name_prefix = "subnet_"
)
}
\arguments{
\item{data}{Data frame containing all of the studies for binary or continuous outcomes, and wide or long format.}

\item{treatment_df}{Data frame containing the names ("Label") and IDs ("Number") of the treatments.}

\item{reference_treatment_name}{The name of the reference treatment. If not specified, then defaults to treatment with "Number" 1 in treatment_df.}

\item{subnet_name_prefix}{Names of the subnetworks. Defaults to "subnet_"}
}
\value{
List of subnetworks, where each subnetwork is a list containing:
- "treatments" = The IDs of the treatments included in the given network
- "studies" = The names of the studies included in the given subnetwork
}
\description{
Identify all of the disconnected subnetworks contained in the data.
}
