% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_storage.R
\name{IncludeLocalStorage}
\alias{IncludeLocalStorage}
\title{Include browser local storage access in the app. Add this anywhere within the app UI.}
\usage{
IncludeLocalStorage()
}
\value{
Script for accessing browser local storage.
}
\description{
Include browser local storage access in the app. Add this anywhere within the app UI.
}
