% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{InferCovariateType}
\alias{InferCovariateType}
\title{Infer whether the covariate data is binary or continuous}
\usage{
InferCovariateType(covariate_data)
}
\arguments{
\item{covariate_data}{Data from which to remove covariate columns}
}
\value{
Data without covariate columns
}
\description{
Infer whether the covariate data is binary or continuous
}
\keyword{internal}
