% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informed_conditional_panel.R
\name{informed_conditional_panel_server}
\alias{informed_conditional_panel_server}
\title{Create the server functions to show and hide the panel and error message. This will hide the given UI and display an
error message when the condition is not met. Note that this will throw an error when called with both a text and UI expression.}
\usage{
informed_conditional_panel_server(
  id,
  condition,
  error_message_text_expression = NULL,
  error_message_ui_expression = NULL,
  inner_server_expression
)
}
\arguments{
\item{id}{ID of the module}

\item{condition}{Reactive containing the state of the condition (TRUE or FALSE).}

\item{error_message_text_expression}{Expression to create text to be shown to the user when the condition is not met.}

\item{error_message_ui_expression}{Expression to create UI elements to be shown to the user when the condition is not met.}

\item{inner_server_expression}{Expression for the server function for this item.}
}
\description{
Create the server functions to show and hide the panel and error message. This will hide the given UI and display an
error message when the condition is not met. Note that this will throw an error when called with both a text and UI expression.
}
