% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informed_conditional_panel.R
\name{informed_conditional_panel_ui}
\alias{informed_conditional_panel_ui}
\title{Create a panel which is replaced with an error message when a condition is not met.}
\usage{
informed_conditional_panel_ui(id, inner_ui_expression, args = list())
}
\arguments{
\item{id}{ID of the module}

\item{inner_ui_expression}{Expression for the UI function for this item.}
}
\value{
Div containing the module
}
\description{
Create a panel which is replaced with an error message when a condition is not met.
}
