% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{InvertMetaAnalysis}
\alias{InvertMetaAnalysis}
\title{Conditionally invert the meta-analysis data based on the sorting order.}
\usage{
InvertMetaAnalysis(meta_analysis, mtorg, outcome_measure, indices)
}
\arguments{
\item{meta_analysis}{Meta-analysis data for direct and indirect evidence}

\item{mtorg}{Sorting order matrix}

\item{outcome_measure}{Type of outcome being plotted}

\item{indices}{Indices of the meta-analysis data to invert}
}
\value{
Inverted meta-analysis data
}
\description{
Conditionally invert the meta-analysis data based on the sorting order.
}
