% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_networks.R
\name{IsNodesplittable}
\alias{IsNodesplittable}
\title{Determine if the network contains any splittable nodes. If not, state the reason.}
\usage{
IsNodesplittable(data, treatments)
}
\arguments{
\item{data}{The dataset from which to create the network.}

\item{treatments}{Vector of treatments in order.}
}
\value{
List:
\itemize{
\item is_nodesplittable: TRUE or FALSE.
\item reason: Text describing the reason there are no splittable nodes, or NULL when is_nodesplittable == TRUE.
}
}
\description{
Determine if the network contains any splittable nodes. If not, state the reason.
}
