% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{KeepOrDeleteControlTreatment}
\alias{KeepOrDeleteControlTreatment}
\title{Keep or delete rows in @param data corresponding to the control treatment in each study.}
\usage{
KeepOrDeleteControlTreatment(data, treatments, keep_delete)
}
\arguments{
\item{data}{Data in long format, plus the column 'Treatment', a text version of 'T'.}

\item{treatments}{Vector of treatments with the reference treatment first.}

\item{keep_delete}{"keep" or "delete".}
}
\value{
@param data with rows corresponding to the control treatment kept or deleted, and a new column 'Control'.
}
\description{
Keep or delete rows in @param data corresponding to the control treatment in each study.
}
