% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{labelmatching.df}
\alias{labelmatching.df}
\title{Adds treatment labels to contrast data.}
\usage{
labelmatching.df(d1, ntx, treat_list)
}
\arguments{
\item{d1}{Data in contrast form, typically created by contrastform.df().}

\item{ntx}{= Number of treatments.}

\item{treat_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}
}
\value{
Input data with the columns treat1 and treat2 now labelled.
}
\description{
Adds treatment labels to contrast data.
}
