% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_deviance_f.R
\name{lev_plot}
\alias{lev_plot}
\title{Creates a plot of leverage versus residual deviance, by study.}
\usage{
lev_plot(deviance)
}
\arguments{
\item{deviance}{gemtc::mtc.deviance() object.}
}
\value{
Leverage vs residual deviance plot.
}
\description{
Creates a plot of leverage versus residual deviance, by study.
}
