% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{levplot}
\alias{levplot}
\title{Leverage plot}
\usage{
levplot(model, package = "gemtc")
}
\arguments{
\item{model}{Various model output created by baye().}
}
\value{
Leverage plot created by levplot.df().
}
\description{
Leverage plot
}
