% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_define_f.R
\name{load_define}
\alias{load_define}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
load_define(
  data,
  treatment_df,
  outcome_type,
  reference_treatment,
  logger = NULL
)
}
\arguments{
\item{data}{dataframe. Uploaded data}

\item{treatment_df}{vector of treatments}

\item{reference_treatment}{character. The reference treatment}
}
\value{
list
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
