% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{loading_spinner}
\alias{loading_spinner}
\title{loading_spinner}
\usage{
loading_spinner(class)
}
\arguments{
\item{class}{character}
}
\description{
For internal use. Show a loading spinner when an output is recalculating.
Visibility is toggled using shinyjs::show(selector = ".class"). Note that you can use
multiple classes and they can be shown / hidden with independent triggers i.e. use one
to hide and a different one to show
}
\keyword{internal}
