% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_storage.R
\name{LocalStorage}
\alias{LocalStorage}
\title{Local Storage}
\description{
Class for accessing the browser local storage from within a \emph{Shiny} app.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{session}}{Root Shiny session.}

\item{\code{storage}}{Reactive containing list of stored values.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{session}}{Root Shiny session.}

\item{\code{storage}}{Reactive containing list of stored values.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Local Storage-new}{\code{LocalStorage$new()}}
\item \href{#method-Local Storage-SetStoredValue}{\code{LocalStorage$SetStoredValue()}}
\item \href{#method-Local Storage-RemoveStoredValue}{\code{LocalStorage$RemoveStoredValue()}}
\item \href{#method-Local Storage-GetStoredValue}{\code{LocalStorage$GetStoredValue()}}
\item \href{#method-Local Storage-GetStoredIds}{\code{LocalStorage$GetStoredIds()}}
\item \href{#method-Local Storage-clone}{\code{LocalStorage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Local Storage-new"></a>}}
\if{latex}{\out{\hypertarget{method-Local Storage-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Local Storage-SetStoredValue"></a>}}
\if{latex}{\out{\hypertarget{method-Local Storage-SetStoredValue}{}}}
\subsection{Method \code{SetStoredValue()}}{
Write a value to the browser local storage. This will overwrite the value if it is already stored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$SetStoredValue(id, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Name of the value to store.}

\item{\code{value}}{(\code{any})\cr
Value to store.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Local Storage-RemoveStoredValue"></a>}}
\if{latex}{\out{\hypertarget{method-Local Storage-RemoveStoredValue}{}}}
\subsection{Method \code{RemoveStoredValue()}}{
Remove a value from the browser local storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$RemoveStoredValue(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Name of the value to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Local Storage-GetStoredValue"></a>}}
\if{latex}{\out{\hypertarget{method-Local Storage-GetStoredValue}{}}}
\subsection{Method \code{GetStoredValue()}}{
Retrieve a value from the browser local storage. This does not take a reactive dependency.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$GetStoredValue(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
Name of the value to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Value if stored in local storage, NULL if not present.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Local Storage-GetStoredIds"></a>}}
\if{latex}{\out{\hypertarget{method-Local Storage-GetStoredIds}{}}}
\subsection{Method \code{GetStoredIds()}}{
Retrieve the IDs of all values in the browser local storage. This does not take a reactive dependency.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$GetStoredIds()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of store item IDs, or NULL if storage is not available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Local Storage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Local Storage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalStorage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
