% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{LongToWide}
\alias{LongToWide}
\title{Convert long format to wide format (including covariate columns)}
\usage{
LongToWide(long_data, outcome_type)
}
\arguments{
\item{long_data}{Data frame of long format}

\item{outcome_type}{Indicator whether outcome is 'Binary' or 'Continuous'}
}
\value{
Data frame in wide format
}
\description{
Convert long format to wide format (including covariate columns)
}
