% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{ma.redu}
\alias{ma.redu}
\title{Function to reduce the matrix size based on user defined plotting range reduction.}
\usage{
ma.redu(ma, rmt)
}
\arguments{
\item{ma}{Meta-analysis data for direct evidence only}

\item{rmt}{Matrix reduction vector}
}
\value{
Reduced meta-analysis data
}
\description{
Function to reduce the matrix size based on user defined plotting range reduction.
}
