% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{ma.sortres}
\alias{ma.sortres}
\title{Function to update the pairwise meta-analysis results after changes to the tx rankings.}
\usage{
ma.sortres(ma, mtc, mtorg, outcome_measure)
}
\arguments{
\item{ma}{Meta-analysis data for direct evidence only}

\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{mtorg}{Sorting order matrix}

\item{outcome_measure}{Type of outcome being plotted}
}
\value{
Updated meta-analysis data
}
\description{
Function to update the pairwise meta-analysis results after changes to the tx rankings.
}
