% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_Incon}
\alias{make_Incon}
\title{2c Creates a data frame of inconsistency data obtained from netmeta::netsplit().}
\usage{
make_Incon(freq, modelranfix)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{modelranfix}{"fixed" or "random".}
}
\value{
Data frame of inconsistency data created by netsplitresult.df().
}
\description{
2c Creates a data frame of inconsistency data obtained from netmeta::netsplit().
}
