% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netComp}
\alias{make_netComp}
\title{2a. Frequentist forest Plot}
\usage{
make_netComp(freq, modelranfix, ref, min, max)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{modelranfix}{"fixed" or "random".}

\item{ref}{Reference treatment.}

\item{min}{Minimum x-axis limit.}

\item{max}{Maximum x-axis limit.}
}
\value{
Forest plot created by forest.df().
}
\description{
2a. Frequentist forest Plot
}
