% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_network_f.R
\name{make_netconnect}
\alias{make_netconnect}
\title{Creates network connectivity info displayed under network plots}
\usage{
make_netconnect(freq)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}
}
\value{
Vector summarising network connectivity created by netmeta::netconnection().
}
\description{
Creates network connectivity info displayed under network plots
}
