% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netgraph}
\alias{make_netgraph}
\title{1c Network plot - number of trials on line}
\usage{
make_netgraph(freq, label_size)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{label_size}{Label size multiplier.}
}
\value{
Network plot created by netmeta::netgraph().
}
\description{
1c Network plot - number of trials on line
}
