% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netgraph_rank}
\alias{make_netgraph_rank}
\title{3c Create network plot for the ranking panel}
\usage{
make_netgraph_rank(freq, order)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{order}{Vector of treatments in SUCRA order.}
}
\value{
Network plot created by netmeta::netgraph() with each line containing the number of trials.
}
\description{
3c Create network plot for the ranking panel
}
