% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netplot}
\alias{make_netplot}
\title{1c Network plot - number of trials by nodesize and line thickness}
\usage{
make_netplot(bugsnetdt, label_size = 1, order = NULL)
}
\arguments{
\item{bugsnetdt}{Long format data.}

\item{label_size}{Node label size (default = 1).}

\item{order}{Order of nodes (default = NULL).}
}
\value{
Network plot created by BUGSnet::net.plot().
}
\description{
1c Network plot - number of trials by nodesize and line thickness
}
