% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netrank}
\alias{make_netrank}
\title{2b Treatment comparison and rank table}
\usage{
make_netrank(freq, modelranfix, rankopts)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{modelranfix}{"fixed" or "random".}

\item{rankopts}{"good" or "bad", referring to smaller outcome values.}
}
\value{
Ranking table created by netmeta::netleague().
}
\description{
2b Treatment comparison and rank table
}
