% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netStudy}
\alias{make_netStudy}
\title{1b Forest plot}
\usage{
make_netStudy(freq, outcome_measure, ForestHeader, ForestTitle)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}

\item{ForestHeader}{Multiplier for size of text in treatment contrast headers.}

\item{ForestTitle}{Multiplier for size of title.}
}
\value{
Plot created by groupforest.df().
}
\description{
1b Forest plot
}
