% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_refText}
\alias{make_refText}
\title{Create text to say outcomes are compared to the reference treatment.}
\usage{
make_refText(ref)
}
\arguments{
\item{ref}{Reference treatment.}
}
\value{
Text as above.
}
\description{
Create text to say outcomes are compared to the reference treatment.
}
