% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainsight-package.R
\name{metainsight-package}
\alias{metainsight-package}
\alias{metainsight}
\title{\emph{metainsight}: A flexible platform for meta-analysis}
\description{
Run the application via the
function \code{\link{run_metainsight}}
}
