% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{mtc.redu}
\alias{mtc.redu}
\title{Function to reduce the matrix size based on user defined plotting range reduction.}
\usage{
mtc.redu(mtc, rmt, p.only, po)
}
\arguments{
\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{rmt}{Matrix reduction vector}

\item{p.only}{Number of treatments to plot}

\item{po}{Matrix containing ranking order of treatments}
}
\value{
Reduced meta-analysis data
}
\description{
Function to reduce the matrix size based on user defined plotting range reduction.
}
