% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{mtc.sortres}
\alias{mtc.sortres}
\title{Function to update the MTC results after changes to the tx rankings.}
\usage{
mtc.sortres(mtc, mtorg, rkgmo, po, outcome_measure)
}
\arguments{
\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{mtorg}{Sorting order matrix}

\item{rkgmo}{Rankgram matrix sorting order}

\item{po}{Matrix containing ranking order of treatments}

\item{outcome_measure}{Type of outcome being plotted}
}
\value{
Updated meta-analysis data
}
\description{
Function to update the MTC results after changes to the tx rankings.
}
