% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{mtcMatrixCont}
\alias{mtcMatrixCont}
\title{Create summary forest matrix plot.}
\usage{
mtcMatrixCont(
  stytitle,
  ntx,
  lstx,
  mtc,
  ma,
  outcome_measure,
  bpredd = TRUE,
  bkey = TRUE,
  p.only = ntx,
  ucex = 1
)
}
\arguments{
\item{stytitle}{Title of plot}

\item{ntx}{Number of treatments}

\item{lstx}{Vector of treatment names}

\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{ma}{Meta-analysis data for direct evidence only}

\item{outcome_measure}{Type of outcome being plotted}

\item{bpredd}{TRUE if predictive interval to be plotted as error bars}

\item{bkey}{TRUE if key should be included in plot}

\item{p.only}{Number of treatments to plot}

\item{ucex}{Font size multiplier. Defaults to 1}
}
\description{
Create summary forest matrix plot.
}
