% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{multiplot}
\alias{multiplot}
\title{Draw the elements of the summary forest matrix plot.}
\usage{
multiplot(
  stytitle,
  ntx,
  lstx,
  mtc,
  ma,
  bpredd = TRUE,
  plt.adj,
  ucex,
  key_text = NULL
)
}
\arguments{
\item{stytitle}{Title of plot}

\item{ntx}{Number of treatments}

\item{lstx}{Vector of treatment names}

\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{ma}{Meta-analysis data for direct evidence only}

\item{bpredd}{TRUE if predictive interval to be plotted as error bars}

\item{plt.adj}{Plot position adjustment}

\item{ucex}{Text size}

\item{key_text}{Vector of length 3 containing strings for the key underneath the plot.}
}
\description{
Draw the elements of the summary forest matrix plot.
}
