% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{netsplitresult.df}
\alias{netsplitresult.df}
\title{Organises output from netmeta::netsplit() into a data frame.}
\usage{
netsplitresult.df(incona, model)
}
\arguments{
\item{incona}{Output from netmeta::netsplit().}

\item{model}{"fixed" or "random".}
}
\value{
Data frame of output from netmeta::netsplit()
\itemize{
\item 'Comparison': Treatment comparison.
\item 'No.Studies': Number of studies.
\item 'NMA': NMA treatment effect estimate.
\item 'Direct': Direct treatment effect estimate.
\item 'Indirect': Indirect treatment effect estimate.
\item 'Difference': Difference between treatment effects.
\item 'Diff_95CI_lower': 2.5\% limit of difference in treatment effects.
\item 'Diff_95CI_upper': 97.5\% limit of difference in treatment effects.
\item 'pValue': p-value for test of "difference in treatment effects == 0".
}
}
\description{
Organises output from netmeta::netsplit() into a data frame.
}
