% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{network.structure}
\alias{network.structure}
\title{Calculate edge.weights for network}
\usage{
network.structure(freq, order = NA)
}
\arguments{
\item{freq}{list. Output from \code{frequentist()}}

\item{order}{character. Vector of treatments names in rank order.}
}
\value{
data.frame containing the number of studies that compare each treatment against the reference treatment.
}
\description{
Calculate edge.weights for network
}
