% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buttons.R
\name{NodeDownloadHeight}
\alias{NodeDownloadHeight}
\title{Calculate download height of a node-split forest plot for a given number of comparisons}
\usage{
NodeDownloadHeight(ncomp, type)
}
\arguments{
\item{ncomp}{The number of comparisons in the plot}

\item{type}{Type of measure, "in" for inches, "px" for pixels}
}
\value{
The height of the plot
}
\description{
Calculate download height of a node-split forest plot for a given number of comparisons
}
