% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buttons.R
\name{NodePixels}
\alias{NodePixels}
\title{Calculate the pixel height of a node-split forest plot for a given number of comparisons for on screen}
\usage{
NodePixels(ncomp)
}
\arguments{
\item{ncomp}{The number of comparisons in the plot}
}
\value{
The height of the plot in pixels
}
\description{
Calculate the pixel height of a node-split forest plot for a given number of comparisons for on screen
}
