% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_numbering.R
\name{PageNumbering}
\alias{PageNumbering}
\title{Page Numbering}
\description{
Class for numbering pages in a logical manner.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{root}}{The top-level for numbering the entire app.}

\item{\code{current}}{The current numbering level.
Find the next identifier at the current level.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{root}}{The top-level for numbering the entire app.}

\item{\code{current}}{The current numbering level.
Find the next identifier at the current level.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PageNumbering-new}{\code{PageNumbering$new()}}
\item \href{#method-PageNumbering-AddChild}{\code{PageNumbering$AddChild()}}
\item \href{#method-PageNumbering-DiveLevel}{\code{PageNumbering$DiveLevel()}}
\item \href{#method-PageNumbering-FloatLevel}{\code{PageNumbering$FloatLevel()}}
\item \href{#method-PageNumbering-clone}{\code{PageNumbering$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageNumbering-new"></a>}}
\if{latex}{\out{\hypertarget{method-PageNumbering-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this \link[R6:R6Class]{R6} class.
Add a child at the current level. The pages will be numbered with:
\itemize{
\item Hindu-Arabic numerals (eg. 1, 2, 3, 4)
\item Latin alphabet letters (eg. a, b, c, d)
\item Roman numerals (eg. i, ii, iii, iv)
These will be arranged in the repeating format "1b iii-...".
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageNumbering$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageNumbering-AddChild"></a>}}
\if{latex}{\out{\hypertarget{method-PageNumbering-AddChild}{}}}
\subsection{Method \code{AddChild()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageNumbering$AddChild()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Formatted page number
Create a sub-level within the current numbering level.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageNumbering-DiveLevel"></a>}}
\if{latex}{\out{\hypertarget{method-PageNumbering-DiveLevel}{}}}
\subsection{Method \code{DiveLevel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageNumbering$DiveLevel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
Move up out of the current sub-level.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageNumbering-FloatLevel"></a>}}
\if{latex}{\out{\hypertarget{method-PageNumbering-FloatLevel}{}}}
\subsection{Method \code{FloatLevel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageNumbering$FloatLevel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PageNumbering-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PageNumbering-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PageNumbering$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
