% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{PairwiseTreatments}
\alias{PairwiseTreatments}
\title{Creates a data frame of values used to set y co-ordinates.}
\usage{
PairwiseTreatments(pairwise, treatment_order)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}

\item{treatment_order}{Vector of the treatments in the order they should appear in the forest plot.}
}
\value{
Data frame with columns:
\itemize{
\item 'treat1': First treatment in the comparison.
\item 'treat2': Second treatment in the comparison.
\item 'y_position_first': y co-ordinate of the lowest point on the graph corresponding to this comparison.
\item 'y_position_last': y co-ordinate of the highest point on the graph corresponding to this comparison.
}
}
\description{
Creates a data frame of values used to set y co-ordinates.
}
