% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scaling.R
\name{PlotDownloadHeight}
\alias{PlotDownloadHeight}
\title{Function to decide the height of pdf or png plot downloads, based on the number of rows in the data}
\usage{
PlotDownloadHeight(n_rows)
}
\arguments{
\item{nrows}{The number of rows in the data (a positive int)}
}
\value{
Number of pixels (a positive int)
}
\description{
Function to decide the height of pdf or png plot downloads, based on the number of rows in the data
}
