% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{PrepDataGemtc}
\alias{PrepDataGemtc}
\title{Function for formatting standard app user data ready for start of {gemtc} chain of commands}
\usage{
PrepDataGemtc(data, treatment_df, outcome_type, covariate, cov_friendly)
}
\arguments{
\item{data}{Uploaded data post-processing}

\item{treatment_df}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively}

\item{outcome_type}{Indicator of whether data is 'Binary' or 'Continuous'}

\item{covariate}{Chosen covariate name as per uploaded data}

\item{cov_friendly}{Friendly name of chosen covariate}
}
\value{
list containing two dataframes: armData containing the core data; studyData containing covariate data
}
\description{
Function for formatting standard app user data ready for start of {gemtc} chain of commands
}
