% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{PrICrI}
\alias{PrICrI}
\title{Function to draw two-tiered error bars for summary relative estimates.}
\usage{
PrICrI(
  offs,
  lower_confidence_limit,
  point_estimate,
  upper_confidence_limit,
  ypos,
  ucol = "black",
  ulwd = 1,
  pcI = FALSE,
  predbd = c(NA, NA)
)
}
\arguments{
\item{offs}{Axis offset}

\item{lower_confidence_limit}{Lower limit of confidence interval}

\item{point_estimate}{Point estimate}

\item{upper_confidence_limit}{Upper limit of confidence interval}

\item{ypos}{Position of error bars in Y axis}

\item{ucol}{Text colour. Defaults to black}

\item{ulwd}{Line width. Defaults to 1}

\item{pcI}{TRUE if both confidence interval and predictive interval to be plotted, otherwise only confidence interval plotted. Defaults to FALSE}

\item{predbd}{Predictive data}
}
\description{
Function to draw two-tiered error bars for summary relative estimates.
}
