% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rank_probs_table}
\alias{rank_probs_table}
\title{Combine probabilities and SUCRA from the list created by rankdata().}
\usage{
rank_probs_table(data)
}
\arguments{
\item{data}{List of ranking data created by rankdata().}
}
\value{
Data frame of probabilities and SUCRA.
}
\description{
Combine probabilities and SUCRA from the list created by rankdata().
}
