% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{ranking_plot}
\alias{ranking_plot}
\title{Treatment ranking plots}
\usage{
ranking_plot(
  ranking_data,
  style,
  colourblind = FALSE,
  simple = FALSE,
  regression_text = "",
  logger = NULL
)
}
\arguments{
\item{ranking_data}{list created by \code{baseline_ranking()}, \code{bayes_ranking()} or
\code{covariate_ranking()}.}

\item{style}{character. The style of plot to produce. Either \code{rankogram} or \code{radial}}

\item{colourblind}{logical. Whether to use a colourblind-friendly palette. Defaults to \code{FALSE}.}

\item{simple}{logical. Whether to display a simplified version of the radial plot. Does
not affect the rankogram plot. Defaults to \code{FALSE}}

\item{regression_text}{Text to show for regression. Defaults to no text.}
}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Produce either a rankogram or radial SUCRA plot ranking the treatments
}
