% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{ranking_table}
\alias{ranking_table}
\title{Ranking probability table}
\usage{
ranking_table(ranking_data)
}
\arguments{
\item{ranking_data}{list created by bayes_ranking().}
}
\value{
dataframe
}
\description{
Ranking probability table
}
