% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{RankingOrder}
\alias{RankingOrder}
\title{Returns the default ranking order, for the example datasets.}
\usage{
RankingOrder(outcome, data)
}
\arguments{
\item{outcome}{"binary" or "continuous".}

\item{data}{Input dataset.}
}
\value{
"good" or "bad".
}
\description{
Returns the default ranking order, for the example datasets.
}
