% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{rankogram}
\alias{rankogram}
\title{Function to draw graphs along diagonal - for NMA SPF Matrix.}
\usage{
rankogram(ntx, rkgram, cumu = FALSE)
}
\arguments{
\item{ntx}{Number of treatments}

\item{rkgram}{Vector containing rankogram data}

\item{cumu}{True if to draw cumulative rankogram. Defaults to False}
}
\description{
Function to draw graphs along diagonal - for NMA SPF Matrix.
}
