% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{redu.matrix}
\alias{redu.matrix}
\title{Function to create the matrix reduction vector.}
\usage{
redu.matrix(ntx, po, p.only, mtorg)
}
\arguments{
\item{ntx}{Number of treatments}

\item{po}{Matrix containing ranking order of treatments}

\item{p.only}{Number of treatments to plot}

\item{mtorg}{Sorting order matrix}
}
\value{
Matrix reduction vector
}
\description{
Function to create the matrix reduction vector.
}
