% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{ReinstateTreatmentIds}
\alias{ReinstateTreatmentIds}
\title{Replace all of the treatment IDs in the data with names, both for long and wide formats.}
\usage{
ReinstateTreatmentIds(data, treatment_ids)
}
\arguments{
\item{data}{Data frame in which to search for treatment IDs.}

\item{treatment_ids}{Data frame containing treatment names (Label) and IDs (Number).}
}
\value{
Data frame where the treatments are given as names, not IDs.
}
\description{
Replace all of the treatment IDs in the data with names, both for long and wide formats.
}
