% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{RemoveCovariates}
\alias{RemoveCovariates}
\title{Create a copy of a data from which does not contain any covariate columns.}
\usage{
RemoveCovariates(data)
}
\arguments{
\item{data}{Data from which to remove covariate columns}
}
\value{
Data without covariate columns
}
\description{
Create a copy of a data from which does not contain any covariate columns.
}
