% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{ReorderColumns}
\alias{ReorderColumns}
\title{Reorder data frame columns to the correct order, both for long and wide formats.}
\usage{
ReorderColumns(data, outcome_type)
}
\arguments{
\item{data}{Data frame to reorder}

\item{outcome_type}{Type of outcome for which to reorder, either 'continuous' or 'binary'}
}
\value{
Data frame with columns reordered
}
\description{
Reorder data frame columns to the correct order, both for long and wide formats.
}
