% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_cinema_f.R
\name{rep_cinema}
\alias{rep_cinema}
\title{rep_cinema}
\usage{
rep_cinema(configured_data, gemtc_results = NULL, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{gemtc_results}{Output from gemtc::mtc.run, as returned in the 'mtcResults' list element from baye(). If this parameter is NULL then the frequentist analysis results found in 'contributions' are used. If it is not NULL then the Bayesian analysis results contained in this parameter are used. Defaults to NULL.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
JSON string with the following structure:
A named list of lists. The top level list contains items:
\itemize{
\item "project" Information for CINeMA project
\itemize{
\item "CM" Contribution matrices
\itemize{
\item "contributionMatrices" output from \code{.PrepareAnalysisForCinema()}
}
\item "format" Data format. Always "long"
\item "type" Outcome type. Either "binary" or "continuous"
\item "Studies" Study data
\itemize{
\item "long" Output from \code{.PrepareDataForCinema()}
}
}
}
}
\description{
Prepare project into a JSON format that CINeMA can read.
}
