% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{ReplaceTreatmentIds}
\alias{ReplaceTreatmentIds}
\title{Find all of the treatment names in the data, both for long and wide formats.}
\usage{
ReplaceTreatmentIds(data, treatment_ids)
}
\arguments{
\item{data}{Data frame in which to search for treatment names}

\item{treatent_ids}{Data frame containing treatment names (Label) and IDs (Number)}
}
\value{
Data frame where the treatments are given as IDs, not names
}
\description{
Find all of the treatment names in the data, both for long and wide formats.
}
