% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_update_f.R
\name{ReplaceTreatments}
\alias{ReplaceTreatments}
\title{Replace treatment IDs with treatment names.}
\usage{
ReplaceTreatments(old_data, treatment_df)
}
\arguments{
\item{old_data}{Data frame with old data format where treatments are specified as IDs}

\item{treatment_df}{Data frame with 'Number' column defining treatment IDs and 'Label' column defining treatment names}
}
\value{
Data frame in the new format with treatment names
}
\description{
Replace treatment IDs with treatment names.
}
