% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{return-svg}
\alias{return-svg}
\title{Return svg}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Return svg
}
\keyword{internal}
