% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_details_f.R
\name{RoundForDisplay}
\alias{RoundForDisplay}
\title{Returns a number rounded in a nice manner.}
\usage{
RoundForDisplay(number)
}
\arguments{
\item{number}{Any number.}
}
\value{
'number' with 1 decimal place or 1 significant figure.
}
\description{
Returns a number rounded in a nice manner.
}
