% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{run_all}
\alias{run_all}
\title{run_all}
\usage{
run_all(component, logger)
}
\arguments{
\item{component}{character. The component to run all the modules of.}

\item{logger}{common$logger}
}
\description{
For internal use. Runs all modules for a given component, excluding the model and nodesplit modules.
}
\keyword{internal}
