% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_deviance_f.R
\name{scat_plot}
\alias{scat_plot}
\title{UME scatter plot}
\usage{
scat_plot(model, deviance, model_type, outcome_measure, seed)
}
\arguments{
\item{model}{Bayesian model produced by bayes_model}

\item{deviance}{Output produced by \code{gemtc::mtc.deviance()}}

\item{model_type}{Model effects type. "random" or "fixed".}

\item{outcome_measure}{Outcome measure being analysed: one of "OR". "RR", "MD".}

\item{seed}{numeric. Seed value to use for calculating UME model.}
}
\description{
UME scatter plot
}
