% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{setup_configure_table}
\alias{setup_configure_table}
\title{Summarise the analysis configuration}
\usage{
setup_configure_table(configured_data)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}
}
\description{
Create a table summarising how the analysis has been configured
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

setup_configure_table(configured_data)

}
