% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{setup_load}
\alias{setup_load}
\title{Load data}
\usage{
setup_load(data_path = NULL, outcome, logger = NULL)
}
\arguments{
\item{data_path}{character. Path to the file (either a \code{.csv} or \code{.xlsx})
to be loaded or if \code{NULL} load the default data}

\item{outcome}{character. Outcome type for the dataset. Either \code{binary} or
\code{continuous}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
List containing:
\item{is_data_valid}{logical. Whether the data is valid}
\item{is_data_uploaded}{logical. Whether the data is uploaded}
\item{data}{dataframe. The data that was uploaded or the default data if
no data_path was provided}
\item{treatments}{Dataframe of the treatments in the data. \code{NULL} if
\code{is_data_valid} is \code{FALSE}}
\item{outcome}{character. Whether the data is \code{binary} or \code{continuous}}
}
\description{
Load data from a spreadsheet or a default dataset and assess the
data for validity. This checks the column names for required columns and
balanced wide format numbered columns. Data can be in either a long or wide
format; long data has one row per study arm whereas wide data has one row
per study. For continuous outcomes, long data should contain the columns:
\code{Study} - an identifier, e.g. author and year, \code{T} - treatment,
\code{N} - number of participants, \code{Mean} - mean value of the outcome,
\code{SD} - standard deviation of the outcome. Wide data for continuous outcomes
should contain: \code{Study}, \code{N.1}, \code{N.2}, \code{Mean.1}, \code{Mean.2}, \code{SD.1}, \code{SD.2}
where the number refers to the arm of the study and extra columns should be
added depending on the number of arms. For binary outcomes, long data should
contain: \code{Study}, \code{T}, \code{N} (as for continuous data) and \code{R} - the number of
participants with the outcome of interest. Wide data for binary outcomes
should follow the same convention: \code{Study}, \code{T.1}, \code{T.2}, \code{R.1}, \code{R.2},
\code{N.1}, \code{N.2}. Additionally, a \verb{covar.<name>} column can be added to all
formats containing covariate data where \verb{<name>} should be replaced with
the name of the covariate. For long data, covariate values must be equal
for every study arm. Risk of bias data can also be included with all columns
containing values ranging from 1 (low risk) to 3 (high risk): \code{rob} for the
overall risk of bias, \code{indirectness} for indirectness and \verb{rob.<name>} for up
to ten individual components.
}
\examples{
# load data from a file
minimal_data_path <- system.file("extdata", "continuous_minimal.csv", package = "metainsight")
loaded_data <- setup_load(data_path = minimal_data_path,
                          outcome = "continuous")

# load default data
loaded_data <- setup_load(outcome = "binary")

}
