% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{SetupRegressionPlotColours}
\alias{SetupRegressionPlotColours}
\title{Setup the main components of the plot panel.}
\usage{
SetupRegressionPlotColours(
  plot,
  comparators,
  include_ghosts,
  include_confidence,
  confidence_opacity = 1
)
}
\arguments{
\item{comparators}{Vector of names of comparison treatments to plot.}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot.}

\item{include_confidence}{TRUE if all other comparator studies should be plotted in grey in the background of the plot.}

\item{confidence_opacity}{The opacity of the confidence regions. Can be any value between 0 and 1, inclusive. Defaults to 1.}

\item{reference}{Name of the reference treatment.}
}
\value{
Created ggplot2 object.
}
\description{
Setup the main components of the plot panel.
}
