% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{shading.vec}
\alias{shading.vec}
\title{Function to create a vector indexing shading in mtcMultiplot.}
\usage{
shading.vec(ntx)
}
\arguments{
\item{ntx}{Number of treatments}
}
\value{
the calculated shading vector
}
\description{
Function to create a vector indexing shading in mtcMultiplot.
}
