% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{ShortenRobNames}
\alias{ShortenRobNames}
\title{Shorten RoB and indirectness variable names.}
\usage{
ShortenRobNames(long_names)
}
\arguments{
\item{long_names}{Vector of risk of bias and indirectness variable names.}
}
\value{
Vector of the same length with shortened names.
}
\description{
Shorten RoB and indirectness variable names.
}
