% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{singleest}
\alias{singleest}
\title{MTC & MA estimates for Forest Matrix plot - used in for loop in multiplot function.}
\usage{
singleest(mtc, pw, xpos = 0, ucex)
}
\arguments{
\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{pw}{Meta-analysis data for direct (pairwise) evidence only}

\item{xpos}{Position of text in X axis}

\item{ucex}{Text size}
}
\description{
MTC & MA estimates for Forest Matrix plot - used in for loop in multiplot function.
}
