% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{singleSFP}
\alias{singleSFP}
\title{Single plot for MTC & MA SFP in NMA SPF Matrix.}
\usage{
singleSFP(mtc, pw, bpredd = TRUE, baxis = TRUE, scaletype)
}
\arguments{
\item{mtc}{Meta-analysis data for direct and indirect evidence}

\item{pw}{Meta-analysis data for direct (pairwise) evidence only}

\item{bpredd}{TRUE if predictive interval to be plotted as error bars}

\item{baxis}{TRUE if axes to be drawn for forest plots}

\item{scaletype}{The outcome type being plotted. "RR" for risk ratio, and "OR" for odds ratio will be plotted on a log scale, anything else will be plotted on a linear scale}
}
\description{
Single plot for MTC & MA SFP in NMA SPF Matrix.
}
