% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{SortByStudyIDThenT}
\alias{SortByStudyIDThenT}
\title{Sort the data by StudyID then T}
\usage{
SortByStudyIDThenT(data, outcome)
}
\arguments{
\item{data}{Data frame to sort}

\item{outcome}{Type of outcome for which to reorder, either 'Continuous' or 'Binary'}
}
\value{
Data frame ordered by StudyID, then T if applicable
}
\description{
Sort the data by StudyID then T
}
