% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{SortLong}
\alias{SortLong}
\title{Sort long data by StudyID then T}
\usage{
SortLong(long_data)
}
\arguments{
\item{long_data}{Data in long format.}
}
\value{
Long data sorted by StudyID then T.
}
\description{
Sort long data by StudyID then T
}
