% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{sortres.matrix}
\alias{sortres.matrix}
\title{Function to create the matrix sorting order - to be used for the MTC & MA numerical results (Upper triangle results),}
\usage{
sortres.matrix(ntx, po)
}
\arguments{
\item{ntx}{Number of treatments}

\item{po}{Ranking order of treatments}
}
\value{
Sorting order matrix
}
\description{
Function to create the matrix sorting order - to be used for the MTC & MA numerical results (Upper triangle results),
}
