% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{sortrkg.ord}
\alias{sortrkg.ord}
\title{Function to create the rankgram matrix sorting order.}
\usage{
sortrkg.ord(ntx, po)
}
\arguments{
\item{ntx}{Number of treatments}

\item{po}{Ranking order of treatments}
}
\value{
Rankgram matrix sorting order
}
\description{
Function to create the rankgram matrix sorting order.
}
