% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{stemplot}
\alias{stemplot}
\title{Stemplot}
\usage{
stemplot(model, package = "gemtc")
}
\arguments{
\item{model}{Various model output created by baye().}
}
\value{
Stemplot created by stemplot.df().
}
\description{
Stemplot
}
