% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_exclusions_panel.R
\name{study_exclusions_panel_server}
\alias{study_exclusions_panel_server}
\title{Module server for the study exclusion panel.}
\usage{
study_exclusions_panel_server(id, data, treatment_df, reference_treatment)
}
\arguments{
\item{id}{ID of the module.}

\item{data}{Reactive containing data to analyse.}

\item{treatment_df}{Reactive containing data frame containing treatment IDs (Number) and names (Label).}

\item{reference_treatment}{Reactive containing the ID of the selected reference treatment.}
}
\value{
List of reactives:
\itemize{
\item "initial_data" contains the data frame of the connected subnetwork of the uploaded data
\item "sensitivity_data" contains the data frame of the connected subnetwork for the sensitivity analysis
\item "sensitivity_treatment_list" contains the treatment names ("Label") and IDs ("Number") for the sensitivity analysis data
}
}
\description{
Module server for the study exclusion panel.
}
