% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_exclusions_panel.R
\name{study_exclusions_panel_ui}
\alias{study_exclusions_panel_ui}
\title{Module UI for the study exclusion panel.}
\usage{
study_exclusions_panel_ui(id)
}
\arguments{
\item{id}{ID of the module.}
}
\value{
Div for the panel.
}
\description{
Module UI for the study exclusion panel.
}
