% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_char_f.R
\name{summary_char}
\alias{summary_char}
\title{Characterise the data}
\usage{
summary_char(configured_data, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
List containing:
\item{network}{network characteristics}
\item{treatments}{treatment characteristics}
\item{pairs}{treatment pair characteristics}
}
\description{
Produce summaries of network characteristics, treatments and
treatment pairs such as the number of participants and and the mean outcome.
Inspired by \code{BUGSnet::net.tab()}
}
