% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_exclude_f.R
\name{summary_exclude}
\alias{summary_exclude}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
summary_exclude(
  data,
  treatment_df,
  reference_treatment,
  metaoutcome,
  outcome_measure,
  model_type,
  exclusions,
  logger = NULL
)
}
\arguments{
\item{data}{dataframe. Uploaded data}

\item{treatment_df}{vector of treatments}

\item{reference_treatment}{character. The reference treatment}
}
\value{
list
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
