% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_network_f.R
\name{summary_network}
\alias{summary_network}
\title{Produce a plot of the network using \code{netmeta::netgraph()}}
\usage{
summary_network(
  configured_data,
  style,
  label_size = 1,
  title = "",
  logger = NULL
)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{style}{character. The plot to produce, either \code{netgraph} or \code{netplot}}

\item{label_size}{numeric. The size of labels in the plots. Default of \code{1}.}

\item{title}{character. Title of plot. Default of no title.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This will display
the plot when at the end of a quarto or rmarkdown chunk. To view in the viewer
panel of Rstudio, use \code{htmltools::browsable()}. The output can be saved
using \code{write_plot()}.
}
\description{
Produce a plot of the network using \code{netmeta::netgraph()}
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

summary_network(configured_data = configured_data,
                style = "netgraph")

}
