% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study}
\alias{summary_study}
\title{Create a forest plot of pairwise comparisons, grouped by treatment pairs.}
\usage{
summary_study(
  connected_data,
  freq,
  outcome_measure,
  plot_area_width = 6,
  colourblind = FALSE,
  x_min = NULL,
  x_max = NULL,
  logger = NULL
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{freq}{list. NMA results created by freq_wrap().}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{plot_area_width}{numeric. The width of the plot area containing the
treatment effects in inches. Defaults to \code{6}.}

\item{colourblind}{logical. Whether to use a colourblind-friendly palette. Defaults to \code{FALSE}}

\item{x_min}{numeric. Minimum value for the x-axis. Defaults to \code{NULL}}

\item{x_max}{numeric. Maximum value for the x-axis. Defaults to \code{NULL}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Create a forest plot of pairwise comparisons, grouped by treatment pairs.
}
