% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study_min_max}
\alias{summary_study_min_max}
\title{Find the default minimum and maximum values for the x-axis and calculate
a sensible step to use in the numeric input}
\usage{
summary_study_min_max(pairwise, outcome)
}
\arguments{
\item{pairwise}{Results of pairwise analysis}

\item{outcome}{character. Outcome type for the dataset. Either \code{binary} or
\code{continuous}.}
}
\value{
Vector of xmin and xmax
}
\description{
Find the default minimum and maximum values for the x-axis and calculate
a sensible step to use in the numeric input
}
