% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summary_table_plot}
\alias{summary_table_plot}
\title{1a Summary table plot}
\usage{
summary_table_plot(bugsnetdt, metaoutcome)
}
\arguments{
\item{bugsnetdt}{Long format data.}

\item{metaoutcome}{"Continuous" or "Binary".}
}
\value{
Network created by BUGSnet::net.tab().
}
\description{
1a Summary table plot
}
