% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{suppress_jags_output}
\alias{suppress_jags_output}
\title{suppress_jags_output}
\usage{
suppress_jags_output(expr)
}
\arguments{
\item{expr}{The model function to be used}
}
\description{
For internal use. Stop JAGS output appearing in the console /
html reports by wrapping model functions
}
\keyword{internal}
