% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{svg_container}
\alias{svg_container}
\title{Put an svg in a container with buttons for fullscreen}
\usage{
svg_container(svg, class = "svg_container")
}
\arguments{
\item{svg}{character. Output from \code{crop_svg()}}

\item{class}{character. Class of the container default \code{svg_container}}
}
\description{
Put an svg in a container with buttons for fullscreen
}
\keyword{internal}
