% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{tau.df}
\alias{tau.df}
\title{Creates the text to be displayed underneath the forest plots, with between-study SD, number of studies and number of treatments.}
\usage{
tau.df(tau, k, n, model, outcome)
}
\arguments{
\item{tau}{Between study standard deviation.}

\item{k}{Number of studies.}

\item{n}{Number of treatments.}

\item{model}{"fixed" or "random".}

\item{outcome}{"MD", "SMD", "OR", "RR", or "RD".}
}
\value{
Text as described above.
}
\description{
Creates the text to be displayed underneath the forest plots, with between-study SD, number of studies and number of treatments.
}
