% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{texttau}
\alias{texttau}
\title{2a. Creates text displayed under frequentist forest plots}
\usage{
texttau(freq, outcome_measure, modelranfix)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}

\item{modelranfix}{"fixed" or "random".}
}
\value{
Text created by tau.df().
}
\description{
2a. Creates text displayed under frequentist forest plots
}
