% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{trace_plots}
\alias{trace_plots}
\title{Creates trace plots of MCMC samples.}
\usage{
trace_plots(model, parameters)
}
\arguments{
\item{model}{Model output.}

\item{parameters}{Vector of parameters to create trace plots for.}
}
\value{
List of ggplot trace plots.
}
\description{
Creates trace plots of MCMC samples.
}
\examples{
configured_data_path <- system.file("extdata", "configured_data.Rds", package = "metainsight")
configured_data <- readRDS(configured_data_path)

# n_adapt and n_iter are set low to run quickly, but should be left as the
# default values in real use

fitted_bayes_model <- bayes_model(configured_data = configured_data,
                                  n_adapt = 100,
                                  n_iter = 100)

mcmc <- bayes_mcmc(model = fitted_bayes_model)

trace_plots(fitted_bayes_model$mtcResults, mcmc$parameters)

}
