% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_guide_page.R
\name{user_guide_page_server}
\alias{user_guide_page_server}
\title{Module server for the user guide page.}
\usage{
user_guide_page_server(id)
}
\arguments{
\item{id}{ID of the module}
}
\description{
Module server for the user guide page.
}
