% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_guide_page.R
\name{user_guide_page_ui}
\alias{user_guide_page_ui}
\title{Module UI for the user guide page.}
\usage{
user_guide_page_ui(id)
}
\arguments{
\item{id}{ID of the module}
}
\value{
Div for the user guide page
}
\description{
Module UI for the user guide page.
}
