% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{ValidateUploadedData}
\alias{ValidateUploadedData}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
ValidateUploadedData(data, outcome)
}
\arguments{
\item{data}{Data frame to validate.}

\item{outcome}{Outcome type selected for the data. Either "Binary" or "Continuous".}
}
\value{
Validation result in the form of a list:
\itemize{
\item "valid" = TRUE or FALSE defining whether data is valid
\item "message" = String describing any issues causing the data to be invalid
}
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
