% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{VectorWithItemFirst}
\alias{VectorWithItemFirst}
\title{Create a copy of a vector with the given item as the first element.}
\usage{
VectorWithItemFirst(vector, first_item)
}
\arguments{
\item{vector}{Vector to reorder}

\item{first_item}{The element to push to the front of the vector}
}
\value{
The reordered vector
}
\description{
Create a copy of a vector with the given item as the first element.
}
