% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{WideToLong}
\alias{WideToLong}
\title{Convert wide format to long format (including covariate columns).}
\usage{
WideToLong(wide_data, outcome)
}
\arguments{
\item{wide_data}{Data frame of wide format}

\item{outcome}{Indicator whether outcome is 'binary' or 'continuous'}
}
\value{
Data frame in long format
}
\description{
Convert wide format to long format (including covariate columns).
}
\keyword{internal}
