% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{WrangleUploadData}
\alias{WrangleUploadData}
\title{Wrangle the uploaded data into a form usable by the internals of the app, both for long and wide formats.}
\usage{
WrangleUploadData(data, treatment_ids, outcome_type)
}
\arguments{
\item{data}{Data frame to wrangle}

\item{treatment_ids}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively}

\item{outcome_type}{Type of outcome for which to reorder, either 'Continuous' or 'Binary'}
}
\value{
Data frame which is uasable by the rest of the app
}
\description{
Wrangle the uploaded data into a form usable by the internals of the app, both for long and wide formats.
}
