% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_svg_plot}
\alias{write_svg_plot}
\title{Write an svg plot to either a png, pdf or svg file}
\usage{
write_svg_plot(file, type, svg)
}
\arguments{
\item{file}{character. The file to which to write.}

\item{type}{character. Type of file to which to write.}

\item{svg}{list. Containing the svg string, width and height returned from \code{crop_svg()}}
}
\description{
Write an svg plot to either a png, pdf or svg file
}
