% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buttons.R
\name{write_to_pdf_or_png}
\alias{write_to_pdf_or_png}
\title{Write some plot to a .pdf or .png file.}
\usage{
write_to_pdf_or_png(file, type, renderFunction, height = NULL, width = NULL)
}
\arguments{
\item{file}{The file to which to write.}

\item{type}{String containing the type of file to which to write.}

\item{renderFunction}{A function to render the plot.}

\item{height}{The height of the plot in inches for pdf, or user specified units for png.}

\item{width}{The width of the plot in inches for pdf, or user specified units for png.}
}
\description{
Write some plot to a .pdf or .png file.
}
