% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{writeLog}
\alias{writeLog}
\title{writeLog}
\usage{
writeLog(logger, ..., type = "default", go_to = NULL)
}
\arguments{
\item{logger}{The logger to write the text to. Can be NULL or a function}

\item{...}{Messages to write to the logger}

\item{type}{One of "default", "info", "error", "warning"}

\item{go_to}{character. The id of a module to navigate to when the modal is closed.
Only used when \code{type = "error"}}
}
\description{
For internal use. Add text to a logger
}
\keyword{internal}
