% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_comparison_f.R
\name{baseline_comparison}
\alias{baseline_comparison}
\title{baseline_comparison}
\usage{
baseline_comparison(model, logger = NULL)
}
\arguments{
\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}

\item{x}{x}
}
\description{
Does x
}
\examples{
{
y <- baseline_comparison(1)
}
}
