% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_deviance_f.R
\name{baseline_deviance}
\alias{baseline_deviance}
\title{Produce deviance plots for baseline models}
\usage{
baseline_deviance(model, async = FALSE)
}
\arguments{
\item{model}{Output model produced by \code{baseline_model()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
list containing:
\item{deviance_mtc}{equivalent summary to that produced by \code{gemtc::mtc.deviance()}}
\item{stem_plot}{plotly object}
\item{lev_plot}{plotly object}
}
\description{
Produce deviance plots for baseline models
}
