% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_deviance_f.R
\name{baseline_deviance}
\alias{baseline_deviance}
\title{Produce deviance plots for baseline models}
\usage{
baseline_deviance(model, logger = NULL)
}
\arguments{
\item{model}{Baseline model produced by \code{baseline_model()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
list containing:
\item{deviance_mtc}{equivalent summary to that produced by \code{gemtc::mtc.deviance()}}
\item{stem_plot}{plotly object}
\item{lev_plot}{plotly object}
}
\description{
Produce deviance plots for baseline models
}
\examples{
{
y <- baseline_deviance(1)
}
}
