% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_forest_f.R
\name{baseline_forest}
\alias{baseline_forest}
\title{baseline_forest}
\usage{
baseline_forest(
  model,
  treatment_df,
  xmin = NULL,
  xmax = NULL,
  title = "Baseline risk regression analysis",
  ranking = FALSE,
  logger = NULL
)
}
\arguments{
\item{model}{model output produced by \code{baseline_model()}}

\item{treatment_df}{dataframe. Treatments}

\item{xmin}{numeric. Minimum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{xmax}{numeric. Maximum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{title}{character. Title for the plot. Defaults to \verb{Baseline risk regression analysis}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Produce a forest plot for a baseline risk model using \code{bnma::network.forest.plot()}
}
