% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{baseline_model}
\alias{baseline_model}
\title{Fit a baseline risk regression model using \code{bnma::network.run()}}
\usage{
baseline_model(configured_data, regressor_type, async = FALSE)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{regressor_type}{character. Type of regression coefficient, either \code{shared}, \code{unrelated}, or \code{exchangeable}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List of bnma related output:
\item{mtcResults}{model object itself carried through (needed to match existing code)}
\item{covariate_value}{The mean covariate value, used for centring}
\item{reference_treatment}{character. The \code{reference_treatment}from \code{configured_data}}
\item{comparator_names}{Vector containing the names of the comparators}
\item{a}{text output stating whether fixed or random effects}
\item{cov_value_sentence}{text output stating the value for which the covariate has been set to for producing output}
\item{slopes}{named list of slopes for the regression equations (unstandardised - equal to one 'increment')}
\item{intercepts}{named list of intercepts for the regression equations at cov_value}
\item{outcome}{character. The \code{outcome} from \code{configured_data}}
\item{outcome_measure}{character. The \code{outcome_measure}from \code{configured_data}}
\item{effects}{character. The \code{effects} from \code{configured_data}}
\item{covariate_min}{Vector of minimum covariate values directly contributing to the regression}
\item{covariate_max}{Vector of maximum covariate values directly contributing to the regression}
\item{dic}{Summary of model fit}
\item{sumresults}{Output of summary(model)}
\item{regressor}{Type of regression coefficient}
}
\description{
Fit a baseline risk regression model using \code{bnma::network.run()}
}
