% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{baseline_model}
\alias{baseline_model}
\title{baseline_model}
\usage{
baseline_model(
  connected_data,
  treatment_df,
  outcome,
  reference_treatment,
  model_type,
  cov_parameters,
  seed,
  logger = NULL
)
}
\arguments{
\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}

\item{x}{x}
}
\description{
Does x
}
\examples{
{
y <- baseline_model(1)
}
}
