% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_model_f.R
\name{baseline_model}
\alias{baseline_model}
\title{baseline_model}
\usage{
baseline_model(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  reference_treatment,
  model_type,
  regressor_type,
  seed,
  async = FALSE
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{treatment_df}{dataframe. Treatments}

\item{outcome}{character. Outcome type for the dataset. Either \code{Binary} or
\code{Continuous}.}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{regressor_type}{character. Type of regression coefficient, either \code{shared}, \code{unrelated}, or \code{exchangeable}}

\item{seed}{numeric. Seed used to fit the model}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List of bnma related output:
mtcResults = model object itself carried through (needed to match existing code).
covariate_value = The mean covariate value, used for centring.
reference_name = The name of the reference treatment.
comparator_names = Vector containing the names of the comparators.
a = text output stating whether fixed or random effects.
cov_value_sentence = text output stating the value for which the covariate has been set to for producing output.
slopes = named list of slopes for the regression equations (unstandardised - equal to one 'increment').
intercepts = named list of intercepts for the regression equations at cov_value.
outcome = The outcome type for the analysis eg. "MD" or "OR".
model = effects type, "fixed" or "random".
covariate_min = Vector of minimum covariate values directly contributing to the regression.
covariate_max = Vector of maximum covariate values directly contributing to the regression.
dic = Summary of model fit
summary = Output of summary(model)
regressor = regressor_type
}
\description{
Fit a baseline risk regression model with \code{bnma::network.run()}
}
