% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_regression_f.R
\name{baseline_regression}
\alias{baseline_regression}
\title{baseline_regression}
\usage{
baseline_regression(
  model_output,
  connected_data,
  covariate_title,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  async = FALSE
)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- baseline_regression(1)
}
}
