% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_regression_f.R
\name{baseline_regression}
\alias{baseline_regression}
\title{Generate data required to produce a metaregression plot
for a baseline risk model.}
\usage{
baseline_regression(model, configured_data, async = FALSE)
}
\arguments{
\item{model}{Output produced by \code{baseline_model()}}

\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{directness}{list. Output from \code{CalculateDirectness()}}
\item{credible_regions}{list. Output from \code{CalculateCredibleRegions()}}
}
\description{
Generate data required to produce a metaregression plot
for a baseline risk model.
}
