% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_summary_f.R
\name{baseline_summary}
\alias{baseline_summary}
\title{Produce a plot summarising baselink risk for each study arm}
\usage{
baseline_summary(configured_data, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\description{
Produce a plot summarising baselink risk for each study arm
}
