% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_compare_f.R
\name{bayes_compare}
\alias{bayes_compare}
\alias{covariate_comparison}
\title{Produce a table of comparisons of all treatment pairs for Bayesian models
using \code{gemtc::relative.effect.table()}}
\usage{
bayes_compare(model, logger = NULL)

covariate_comparison(...)
}
\arguments{
\item{model}{list. Object created by \code{bayes_model()} or \code{covariate_model()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{...}{Parameters passed to \code{bayes_compare()}}
}
\value{
Relative effects table
}
\description{
Produce a table of comparisons of all treatment pairs for Bayesian models
using \code{gemtc::relative.effect.table()}
}
