% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_compare_f.R
\name{bayes_compare}
\alias{bayes_compare}
\title{Creates a table of comparisons of all treatment pairs}
\usage{
bayes_compare(model, logger = NULL)
}
\arguments{
\item{model}{Model output created by bayes_model().}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
Relative effects table created by gemtc::relative.effect.table().
}
\description{
Creates a table of comparisons of all treatment pairs
}
