% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_compare_f.R
\name{bayes_compare}
\alias{bayes_compare}
\title{Creates a table of comparisons of all treatment pairs}
\usage{
bayes_compare(model, outcome_measure)
}
\arguments{
\item{model}{Various model output created by baye().}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}
}
\value{
Relative effects table created by gemtc::relative.effect.table().
}
\description{
Creates a table of comparisons of all treatment pairs
}
