% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_deviance_f.R
\name{bayes_deviance}
\alias{bayes_deviance}
\title{Produce deviance plots}
\usage{
bayes_deviance(model, model_type, outcome_measure)
}
\arguments{
\item{model}{Bayesian model produced by bayes_model}

\item{model_type}{Model effects type. "random" or "fixed".}

\item{outcome_measure}{Outcome measure being analysed: one of "OR". "RR", "MD".}
}
\value{
list containing:
 \item{scat_plot}{plotly object}
 \item{stem_plot}{plotly object}
 \item{lev_plot}{plotly object}
}
\description{
Produce deviance plots
}
