% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_forest_f.R
\name{bayes_forest}
\alias{bayes_forest}
\title{Make a Bayesian forest plot}
\usage{
bayes_forest(model, logger = NULL)
}
\arguments{
\item{model}{list Objected created by bayes_model()}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
None - produces a base plot
}
\description{
Make a Bayesian forest plot
}
