% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_forest_f.R
\name{bayes_forest}
\alias{bayes_forest}
\title{Make a Bayesian forest plot}
\usage{
bayes_forest(
  model,
  treatment_df,
  reference_treatment,
  title,
  ranking = FALSE,
  logger = NULL
)
}
\arguments{
\item{model}{list. Object created by bayes_model()}

\item{treatment_df}{dataframe.}

\item{reference_treatment}{character.}

\item{title}{character. Title for the plot}

\item{ranking}{logical. Whether the function is being used in `bayes_ranking`}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
List containing:
 \item{svg}{character. SVG code to produce the plot}
 \item{height}{numeric. Plot height in pixels}
 \item{width}{numeric. Plot width in pixels}
}
\description{
Make a Bayesian forest plot
}
