% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{bayes_mcmc}
\alias{bayes_mcmc}
\title{bayes_mcmc}
\usage{
bayes_mcmc(model, logger = NULL)
}
\arguments{
\item{model}{Output from `bayes_model()`}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
list containing:
\item{gelman_plots}{Gelman plots}
\item{trace_plots}{Gelman plots}
\item{density_plots}{Gelman plots}
\item{n_rows}{The number of rows for each plot}
}
\description{
Produces Markov chain Monte Carlo plots
}
