% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model_f.R
\name{bayes_model}
\alias{bayes_model}
\title{Fit a Bayesian model using \CRANpkg{gemtc}}
\usage{
bayes_model(configured_data, async = FALSE)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{mtcResults}{mtc.result. Output from \code{gemtc::mtc.run()}}
\item{mtcRelEffects}{mtc.result. Output from \code{gemtc::relative.effect()}}
\item{rel_eff_tbl}{mtc.relative.effect.table. Output from \code{gemtc::relative.effect.table()}}
\item{sumresults}{summary.mtc.result. Output from \code{summary(mtcRelEffects)}}
\item{mtcNetwork}{mtc.network. Output from \code{gemtc::mtc.network()}}
\item{dic}{dataframe. Containing the statistics 'Dbar', 'pD', 'DIC', and 'data points'}
\item{outcome}{character. The \code{outcome} from \code{configured_data}}
\item{outcome_measure}{character. The \code{outcome_measure}from \code{configured_data}}
\item{reference_treatment}{character. The \code{reference_treatment}from \code{configured_data}}
\item{effects}{character. The \code{effects} from \code{configured_data}}
\item{seed}{numeric. The \code{seed} from \code{configured_data}}
}
\description{
Fit a Bayesian model using \CRANpkg{gemtc}
}
