% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model_f.R
\name{bayes_model}
\alias{bayes_model}
\title{Run Bayesian models}
\usage{
bayes_model(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  reference_treatment,
  seed,
  async = FALSE
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{treatment_df}{dataframe. Treatments}

\item{outcome}{character. Outcome type for the dataset. Either \code{Binary} or
\code{Continuous}.}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{seed}{numeric. Seed used to fit the model}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{mtcResults}{mtc.result. Output from \code{gemtc::mtc.run()}}
\item{mtcRelEffects}{mtc.result. Output from \code{gemtc::relative.effect()}}
\item{rel_eff_tbl}{mtc.relative.effect.table. Output from \code{gemtc::relative.effect.table()}}
\item{sumresults}{summary.mtc.result. Output from \code{summary(mtcRelEffects)}}
\item{mtcNetwork}{mtc.network. Output from \code{gemtc::mtc.network()}}
\item{dic}{dataframe. Containing the statistics 'Dbar', 'pD', 'DIC', and 'data points'}
\item{outcome_measure}{character. The input \code{outcome_measure}}
\item{model_type}{dataframe. The input \code{model_type}}
}
\description{
Run Bayesian models
}
