% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model_f.R
\name{bayes_model}
\alias{bayes_model}
\title{Run Bayesian models}
\usage{
bayes_model(
  data,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  reference_treatment,
  logger = NULL
)
}
\arguments{
\item{data}{dataframe. Input data set.}

\item{treatment_df}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}

\item{outcome}{"Continuous" or "Binary".}

\item{outcome_measure}{"MD", "OR" or "RR".}

\item{model_type}{"fixed" or "random".}

\item{reference_treatment}{Reference treatment}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
List:
 - 'mtcResults' = Output from gemtc::mtc.run
 - 'mtcRelEffects' = Output from gemtc::relative.effect
 - 'rel_eff_tbl = Output from gemtc::relative.effect.table
 - 'sumresults' = summary(mtcRelEffects)
 - 'a' = "fixed effect" or "random effect"
 - 'mtcNetwork' = Output from gemtc::mtc.network
 - 'dic' = Data frame containing the statistics 'Dbar', 'pD', 'DIC', and 'data points'
}
\description{
Run Bayesian models
}
