% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_model_f.R
\name{bayes_model}
\alias{bayes_model}
\title{Run Bayesian models}
\usage{
bayes_model(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  reference_treatment,
  async = FALSE
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{treatment_df}{dataframe containing the treatment ID ('Number') and the treatment name ('Label').}

\item{outcome}{character. The type of outcome being measured either \code{Continuous} or \code{Binary}}

\item{outcome_measure}{character. The analysis outcome measure. Either \code{MD}, \code{OR} or \code{RR}}

\item{model_type}{character. The type of model. Either \code{random} or \code{fixed}}

\item{reference_treatment}{character. Reference treatment}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List containing:
\item{mtcResults}{mtc.result. Output from \code{gemtc::mtc.run()}}
\item{mtcRelEffects}{mtc.result. Output from \code{gemtc::relative.effect()}}
\item{rel_eff_tbl}{mtc.relative.effect.table. Output from \code{gemtc::relative.effect.table()}}
\item{sumresults}{summary.mtc.result. Output from \code{summary(mtcRelEffects)}}
\item{mtcNetwork}{mtc.network. Output from \code{gemtc::mtc.network()}}
\item{dic}{dataframe. Containing the statistics 'Dbar', 'pD', 'DIC', and 'data points'}
\item{outcome_measure}{character. The input \code{outcome_measure}}
\item{model_type}{dataframe. The input \code{model_type}}
}
\description{
Run Bayesian models
}
