% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_nodesplit_f.R
\name{bayes_nodesplit}
\alias{bayes_nodesplit}
\title{Run the nodesplit model}
\usage{
bayes_nodesplit(
  data,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  async = FALSE
)
}
\arguments{
\item{data}{Data to analyse.}

\item{treatment_df}{Dataframe containing treatment names ("Label") and IDs ("Number").}

\item{outcome}{character. The type of outcome being measured either `Continuous` or `Binary`}

\item{outcome_measure}{character. The analysis outcome measure. Either `MD`, `OR` or `RR`}

\item{model_type}{character. The type of model. Either `random` or `fixed`}
}
\value{
The created nodesplit model
}
\description{
Run the nodesplit model
}
