% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_nodesplit_f.R
\name{bayes_nodesplit}
\alias{bayes_nodesplit}
\title{Run the nodesplit model}
\usage{
bayes_nodesplit(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  async = FALSE
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{treatment_df}{Dataframe containing treatment names ("Label") and IDs ("Number").}

\item{outcome}{character. The type of outcome being measured either \code{Continuous} or \code{Binary}}

\item{outcome_measure}{character. The analysis outcome measure. Either \code{MD}, \code{OR} or \code{RR}}

\item{model_type}{character. The type of model. Either \code{random} or \code{fixed}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
mtc.nodesplit object containing an mtc.result object for each node
}
\description{
Run the nodesplit model
}
