% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_nodesplit_f.R
\name{bayes_nodesplit}
\alias{bayes_nodesplit}
\title{Fit a nodesplitting model with \code{gemtc::mtc.nodesplit()}. This is not possible
for all networks and the function will return an error if the nodes cannot
be split.}
\usage{
bayes_nodesplit(configured_data, async = FALSE)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
\code{mtc.nodesplit} object containing an \code{mtc.result} object for each node
}
\description{
Fit a nodesplitting model with \code{gemtc::mtc.nodesplit()}. This is not possible
for all networks and the function will return an error if the nodes cannot
be split.
}
