% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_nodesplit_f.R
\name{bayes_nodesplit_plot}
\alias{bayes_nodesplit_plot}
\title{Produce a forest plot from nodesplitting results}
\usage{
bayes_nodesplit_plot(nodesplit, main_analysis = TRUE, logger = NULL)
}
\arguments{
\item{nodesplit}{\code{mtc.nodesplit} object produced by \code{bayes_nodesplit()}}

\item{main_analysis}{logical. Whether the analysis is the main or sensitivity analysis. Default \code{TRUE}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. svg document containing the plot
}
\description{
Produce a forest plot from nodesplitting results
}
