% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{bayes_ranking}
\alias{bayes_ranking}
\title{Function to create data regarding rank results - CRN}
\usage{
bayes_ranking(
  connected_data,
  treatment_df,
  model,
  ranking_option,
  logger = NULL
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by `setup_configure()` or `setup_exclude`}

\item{treatment_df}{Dataframe containing the treatment ID ('Number') and the treatment name ('Label').}

\item{model}{List. Output produced by bayes_model().}

\item{ranking_option}{"good" or "bad", referring to small outcome values.}
}
\value{
List of output created by rankdata().
}
\description{
Function to create data regarding rank results - CRN
}
