% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{bayes_ranking}
\alias{bayes_ranking}
\title{Function to create data regarding rank results - CRN}
\usage{
bayes_ranking(
  data,
  metaoutcome,
  treatment_list,
  bayesmodel,
  rankdir,
  cov_value = NA,
  excluded = c(),
  package = "gemtc"
)
}
\arguments{
\item{data}{Input data set.}

\item{metaoutcome}{"Continuous" or "Binary".}

\item{treatment_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}

\item{bayesmodel}{List of various model output created by baye().}

\item{rankdir}{"good" or "bad", referring to small outcome values.}

\item{excluded}{Vector of excluded studies for sensitivity analysis.}
}
\value{
List of output created by rankdata().
}
\description{
Function to create data regarding rank results - CRN
}
