% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{bayes_ranking}
\alias{bayes_ranking}
\alias{baseline_ranking}
\alias{covariate_ranking}
\title{Generate ranking data required to produce SUCRA plots from Bayesian models}
\usage{
bayes_ranking(
  connected_data,
  treatment_df,
  model,
  ranking_option,
  cov_value = NA,
  logger = NULL
)

baseline_ranking(...)

covariate_ranking(...)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{treatment_df}{dataframe. Treatments}

\item{model}{list. Output produced by \code{baseline_model()}, \code{bayes_model()} or \code{covariate_model()}.}

\item{ranking_option}{character. \code{good} if the treatment effect is desirable, else \code{bad}}

\item{cov_value}{numeric. Covariate value if a meta-regression. Default \code{NA}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
List of output created by \code{rankdata()}.
}
\description{
Generate ranking data required to produce SUCRA plots from Bayesian models
}
