% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{bayes_ranking}
\alias{bayes_ranking}
\alias{baseline_ranking}
\alias{covariate_ranking}
\title{Generate ranking data required to produce SUCRA plots from Bayesian models}
\usage{
bayes_ranking(model, configured_data, logger = NULL)

baseline_ranking(...)

covariate_ranking(...)
}
\arguments{
\item{model}{list. Output produced by \code{baseline_model()}, \code{bayes_model()} or \code{covariate_model()}.}

\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{...}{Parameters passed to \code{bayes_ranking()}}
}
\value{
List of output created by \code{rankdata()}
\item{SUCRA}{Dataframe of SUCRA data}
\item{Colour}{Dataframe of colours}
\item{Cumulative}{Dataframe of cumulative ranking probabilities}
\item{Probabilities}{Dataframe of ranking probabilities}
\item{BUGSnetData}{ output from \code{BUGSnet::data.prep()}}
}
\description{
Generate ranking data required to produce SUCRA plots from Bayesian models
}
