% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_results_f.R
\name{bayes_results}
\alias{bayes_results}
\alias{covariate_results}
\alias{baseline_results}
\title{Produce a summary of a Bayesian model}
\usage{
bayes_results(model, logger = NULL)

covariate_results(...)

baseline_results(...)
}
\arguments{
\item{model}{list. Output produced by \code{baseline_model()}, \code{bayes_model()} or \code{covariate_model()}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{...}{Parameters passed to \code{bayes_results()}}
}
\value{
HTML summary of the model
}
\description{
Produce a summary of a Bayesian model
}
