% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_results_f.R
\name{bayes_results}
\alias{bayes_results}
\alias{covariate_results}
\title{Produces a summary of a Bayesian model}
\usage{
bayes_results(model, logger = NULL)

covariate_results(...)
}
\arguments{
\item{model}{list. Output produced by \code{bayes_model()} or \code{covariate_model()}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
HTML summary of the model
}
\description{
Produces a summary of a Bayesian model
}
