% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_results_f.R
\name{bayes_results}
\alias{bayes_results}
\title{Produces a summary of the fitted Bayesian model}
\usage{
bayes_results(model, logger = NULL)
}
\arguments{
\item{model}{Various model output created by bayes_model().}
}
\value{
HTML summary of the model
}
\description{
Produces a summary of the fitted Bayesian model
}
