% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_results_f.R
\name{bayes_results}
\alias{bayes_results}
\title{Produces a summary of the fitted Bayesian model}
\usage{
bayes_results(model, logger = NULL)
}
\arguments{
\item{model}{Various model output created by bayes_model()}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
HTML summary of the model
}
\description{
Produces a summary of the fitted Bayesian model
}
