% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{common_params}
\alias{common_params}
\title{Common parameters}
\arguments{
\item{treatments}{dataframe. Treatments}

\item{outcome}{character. Outcome type for the dataset. Either \code{binary} or
\code{continuous}.}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{continuous}}

\item{ranking_option}{character. \code{good} if the treatment effect is desirable, else \code{bad}}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{effects}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{seed}{numeric. Seed used to fit the models.}

\item{xmin}{numeric. Minimum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{xmax}{numeric. Maximum x-axis value. Default \code{NULL} in which case it is calculated internally}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\description{
Common parameters
}
\keyword{internal}
