% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{contrastform.df}
\alias{contrastform.df}
\title{Puts data in contrast form using netmeta::pairwise().}
\usage{
contrastform.df(data, outcome, CONBI)
}
\arguments{
\item{data}{Input dataset.}

\item{outcome}{"MD", "SMD", "OR", "RR", or "RD".}

\item{CONBI}{"Continuous" or "Binary".}
}
\value{
Input data in contrast form.
}
\description{
Puts data in contrast form using netmeta::pairwise().
}
