% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{covariate_model}
\alias{covariate_model}
\title{covariate_model}
\usage{
covariate_model(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  covariate,
  cov_friendly,
  cov_value,
  model_type,
  regressor_type,
  reference_treatment,
  mtc_run_output = NULL,
  aync = FALSE
)
}
\arguments{
\item{connected_data}{dataframe. Uploaded data post-processing}

\item{treatment_df}{dataframe. Containing the treatment ID ('Number') and the treatment name ('Label').}

\item{outcome}{character. The type of outcome being measured either \code{Continuous} or \code{Binary}}

\item{outcome_measure}{character. The analysis outcome measure. Either \code{MD}, \code{OR} or \code{RR}}

\item{covariate}{character. Chosen covariate name as per uploaded data}

\item{cov_friendly}{character. Friendly name of chosen covariate}

\item{cov_value}{numeric.}

\item{model_type}{character. The type of model. Either \code{random} or \code{fixed}}

\item{regressor_type}{character. Type of regression coefficient, either \code{shared}, \code{unrelated}, or \code{exchangeable}}

\item{reference_treatment}{character. Reference treatment}

\item{covariate_model_output}{list. The output of the function. Default NULL.
When supplied, only the output is recalculated for a given covariate value,
rather than refitting the model.}
}
\description{
Does x
}
