% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{covariate_model}
\alias{covariate_model}
\title{covariate_model}
\usage{
covariate_model(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  covariate,
  cov_friendly,
  cov_value,
  model_type,
  regressor_type,
  reference_treatment,
  covariate_model_output = NULL,
  async = FALSE
)
}
\arguments{
\item{connected_data}{dataframe. Input data set created by \code{setup_configure()} or \code{setup_exclude}}

\item{treatment_df}{dataframe. Treatments}

\item{outcome}{character. Outcome type for the dataset. Either \code{Binary} or
\code{Continuous}.}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{covariate}{character. Chosen covariate name as per uploaded data}

\item{cov_friendly}{character. Friendly name of chosen covariate}

\item{cov_value}{numeric. Covariate value at which to produce predictions}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{regressor_type}{character. Type of regression coefficient, either \code{shared}, \code{unrelated}, or \code{exchangeable}}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{covariate_model_output}{list. The output of the function. Default NULL.
When supplied, only the output is recalculated for a given covariate value,
rather than refitting the model.}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
List of gemtc related output:
\item{mtc_run_output#'  \item{output from \code{gemtc::mtc.run()}
\item{mtcResults}{model object itself carried through (needed to match existing code)}
\item{mtcRelEffects}{data relating to presenting relative effects}
\item{rel_eff_tbl}{table of relative effects for each comparison}
\item{covariate_value}{The covariate value originally passed into this function}
\item{reference_name}{The name of the reference treatment}
\item{comparator_names}{Vector containing the names of the comparators}
\item{a}{text output stating whether fixed or random effects}
\item{sumresults}{summary output of relative effects}
\item{dic}{data frame of model fit statistics}
\item{cov_value_sentence}{text output stating the value for which the covariate has been set to for producing output}
\item{slopes}{named list of slopes for the regression equations (unstandardised - equal to one 'increment')}
\item{intercepts}{named list of intercepts for the regression equations at cov_value}
\item{outcome}{The outcome type for the analysis eg. "MD" or "OR"}
\item{mtcNetwork}{The network object from GEMTC}
\item{model}{The type of linear model, either "fixed" or "random"}
\item{covariate_min}{Vector of minimum covariate values directly contributing to the regression}
\item{covariate_max}{Vector of maximum covariate values directly contributing to the regression}
}
\description{
Does x
}
