% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{covariate_model}
\alias{covariate_model}
\title{covariate_model}
\usage{
covariate_model(
  connected_data,
  treatment_df,
  outcome,
  outcome_measure,
  covariate,
  cov_friendly,
  cov_value,
  model_type,
  regressor_type,
  reference_treatment,
  mtc_run_output = NULL,
  logger = NULL
)
}
\arguments{
\item{connected_data}{dataframe. Uploaded data post-processing}

\item{treatment_df}{dataframe. Containing the treatment ID ('Number') and the treatment name ('Label').}

\item{outcome}{character. The type of outcome being measured either `Continuous` or `Binary`}

\item{outcome_measure}{character. The analysis outcome measure. Either `MD`, `OR` or `RR`}

\item{covariate}{character. Chosen covariate name as per uploaded data}

\item{cov_friendly}{character. Friendly name of chosen covariate}

\item{cov_value}{numeric.}

\item{model_type}{character. The type of model. Either `random` or `fixed`}

\item{regressor_type}{character. Type of regression coefficient, either `shared`, `unrelated`, or `exchangeable`}

\item{reference_treatment}{character. Reference treatment}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
