% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{covariate_regression}
\alias{covariate_regression}
\title{covariate_regression}
\usage{
covariate_regression(
  model_output,
  connected_data,
  covariate_column,
  covariate_name,
  treatment_df,
  outcome,
  outcome_measure,
  model_type,
  absolute_or_percentage,
  weight_or_contribution,
  treatment_or_covariate_effect,
  logger = NULL
)
}
\arguments{
\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}

\item{x}{x}
}
\description{
Does x
}
\examples{
{
y <- covariate_regression(1)
}
}
