% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_regression_f.R
\name{covariate_regression}
\alias{covariate_regression}
\title{covariate_regression}
\usage{
covariate_regression(x, logger = NULL)
}
\arguments{
\item{x}{x}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\description{
Does x
}
\examples{
{
y <- covariate_regression(1)  
}
}
