% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_summary_f.R
\name{covariate_summary}
\alias{covariate_summary}
\title{Produce a plot summarising the covariate value for each study arm}
\usage{
covariate_summary(configured_data, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Produce a plot summarising the covariate value for each study arm
}
