% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CreateTauSentence}
\alias{CreateTauSentence}
\title{Create text with the point estimate and 95\\% CrI of between-trial SD of treatment effects.}
\usage{
CreateTauSentence(model)
}
\arguments{
\item{model}{Output created by \code{bayes_model()}}
}
\value{
Text with the point estimate and 95\\% CrI of between-trial SD of treatment effects (all 0 if fixed effects)
}
\description{
Create text with the point estimate and 95\\% CrI of between-trial SD of treatment effects.
}
