% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{crop_svg}
\alias{crop_svg}
\title{Crops an svg object, by rendering it, locating the non-white pixels and
editing the svg viewBox property. The width and height are also returned
to facilitate rendering to other formats.}
\usage{
crop_svg(input_svg, margin = 10)
}
\arguments{
\item{input_svg}{xml_document. Output from `svglite::xmlSVG()`}

\item{margin}{numeric. The margin in pixels to leave around the edge
of the plot content. Defaults to 10.}
}
\value{
List containing:
 \item{svg}{character. The cropped svg}
 \item{width}{numeric. The width of the viewBox in pixels}
 \item{height}{numeric. The height of the viewBox in pixels}
}
\description{
Crops an svg object, by rendering it, locating the non-white pixels and
editing the svg viewBox property. The width and height are also returned
to facilitate rendering to other formats.
}
