% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_configure_f.R
\name{dataform.df}
\alias{dataform.df}
\title{Converts wide data to long, adds the variable 'se' for a continuous outcomes, and does some formatting.}
\usage{
dataform.df(newData1, treat_list, CONBI)
}
\arguments{
\item{newData1}{Input dataset.}

\item{treat_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}

\item{CONBI}{"continuous" or "binary".}
}
\value{
Input data set in long format with the variable 'se' for a continuous outcome.
}
\description{
Converts wide data to long, adds the variable 'se' for a continuous outcomes, and does some formatting.
}
