% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.AddOutcomeText}
\alias{.AddOutcomeText}
\title{Add the effects size and confidence interval text for a single comparison in a single study.}
\usage{
.AddOutcomeText(
  pairwise,
  x_position,
  y_position,
  studlab,
  treatment1,
  treatment2
)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}

\item{x_position}{x co-ordinate for the start of the text.}

\item{y_position}{y co-ordinate for the effect size and confidence interval.}

\item{studlab}{Study.}

\item{treatment1}{First treatment in the comparison.}

\item{treatment2}{Second treatment in the comparison.}
}
\description{
Add the effects size and confidence interval text for a single comparison in a single study.
}
