% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{.AddTreatmentEffectBlock}
\alias{.AddTreatmentEffectBlock}
\title{Add the treatment effect lines, confidence intervals, studies, and comparison header, for a single comparison.}
\usage{
.AddTreatmentEffectBlock(
  pairwise,
  label_x_position,
  outcome_x_position,
  rob_x_positions,
  y_header_position,
  y_positions,
  treatment1,
  treatment2,
  rob_variables,
  palette,
  interactive
)
}
\arguments{
\item{pairwise}{Output from meta::pairwise().}

\item{label_x_position}{x co-ordinate for the start of the labels.}

\item{outcome_x_position}{x co-ordinate for the start of the text.}

\item{rob_x_positions}{x co-ordinates for the circles.}

\item{y_header_position}{y co-ordinate for the header.}

\item{y_positions}{Vector of y co-ordinates for the study labels and confidence intervals.}

\item{treatment1}{First treatment in the comparison.}

\item{treatment2}{Second treatment in the comparison.}

\item{rob_variables}{Vector of RoB and indirectness variable names from 'pairwise'.}

\item{palette}{Vector of colours to use}

\item{interactive}{Whether the plot will be interactive}
}
\description{
Add the treatment effect lines, confidence intervals, studies, and comparison header, for a single comparison.
}
