% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{.ValidateMissingColumns}
\alias{.ValidateMissingColumns}
\title{Validate that there are no missing columns in the data.}
\usage{
.ValidateMissingColumns(data, required_columns, outcome_type)
}
\arguments{
\item{data}{Data frame to validate.}

\item{required_columns}{Data frame containing data definitions.}

\item{outcome_type}{Outcome type selected for the data. Either "Binary" or "Continuous".}
}
\value{
Validation result in the form of a list:
- "valid" = TRUE or FALSE defining whether data is valid
- "message" = String describing any issues causing the data to be invalid
}
\description{
Validate that there are no missing columns in the data.
}
