% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{FindCovariateNames}
\alias{FindCovariateNames}
\title{Find the names of all columns which contain a covariate.}
\usage{
FindCovariateNames(df)
}
\arguments{
\item{df}{Data frame in which to find covariate columns.}
}
\value{
Names of all covariate columns
}
\description{
Find the names of all columns which contain a covariate.
}
