% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{forest_annotation}
\alias{forest_annotation}
\title{Creates the text to be displayed underneath the forest plots, with between-study SD, number of studies and number of treatments.}
\usage{
forest_annotation(freq, model_type, outcome_measure)
}
\arguments{
\item{model_type}{"fixed" or "random".}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}

\item{tau}{Between study standard deviation.}

\item{k}{Number of studies.}

\item{n}{Number of treatments.}
}
\value{
Text as described above.
}
\description{
Creates the text to be displayed underneath the forest plots, with between-study SD, number of studies and number of treatments.
}
