% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_analysis.R
\name{freq.df}
\alias{freq.df}
\title{Frequentist NMA.}
\usage{
freq.df(effects, outcome_measure, dataf, reference_treatment)
}
\arguments{
\item{effects}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{continuous}}

\item{dataf}{Data in contrast form with treatment labels, typically output from labelmatching.df().}

\item{reference_treatment}{character. The reference treatment of the dataset}
}
\value{
NMA results from netmeta::netmeta().
}
\description{
Frequentist NMA.
}
