% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_compare_f.R
\name{freq_compare}
\alias{freq_compare}
\title{Produce a comparison table of treatments using
\code{netmeta::netleague()}.}
\usage{
freq_compare(configured_data, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
Dataframe of comparisons with one row and one column
per treatment
}
\description{
Produce a comparison table of treatments using
\code{netmeta::netleague()}.
}
