% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_compare_f.R
\name{freq_compare}
\alias{freq_compare}
\title{Produce comparison tables}
\usage{
freq_compare(freq, model_type, ranking_option, logger = NULL)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}

\item{model_type}{Character. `fixed` or `random`.}

\item{ranking_option}{Character. `good` or `bad`, referring to smaller outcome values.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
Ranking table created by netmeta::netleague().
}
\description{
Produce comparison tables
}
