% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{freq_forest}
\alias{freq_forest}
\title{Produce a frequentist forest plot and annotation using \code{meta::forest()}}
\usage{
freq_forest(configured_data, xmin, xmax, title, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{xmin}{numeric. Minimum x-axis limit.}

\item{xmax}{numeric. Maximum x-axis limit.}

\item{title}{character. Title for the plot.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Produce a frequentist forest plot and annotation using \code{meta::forest()}
}
