% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{freq_forest}
\alias{freq_forest}
\title{Produce a forest plot and annotation}
\usage{
freq_forest(
  freq,
  reference_treatment,
  model_type,
  outcome_measure,
  xmin,
  xmax,
  title,
  logger = NULL
)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{xmin}{numeric. Minimum x-axis limit.}

\item{xmax}{numeric. Maximum x-axis limit.}

\item{title}{character. Title for the plot.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
List containing:
\item{svg}{character. SVG code to produce the plot}
\item{height}{numeric. Plot height in pixels}
\item{width}{numeric. Plot width in pixels}
}
\description{
Produce a forest plot and annotation
}
