% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_forest_f.R
\name{freq_forest}
\alias{freq_forest}
\title{Produce a forest plot and annotation}
\usage{
freq_forest(
  freq,
  reference_treatment,
  model_type,
  outcome_measure,
  xmin,
  xmax,
  logger = NULL
)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{model_type}{character. Type of model to fit, either `random` or `fixed`}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
`OR`, `RR` or `RD` when `outcome` is `Binary` or `MD` or `SMD` when
`outcome` is `Continuous`}

\item{xmin}{numeric. Minimum x-axis limit.}

\item{xmax}{numeric. Maximum x-axis limit.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
List containing:
 \item{plot}{function. Forest plot created using `meta::forest()`}
 \item{annotation}{character. Annotation to add to the plot}
 \item{height}{numeric. Plot height in inches}
 \item{width}{numeric. Plot width in inches}
}
\description{
Produce a forest plot and annotation
}
