% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{freq_summary}
\alias{freq_summary}
\title{Produce a summary forest plot matrix with treatments ranked by SUCRA score,
determined by \code{netmeta::rankogram()}. This function can only be used when
\code{configured_data} contains between 3 and 10 treatments.}
\usage{
freq_summary(configured_data, plot_title = "", logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{plot_title}{character. Title of the plot. Default is no title.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Produce a summary forest plot matrix with treatments ranked by SUCRA score,
determined by \code{netmeta::rankogram()}. This function can only be used when
\code{configured_data} contains between 3 and 10 treatments.
}
