% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{freq_summary}
\alias{freq_summary}
\title{Create a summary forest plot matrix.}
\usage{
freq_summary(
  freq,
  treatment_df,
  plot_title,
  outcome_measure,
  ranking_option,
  model_type,
  logger = NULL
)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{treatment_df}{dataframe. containing the names and IDs of all treatments}

\item{plot_title}{character. Title of the plot}

\item{outcome_measure}{character. Type of outcome being plotted}

\item{ranking_option}{character. `good` if the treatment effect is desirable, else `bad`}

\item{model_type}{character. `random` or `fixed` effects model being plotted}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
Summary forest matrix plot
}
\description{
Create a summary forest plot matrix.
}
