% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_summary_f.R
\name{freq_summary}
\alias{freq_summary}
\title{Create a summary forest plot matrix.}
\usage{
freq_summary(
  freq,
  treatment_df,
  plot_title,
  outcome_measure,
  ranking_option,
  model_type,
  seed,
  logger = NULL
)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{treatment_df}{dataframe. Treatments}

\item{plot_title}{character. Title of the plot}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{ranking_option}{character. \code{good} if the treatment effect is desirable, else \code{bad}}

\item{model_type}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{seed}{numeric. Seed used to fit the model}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Create a summary forest plot matrix.
}
