% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_analysis.R
\name{frequentist}
\alias{frequentist}
\title{Converts long data to wide if necessary, then applies freq_wrap().}
\usage{
frequentist(
  data,
  metaoutcome,
  treatment_list,
  outcome_measure,
  modelranfix,
  reference
)
}
\arguments{
\item{data}{Input dataset.}

\item{metaoutcome}{"Continuous" or "Binary".}

\item{treatment_list}{Data frame containing the treatment ID ('Number') and the treatment name ('Label').}

\item{outcome_measure}{"MD", "SMD", "OR", "RR", or "RD".}

\item{modelranfix}{"fixed" or "random".}

\item{reference}{Name of the reference treatment.}
}
\value{
See output from freq_wrap().
}
\description{
Converts long data to wide if necessary, then applies freq_wrap().
}
