% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_analysis.R
\name{frequentist}
\alias{frequentist}
\title{Converts long data to wide if necessary, then applies freq_wrap().}
\usage{
frequentist(
  non_covariate_data,
  outcome,
  treatments,
  outcome_measure,
  effects,
  reference_treatment
)
}
\arguments{
\item{non_covariate_data}{Input dataset with any covariates removed.}

\item{outcome}{character. Outcome type for the dataset. Either \code{binary} or
\code{continuous}.}

\item{treatments}{dataframe. Treatments}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{continuous}}

\item{effects}{character. Type of model to fit, either \code{random} or \code{fixed}}

\item{reference_treatment}{character. The reference treatment of the dataset}
}
\value{
List containing:
\item{net1}{list. NMA results from netmeta::netmeta()}
\item{lstx}{character. Vector of treatment labels}
\item{ntx}{numeric. Number of treatments}
\item{d0}{dataframe.Data in contrast form}
\item{d1}{dataframe.Same as 'd0' but with treatment labels}
}
\description{
Converts long data to wide if necessary, then applies freq_wrap().
}
