% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{gelman_plots}
\alias{gelman_plots}
\title{Creates Gelman plots for a gemtc or bnma model.}
\usage{
gelman_plots(gelman_data, parameters)
}
\arguments{
\item{parameters}{Vector of parameters mentioned in the previous argument.}

\item{gelman_plots}{List of outputs from \code{coda::gelman.plot(model$samples[, parm])}, where parm is a parameter from the model.}
}
\value{
List of ggplot Gelman plots
}
\description{
Creates Gelman plots for a gemtc or bnma model.
}
