% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{gelman_plots}
\alias{gelman_plots}
\title{Creates Gelman plots for a gemtc or bnma model.}
\usage{
gelman_plots(gelman_data, parameters)
}
\arguments{
\item{gelman_data}{List of outputs from \code{coda:::gelman.preplot}}

\item{parameters}{Vector of parameters mentioned in the previous argument.}
}
\value{
List of ggplot Gelman plots
}
\description{
Creates Gelman plots for a gemtc or bnma model.
}
