% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_load_f.R
\name{GetFriendlyCovariateName}
\alias{GetFriendlyCovariateName}
\title{Convert a covariate column name to a display name.}
\usage{
GetFriendlyCovariateName(column_name)
}
\arguments{
\item{column_name}{Covariate column name to convert}
}
\value{
Friendly covariate name
}
\description{
Convert a covariate column name to a display name.
}
