% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_model_f.R
\name{GetReferenceOutcome}
\alias{GetReferenceOutcome}
\title{Get the outcome in the reference arm.}
\usage{
GetReferenceOutcome(
  connected_data,
  treatment_df,
  outcome,
  observed,
  model = NULL
)
}
\arguments{
\item{connected_data}{Data in long format.}

\item{treatment_df}{Data frame containing treatment IDs and names in columns named 'Number' and 'Label' respectively.}

\item{outcome}{"Binary" or "Continuous".}

\item{observed}{"Observed" or "Imputed". See @return.}

\item{model}{Model created by bnma::network.run(). Only required when @param observed == "Imputed". Defaults to NULL.}
}
\value{
Vector of reference arm outcomes, named by study.
If a study contains the reference treatment then the value returned is the observed outcome in the reference arm.
If a study does not contain the reference treatment then the value returned is:
- NA if @param observed == "Observed";
- The median of the study-specific intercept parameter if @param observed == "Imputed".
}
\description{
Get the outcome in the reference arm.
}
