% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_analysis.R
\name{labelmatching.df}
\alias{labelmatching.df}
\title{Adds treatment labels to contrast data.}
\usage{
labelmatching.df(d1, ntx, treatments)
}
\arguments{
\item{d1}{Data in contrast form, typically created by contrastform.df().}

\item{ntx}{= Number of treatments.}

\item{treatments}{dataframe. Treatments}
}
\value{
Input data with the columns treat1 and treat2 now labelled.
}
\description{
Adds treatment labels to contrast data.
}
