% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{LitmusRankOGram}
\alias{LitmusRankOGram}
\title{Litmus Rank-O-Gram}
\usage{
LitmusRankOGram(
  CumData,
  SUCRAData,
  ColourData,
  colourblind = FALSE,
  regression_text = ""
)
}
\arguments{
\item{CumData}{Cumulative ranking probabilities, created by rankdata().}

\item{SUCRAData}{SUCRA data, created by rankdata().}

\item{ColourData}{Colour data, created by rankdata().}

\item{colourblind}{TRUE for colourblind friendly colours (default = FALSE).}

\item{regression_text}{Text to show for regression (default = "").}
}
\value{
Litmus rank-o-gram.
}
\description{
Litmus Rank-O-Gram
}
