% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{LitmusRankOGram}
\alias{LitmusRankOGram}
\title{Litmus Rank-O-Gram}
\usage{
LitmusRankOGram(ranking_data, colourblind = FALSE, regression_text = "")
}
\arguments{
\item{ranking_data}{list created by bayes_ranking().}

\item{colourblind}{TRUE for colourblind friendly colours (default = FALSE).}

\item{regression_text}{Text to show for regression (default = "").}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Litmus Rank-O-Gram
}
