% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_netconnect}
\alias{make_netconnect}
\title{1c Creates network connectivity info displayed under network plots}
\usage{
make_netconnect(freq)
}
\arguments{
\item{freq}{List of NMA results created by freq_wrap().}
}
\value{
Network connectivity created by netmeta::netconnection().
}
\description{
1c Creates network connectivity info displayed under network plots
}
