% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_summary_f.R
\name{PasteCaptionText}
\alias{PasteCaptionText}
\title{Paste the caption text together}
\usage{
PasteCaptionText(caption_setting, error_bar_text = NULL)
}
\arguments{
\item{caption_setting}{Text string to describe type of plot. Can be "baseline risk" or "covariate"}

\item{error_bar_text}{Text string to explain the error bar (optional)}
}
\value{
Text string to be used for caption
}
\description{
Paste the caption text together
}
