% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{RadialSUCRA}
\alias{RadialSUCRA}
\title{Radial SUCRA Plot}
\usage{
RadialSUCRA(
  ranking_data,
  original = TRUE,
  colourblind = FALSE,
  regression_text = ""
)
}
\arguments{
\item{ranking_data}{list created by bayes_ranking().}

\item{original}{logical. Whether to display the original or simplified version (default = \code{TRUE})}

\item{colourblind}{logical. \code{TRUE} for colourblind friendly colours (default = \code{FALSE}).}

\item{regression_text}{Text to show for regression (default = "").}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Radial SUCRA Plot
}
