% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{RadialSUCRA}
\alias{RadialSUCRA}
\title{Radial SUCRA Plot}
\usage{
RadialSUCRA(ranking_data, colourblind = FALSE, regression_text = "")
}
\arguments{
\item{ranking_data}{list created by bayes_ranking().}

\item{colourblind}{TRUE for colourblind friendly colours (default = FALSE).}

\item{regression_text}{Text to show for regression (default = "").}
}
\value{
Radial SUCRA plot.
}
\description{
Radial SUCRA Plot
}
