% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{RadialSUCRA}
\alias{RadialSUCRA}
\title{Radial SUCRA Plot}
\usage{
RadialSUCRA(
  SUCRAData,
  ColourData,
  BUGSnetData,
  colourblind = FALSE,
  regression_text = ""
)
}
\arguments{
\item{SUCRAData}{SUCRA data, created by rankdata().}

\item{ColourData}{Colour data, created by rankdata().}

\item{BUGSnetData}{Output created by BUGSnet functions in rankdata().}

\item{colourblind}{TRUE for colourblind-friendly colours (default = FALSE).}

\item{regression_text}{Text to show for regression (default = "").}

\item{Type}{"Original" or "Alternative".}
}
\value{
Radial SUCRA plot.
Creates the network part of the radical SUCRA plot, excluding the nodes.

ggplot object.
Creates the nodes for the network part of the radial SUCRA plot.

ggplot object.
}
\description{
Radial SUCRA Plot
}
