% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{RadialSUCRA}
\alias{RadialSUCRA}
\title{Radial SUCRA Plot}
\usage{
RadialSUCRA(
  ranking_data,
  original = TRUE,
  colourblind = FALSE,
  regression_text = ""
)
}
\arguments{
\item{ranking_data}{list created by bayes_ranking().}

\item{colourblind}{TRUE for colourblind friendly colours (default = FALSE).}

\item{regression_text}{Text to show for regression (default = "").}
}
\value{
List containing:
\item{svg}{character. SVG code to produce the plot}
\item{height}{numeric. Plot height in pixels}
\item{width}{numeric. Plot width in pixels}
}
\description{
Radial SUCRA Plot
}
