% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ranking_f.R
\name{rankdata}
\alias{rankdata}
\title{Get SUCRA data.}
\usage{
rankdata(NMAdata, rankdirection, longdata, cov_value = NA, package = "gemtc")
}
\arguments{
\item{NMAdata}{Output from 'baye' function or from bnma::network.run.}

\item{rankdirection}{"good" or "bad" (referring to smaller outcome values).}

\item{longdata}{Output from 'dataform.df' function. This should be the same dataset that was passed as the 'data' argument to baye(), which resulted in @param NMAdata.
(TM: Suggested improvement: baye() should output its 'data' argument, then @param longdata becomes superfluous, and there is no possibility of a mismatch between @param NMAdata and @param longdata.)}

\item{cov_value}{covariate value if a meta-regression}

\item{package}{"gemtc" or "bnma", defaults to "gemtc".}
}
\value{
List:
- 'SUCRA' = Data frame of SUCRA data.
    - 'Treatment'
    - 'SUCRA' = Sucra percentages.
    - 'N' = Total number of patients in 'Treatment' arms (summed over all studies).
    - 'SizeO' = Size of points (relative to 'N') in original SUCRA plot.
    - 'SizeA' = Size of points (relative to 'N') in alternative SUCRA plot.
- 'Colour' = Data frame of colours.
    - 'SUCRA' = Possible SUCRA values.
    - 'colour' = Colour values corresponding to 'SUCRA'.
- 'Cumulative' = Data frame of cumulative ranking probabilities, in long format.
    - 'Treatment'
    - 'Rank'
    - 'Cumulative_Probability'
    - 'SUCRA'
- 'Probabilities' = Data frame of ranking probabilities.
    - 'Treatment'
    - 'Rank 1' = Probability 'Treatment' is ranked first.
    - ...
    - 'Rank n_t' = Probability 'Treatment' is ranked last (n_t = number of treatments).
- 'BUGSnetData' = Output from BUGSnet::data.prep with arguments from @param longdata.
}
\description{
Get SUCRA data.
}
