% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_generic.R, R/setup_define_f.R
\name{RankingOrder}
\alias{RankingOrder}
\title{Returns the default ranking order, for the example datasets.}
\usage{
RankingOrder(outcome, data)

RankingOrder(outcome, data)
}
\arguments{
\item{outcome}{"Binary" or "Continuous".}

\item{data}{Input dataset.}
}
\value{
"good" or "bad".

"good" or "bad".
}
\description{
Returns the default ranking order, for the example datasets.

Returns the default ranking order, for the example datasets.
}
