% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_docs.R
\name{return-svg}
\alias{return-svg}
\title{Return svg}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Return svg
}
\keyword{internal}
