% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scat_plot}
\alias{scat_plot}
\title{UME scatter plot}
\usage{
scat_plot(model)
}
\arguments{
\item{model}{Various model output created by baye().}
}
\value{
UME scatter plot created by umeplot.df().
}
\description{
UME scatter plot
}
