% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_exclude_f.R
\name{setup_exclude}
\alias{setup_exclude}
\title{Takes the configured data, removes any excluded studies and returns
subsets of the data to be passed to other functions.}
\usage{
setup_exclude(configured_data, exclusions, async = FALSE)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{exclusions}{character. Vector of study names to exclude.}

\item{async}{Whether or not the function is being used asynchronously. Default \code{FALSE}}
}
\value{
\code{configured_data} containing:
\item{treatments}{dataframe. Treatment names and IDs}
\item{reference_treatment}{character. The selected reference treatment}
\item{connected_data}{dataframe. A subset of the data containing only connected studies}
\item{covariate}{A list containing these items if covariate data exists or
else empty:}
\itemize{
\item \code{cross}: Crosses
\item \code{circle_open}: Open circles
\item \code{none}: No symbols in which case only the plot of direct evidence is
}
\item{bugsnet}{dataframe. Processed data for bugsnet analyses created by \code{bugsnetdata}}
\item{freq}{list. Processed data for frequentist analyses created by \code{frequentist()}}
\item{outcome}{character. Whether the data is \code{binary} or \code{continuous}}
\item{outcome_measure}{character. Outcome measure of the dataset.}
\item{effects}{character. Whether the models are \code{fixed} or \code{random} effects}
\item{ranking_option}{character. Whether higher values in the data are \code{good} or \code{bad}}
\item{seed}{numeric. A seed value to be passed to models}
}
\description{
Takes the configured data, removes any excluded studies and returns
subsets of the data to be passed to other functions.
}
