% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_exclude_f.R
\name{setup_exclude}
\alias{setup_exclude}
\title{Takes the uploaded data, removes any excluded studies and returns
subsets of the data in formats for bugsnet and frequentist analyses}
\usage{
setup_exclude(
  non_covariate_data,
  treatment_df,
  reference_treatment,
  outcome,
  outcome_measure,
  model_type,
  exclusions,
  logger = NULL
)
}
\arguments{
\item{non_covariate_data}{dataframe. Data processed by setup_define()}

\item{treatment_df}{dataframe. Treatments and their IDs}

\item{reference_treatment}{character. The reference treatment of the dataset}

\item{outcome}{character. Outcome type for the dataset. Either `Binary` or
`Continuous`.}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
`OR`, `RR` or `RD` when `outcome` is `Binary` or `MD` or `SMD` when
`outcome` is `Continuous`}

\item{model_type}{character. Type of model to fit, either `random` or `fixed`}

\item{exclusions}{character. Vector of study names to exclude.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
List containing:
 \item{bugsnet_sub}{dataframe. Processed data for bugsnet analyses created by `bugsnetdata()`}
 \item{freq_sub}{list. Processed data for frequentist analyses created by `frequentist()`}
 \item{reference_treatment_sub}{character. Reference treatment for sensitivity analyses}
 \item{subsetted_data}{dataframe. Processed data}
 \item{subsetted_treatment_df}{dataframe. Treatments in processed data}
}
\description{
Takes the uploaded data, removes any excluded studies and returns
subsets of the data in formats for bugsnet and frequentist analyses
}
