% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{setup_load}
\alias{setup_load}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
setup_load(data_path = NULL, outcome_type, logger = NULL)
}
\arguments{
\item{data_path}{character. Path to the file to be loaded or if `NULL` load the default data}

\item{outcome_type}{character. Outcome type selected for the data. Either `Binary` or `Continuous`.}
}
\value{
List containing:
 \item{valid}{logical. Whether the data is valid}
 \item{uploaded}{logical. Whether the data is uploaded}
 \item{data}{Dataframe. The data that was uploaded, `NULL` if none was uploaded}
 \item{treatments}{Vector of all treatment names either in the data if it is valid or in the default data if not}
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
