% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_upgrade_f.R
\name{setup_upgrade}
\alias{setup_upgrade}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
setup_upgrade(data_path, treatments, logger = NULL)
}
\arguments{
\item{data_path}{character. path to the file to be upgraded}

\item{treatments}{character vector. List of the treatments in the data}
}
\value{
Dataframe containing the upgraded data
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
