% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_upgrade_f.R
\name{setup_upgrade}
\alias{setup_upgrade}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
setup_upgrade(data_path, treatments, logger = NULL)
}
\arguments{
\item{data_path}{character. Path to the file to be upgraded}

\item{treatments}{character. The treatments in the data seperated by commas.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
Dataframe containing the upgraded data
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
