% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregression_plot_f.R
\name{SetupRegressionPlotColours}
\alias{SetupRegressionPlotColours}
\title{Setup the main components of the plot panel.}
\usage{
SetupRegressionPlotColours(
  plot,
  comparators,
  include_ghosts,
  include_credible,
  credible_opacity = 1
)
}
\arguments{
\item{plot}{The ggplot2 object}

\item{comparators}{Vector of names of comparison treatments to plot.}

\item{include_ghosts}{TRUE if all other comparator studies should be plotted in grey in the background of the plot.}

\item{include_credible}{TRUE if all other comparator studies should be plotted in grey in the background of the plot.}

\item{credible_opacity}{The opacity of the credible regions. Can be any value between 0 and 1, inclusive. Defaults to 1.}
}
\value{
Created ggplot2 object.
}
\description{
Setup the main components of the plot panel.
}
