% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_char_f.R
\name{summary_char}
\alias{summary_char}
\title{Takes bugsnet data and produces a summary dataframe of the studies}
\usage{
summary_char(bugsnet_data, outcome, logger = NULL)
}
\arguments{
\item{bugsnet_data}{dataframe. Dataframe produced by \code{bugsnetdata()}}

\item{outcome}{character. Outcome type for the dataset. Either \code{Binary} or
\code{Continuous}.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
dataframe of characteristics
}
\description{
Takes bugsnet data and produces a summary dataframe of the studies
}
