% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_char_f.R
\name{summary_char}
\alias{summary_char}
\title{Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.}
\usage{
summary_char(data, metaoutcome)
}
\arguments{
\item{data}{dataframe. Uploaded data}

\item{treatment_df}{vector of treatments}

\item{reference_treatment}{character. The reference treatment}
}
\value{
list
}
\description{
Assess the data for validity. this checks the column names for required columns, and balanced wide format numbered columns.
}
