% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_char_f.R
\name{summary_char}
\alias{summary_char}
\title{Produce a summary dataframe of the studies using \code{BUGSnet::net.tab()}}
\usage{
summary_char(configured_data, logger = NULL)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
dataframe of characteristics
}
\description{
Produce a summary dataframe of the studies using \code{BUGSnet::net.tab()}
}
