% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_network_f.R
\name{summary_network}
\alias{summary_network}
\title{Produce either a netgraph or netplot}
\usage{
summary_network(freq, bugsnetdt, style, label_size = 1, logger = NULL)
}
\arguments{
\item{freq}{dataframe. Created by `frequentist()`}

\item{bugsnetdt}{dataframe. Created by `bugsnetdata()`}

\item{style}{character. The plot to produce, either `netgraph` or `netplot`}

\item{label_size}{numeric. The size of labels in the plots. Default of 1.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
The plotting function
}
\description{
Produce either a netgraph or netplot
}
