% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study}
\alias{summary_study}
\title{Create a forest plot of pairwise comparisons, grouped by treatment pairs.}
\usage{
summary_study(
  connected_data,
  freq,
  outcome_measure,
  plot_area_width = 6,
  colourblind = FALSE,
  x_min = NA,
  x_max = NA
)
}
\arguments{
\item{connected_data}{dataframe. Uploaded data created by `setup_load()`}

\item{freq}{list. Frequentist analysis produced by `setup_configure()` or `setup_exclude()`}

\item{outcome_measure}{character. One of "MD", "SMD", "OR", "RR", or "RD".}

\item{plot_area_width}{numeric. The width of the plot area containing the
treatment effects in inches. Defaults to `6`.}

\item{colourblind}{logical. Whether to use a colourblind-friendly palette. Defaults to `FALSE`}

\item{x_min}{numeric. Minimum value for the x-axis. Defaults to `NA`}

\item{x_max}{numeric. Maximum value for the x-axis. Defaults to `NA`}
}
\value{
List containing:
 \item{svg}{character. SVG code to produce the plot}
 \item{height}{numeric. Plot height in pixels}
 \item{width}{numeric. Plot width in pixels}
}
\description{
Create a forest plot of pairwise comparisons, grouped by treatment pairs.
}
