% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study}
\alias{summary_study}
\title{Creates a summary forest plot of all the studies.}
\usage{
summary_study(freq, outcome_measure, header, title, logger = NULL)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{outcome_measure}{character. "MD", "SMD", "OR", "RR", or "RD".}

\item{header}{numeric. Multiplier for size of text in treatment contrast headers.}

\item{title}{numeric. Multiplier for size of title.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default `NULL`}
}
\value{
Plot created by groupforest.df().
}
\description{
Creates a summary forest plot of all the studies.
}
