% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study}
\alias{summary_study}
\title{Create a forest plot of pairwise comparisons, grouped by treatment pairs.}
\usage{
summary_study(data, freq, outcome_measure, x_limits)
}
\arguments{
\item{outcome_measure}{One of "MD", "SMD", "OR", "RR", or "RD".}

\item{x_limits}{Named vector of x-axis limits in the form (lower, upper).}

\item{pairwise}{Output from meta::pairwise().}

\item{treatment_order}{Vector of the treatments in the order they should appear in the forest plot.}
}
\value{
Forest plot.
}
\description{
Create a forest plot of pairwise comparisons, grouped by treatment pairs.
}
