% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study}
\alias{summary_study}
\title{Creates a summary forest plot of all the studies.}
\usage{
summary_study(freq, outcome_measure, header, title, logger = NULL)
}
\arguments{
\item{freq}{list. NMA results created by freq_wrap().}

\item{outcome_measure}{character. Outcome measure of the dataset. Either
\code{OR}, \code{RR} or \code{RD} when \code{outcome} is \code{Binary} or \code{MD} or \code{SMD} when
\code{outcome} is \code{Continuous}}

\item{header}{numeric. Multiplier for size of text in treatment contrast headers.}

\item{title}{numeric. Multiplier for size of title.}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
Plot created by \code{metafor::forest()}
}
\description{
Creates a summary forest plot of all the studies.
}
