% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study}
\alias{summary_study}
\title{Produce a forest plot of pairwise comparisons, grouped by treatment pairs.
If risk of bias data was loaded these are also included.}
\usage{
summary_study(
  configured_data,
  plot_area_width = 6,
  colourblind = FALSE,
  x_min = NULL,
  x_max = NULL,
  logger = NULL
)
}
\arguments{
\item{configured_data}{list. Input dataset created by \code{setup_configure()} or \code{setup_exclude()}}

\item{plot_area_width}{numeric. The width of the plot area containing the
treatment effects in inches. Defaults to \code{6}.}

\item{colourblind}{logical. Whether to use a colourblind-friendly palette. Defaults to \code{FALSE}}

\item{x_min}{numeric. Minimum value for the x-axis. Defaults to \code{NULL}. For binary outcomes
values should be wrapped in \code{log()}}

\item{x_max}{numeric. Maximum value for the x-axis. Defaults to \code{NULL}. For binary outcomes
values should be wrapped in \code{log()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default \code{NULL}}
}
\value{
html. Contains the svg string to generate the plot. This can be saved
using \code{write_plot()}.
}
\description{
Produce a forest plot of pairwise comparisons, grouped by treatment pairs.
If risk of bias data was loaded these are also included.
}
