% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_study_f.R
\name{summary_study_min_max}
\alias{summary_study_min_max}
\title{Find the default minimum and maximum values for the x-axis and calculate
a sensible step to use in the numeric input}
\usage{
summary_study_min_max(connected_data, freq)
}
\arguments{
\item{connected_data}{dataframe. Uploaded data created by \code{setup_load()}}

\item{freq}{list. Frequentist analysis produced by \code{setup_configure()} or \code{setup_exclude()}}
}
\value{
List containing:
\item{x_min}{numeric. Minimum value for the x-axis}
\item{x_max}{numeric. Maximum value for the x-axis}
\item{step}{numeric. Step value to use in numericInput}
}
\description{
Find the default minimum and maximum values for the x-axis and calculate
a sensible step to use in the numeric input
}
