% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mcmc_f.R
\name{trace_plots}
\alias{trace_plots}
\title{Creates trace plots of MCMC samples.}
\usage{
trace_plots(model, parameters)
}
\arguments{
\item{model}{Model output.}

\item{parameters}{Vector of parameters to create trace plots for.}
}
\value{
List of ggplot trace plots.
}
\description{
Creates trace plots of MCMC samples.
}
