% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_load_f.R
\name{WideToLong}
\alias{WideToLong}
\title{Find all of the treatment names in the data, both for long and wide formats.}
\usage{
WideToLong(wide_data, outcome_type)
}
\arguments{
\item{wide_data}{Data frame of wide format}

\item{outcome_type}{Indicator whether outcome is 'Binary' or 'Continuous'}
}
\value{
Data frame in long format
}
\description{
Find all of the treatment names in the data, both for long and wide formats.
}
