% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_plot}
\alias{write_plot}
\title{Write a plot to a .pdf, .png or .svg file.}
\usage{
write_plot(file, type, renderFunction, height = NULL, width = NULL)
}
\arguments{
\item{file}{The file to which to write.}

\item{type}{String containing the type of file to which to write.}

\item{renderFunction}{A function to render the plot.}

\item{height}{The height of the plot in inches for pdf, or user specified units for png.}

\item{width}{The width of the plot in inches for pdf, or user specified units for png.}
}
\description{
Write a plot to a .pdf, .png or .svg file.
}
